### 2024-02-27 #6422 @editorialbot
[PRE REVIEW]: doped: Python toolkit for robust and repeatable charged defect supercell calculations
**Submitting author:** @kavanase (<a href="http://orcid.org/0000-0003-4577-9647">Seán R. Kavanagh</a>)
**Repository:** https://github.com/SMTG-Bham/doped
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.3.0
