### 2024-02-28 #6422 @rkurchin
@kavanase a question regarding reviewer selection – how crucial would you say it is to testing out your package that reviewers have access to VASP?
