### 2024-02-28 #6422 @rkurchin
👋 @yuan-gist, @yyx5048, and/or @mturiansky, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
