### 2024-02-28 #6422 @kavanase
> @kavanase a question regarding reviewer selection – how crucial would you say it is to testing out your package that reviewers have access to VASP?

@rkurchin access to VASP should not be critical for testing out the package, as we provide several example output files from VASP on the repo for examples and testing (all output data required to perform all steps in the [tutorials](https://doped.readthedocs.io/en/latest/Tutorials.html) is available on the repo). Some general familiarity with VASP might be preferable, just so the reviewers are more familiar with the file I/O process, but this is only a small portion of the code and not totally necessary (same concept as most electronic structure codes).
