### 2024-02-28 4c5d190
Add test for SnB `groundstate` compatibility
