+++ tests/test_thermodynamics.py
@@ -1447,45 +1447,45 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
             ["vac_1_Cd_C2v_Bipolaron_S0", "0", "Td", "C2v", 6.0, 1, 6.0, 1.0],
             ["vac_1_Cd_C2v_Bipolaron_S1", "0", "Td", "C2v", 6.0, 1, 6.0, 1.0],
             ["vac_1_Cd_Td", "0", "Td", "Td", 1.0, 1, 1.0, 1.0],
-            ["vac_2_Te", "-1", "Td", "Cs", 12.0, 2, 24.0, 1.0],
-            ["vac_2_Te", "0", "Td", "C2v", 6.0, 1, 6.0, 1.0],
-            ["vac_2_Te", "-2", "Td", "Cs", 12.0, 1, 12.0, 1.0],
-            ["vac_2_Te", "+1", "Td", "Td", 1.0, 2, 2.0, 1.0],
             ["vac_2_Te", "+2", "Td", "Td", 1.0, 1, 1.0, 1.0],
+            ["vac_2_Te", "+1", "Td", "Td", 1.0, 2, 2.0, 1.0],
+            ["vac_2_Te", "0", "Td", "C2v", 6.0, 1, 6.0, 1.0],
+            ["vac_2_Te", "-1", "Td", "Cs", 12.0, 2, 24.0, 1.0],
+            ["vac_2_Te", "-2", "Td", "Cs", 12.0, 1, 12.0, 1.0],
             ["vac_2_Te_C3v_low_energy_metastable", "0", "Td", "C3v", 4.0, 1, 4.0, 1.0],
             ["vac_2_Te_orig_metastable", "-1", "Td", "Cs", 12.0, 2, 24.0, 1.0],
             ["vac_2_Te_orig_non_JT_distorted", "0", "Td", "Td", 1.0, 1, 1.0, 1.0],
             ["vac_2_Te_shaken", "-2", "Td", "C2v", 6.0, 1, 6.0, 1.0],
             ["vac_2_Te_unperturbed", "-2", "Td", "C2", 12.0, 1, 12.0, 1.0],
-            ["as_1_Cd_on_Te", "-2", "Td", "C1", 24.0, 1, 24.0, 1.0],
-            ["as_1_Cd_on_Te", "-1", "Td", "Td", 1.0, 2, 2.0, 1.0],
+            ["as_1_Cd_on_Te", "+2", "Td", "Td", 1.0, 1, 1.0, 1.0],
             ["as_1_Cd_on_Te", "+1", "Td", "C2v", 6.0, 2, 12.0, 1.0],
             ["as_1_Cd_on_Te", "0", "Td", "Cs", 12.0, 1, 12.0, 1.0],
-            ["as_1_Cd_on_Te", "+2", "Td", "Td", 1.0, 1, 1.0, 1.0],
+            ["as_1_Cd_on_Te", "-1", "Td", "Td", 1.0, 2, 2.0, 1.0],
+            ["as_1_Cd_on_Te", "-2", "Td", "C1", 24.0, 1, 24.0, 1.0],
+            ["as_1_Te_on_Cd", "+2", "Td", "Td", 1.0, 1, 1.0, 1.0],
+            ["as_1_Te_on_Cd", "+1", "Td", "C3v", 4.0, 2, 8.0, 1.0],
+            ["as_1_Te_on_Cd", "0", "Td", "C3v", 4.0, 1, 4.0, 1.0],
+            ["as_1_Te_on_Cd", "-1", "Td", "Cs", 12.0, 2, 24.0, 1.0],
[…]
