### 2024-02-28 ff39f94
Add test for SnB `groundstate` compatibility
Former-commit-id: 4c5d190f3018901cd43ce246409745ae0e8b8bae
