### 2024-02-29 b55ca53
Update `SnB` requirement to match updated `format_defect_name` name
Former-commit-id: dccfb8730bd87cab5c8de95739138b80f2d0a398 [formerly 9ab614f0705b73ad427a27f3fcc724f6296ea474]
Former-commit-id: e06d8a20fee3dd59c70b27b66775a8085c3010bd
