+++ pyproject.toml
@@ -23,7 +23,7 @@ dependencies = [
     "numpy>=1.21.0",
     "pymatgen>=2023.11.12",  # limiting factor: POTCAR_STATS_PATH available from pymatgen.io.vasp.inputs
     "pymatgen-analysis-defects>=2023.8.22",
-    "shakenbreak>=3.3.0",
+    "shakenbreak>=3.3.1",
     "pandas",
     "pydefect",  # for Kumagai (eFNV) correction
     "filelock",
