### 2024-02-29 d89d2e5
Update handling of LHFCALC/SOC warnings so only shown if user tries to write nkred/ncl folders
Former-commit-id: 8f015b95d6dd3468150871ce624bb2db33e5b2ae
