+++ doped/vasp.py
@@ -813,11 +813,6 @@ class DefectRelaxSet(MSONable):
         chemical potential) calculations.
         """
         if self.user_incar_settings.get("LHFCALC", True) is False:  # GGA
-            warnings.warn(
-                "`LHFCALC` is set to `False` in user_incar_settings, so `vasp_nkred_std` (as NKRED "
-                "acts to downsample the Fock exchange potential in _hybrid DFT_ calculations), "
-                "and so `vasp_std` should be used instead."
-            )
             return None
 
         std_defect_set = self.vasp_std
@@ -896,11 +891,6 @@ class DefectRelaxSet(MSONable):
         chemical potential) calculations.
         """
         if not self.soc:
-            warnings.warn(
-                "DefectRelaxSet.soc is False, so vasp_ncl is None (i.e. no `vasp_ncl` input files have "
-                "been generated). If SOC calculations are desired, set soc=True when initializing "
-                "DefectRelaxSet. Otherwise, use vasp_std or vasp_gam instead."
-            )
             return None
 
         user_incar_settings = copy.deepcopy(self.user_incar_settings)
@@ -1147,11 +1137,6 @@ class DefectRelaxSet(MSONable):
             return None
 
         if not self.soc:
-            warnings.warn(
-                "DefectRelaxSet.soc is False, so bulk_vasp_ncl is None (i.e. no `vasp_ncl` input files "
-                "have been generated). If SOC calculations are desired, set soc=True when initializing "
-                "DefectRelaxSet. Otherwise, use bulk_vasp_std or bulk_vasp_gam instead."
-            )
             return None
 
         user_incar_settings = copy.deepcopy(self.user_incar_settings)
@@ -1249,7 +1234,10 @@ class DefectRelaxSet(MSONable):
                 the recommended workflow is to initially perform ``vasp_gam``
                 ground-state structure searching using ShakeNBreak
                 (https://shakenbreak.readthedocs.io), then continue the
-                ``vasp_std`` relaxations from the 'Groundstate' ``CONTCAR``\s.
+                ``vasp(_nkred)_std`` relaxations from the ground-state structures
+                (e.g. using ``-d vasp_nkred_std`` with `snb-groundstate` (CLI) or
+                ``groundstate_folder="vasp_nkred_std"`` with
+                ``write_groundstate_structure`` (Python API)).
                 (default: True)
             bulk (bool):
                 If True, the input files for a singlepoint calculation of the
@@ -1297,12 +1285,12 @@ class DefectRelaxSet(MSONable):
         generate ``POSCAR`` (input structure) files, as these should be taken
         from the ``CONTCAR``\s of ``vasp_std`` relaxations using ``NKRED(X,Y,Z)``
         (originally from ``ShakeNBreak`` relaxations) if using hybrid DFT, or
-        from ``ShakeNBreak`` calculations (via ``snb-groundstate``) if using GGA,
-        or, if not following the recommended structure-searching workflow, from
-        the ``CONTCAR``\s of ``vasp_gam`` calculations. If unperturbed ``POSCAR``
-        files are desired, set ``unperturbed_poscar=True``. If bulk is True, the
-        input files for a singlepoint calculation of the bulk supercell are
-        also written to "{formula}_bulk/{subfolder}".
+        from ``ShakeNBreak`` calculations (via ``snb-groundstate -d vasp_std``)
+        if using GGA, or, if not following the recommended structure-searching
+        workflow, from the ``CONTCAR``\s of ``vasp_gam`` calculations. If
+        unperturbed ``POSCAR`` files are desired, set ``unperturbed_poscar=True``.
+        If bulk is True, the input files for a singlepoint calculation of the bulk
+        supercell are also written to "{formula}_bulk/{subfolder}".
 
[…]
