### 2024-02-29 8cd910a
Add MgO VASP tests (tests warnings and defaults)
Former-commit-id: c7330b88e90787ddfd17f0291c4b218d25126dc2
