+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "doped"
-version = "2.3.0"
+version = "2.3.1"
 description = "Python package to setup, process and analyse solid-state defect calculations with VASP"
 authors = [{name = "Seán Kavanagh", email = "sean.kavanagh.19@ucl.ac.uk"}]
 readme = "README.md"
