### 2024-02-29 4a48443
Update tutorials to show use of `groundstate_folder="vasp_nkred_std"` etc (also updated in SnB & CLI tutorials) which streamlines `SnB` -> `doped` workflow
