### 2024-03-01 99b1c16
Update docstrings prints and `install doped` in notebooks
Former-commit-id: 68e62dc3da7ef7481d72c295d92f59ef4f8258bd
