+++ docs/Dev_ToDo.md
@@ -110,5 +110,4 @@
 - Add sanity check warning with `DefectsGenerator`, if input structure symmetry is `P1` – you sure about this??
 - Add spin degeneracy auto determination to get symmetries/degeneracies later (if hasn't been parsed with current code for some reason e.g. Kanta...). Should implement this for all degeneracy factors whenever concentration methods used?
 - Deal with cases where "X-rich"/"X-poor" corresponds to more than one limit (pick one and warn user?)
-- Avoid `invalid value encountered in scalar divide` warning when using `plot_site_displacements(vector_to_project_on...)` and update advanced analysis tutorial. Also note in docs where in `calculation_metadata` this data is stored.
 - Note in chempot tutorial you may need to increase the kpoints range for certain metals to reach convergence. Also show writing `KPOINTS` files from convergence dict (as in Alkali_Doping notebook) to folders.
