### 2024-03-01 68e62dc
Update docstrings prints and `install doped` in notebooks
