### 2024-03-01 d14aef4
Refactor `facet(s)` to `limit(s)` for chemical potential limits as this is more intuitive for most users
Former-commit-id: f4ccfba72c6e86c4747c52a6ae61d22ab8692dab
