+++ tests/test_thermodynamics.py
@@ -260,7 +260,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
         assert isinstance(df, pd.DataFrame)
         if chempots is not None:
             assert any(
-                "No facet (chemical potential limit) specified! Using" in str(warn.message)
+                "No limit (chemical potential limit) specified! Using" in str(warn.message)
                 for warn in conc_w
             )
 
@@ -345,7 +345,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
                         {
                             "Compensating Defect",
                             "Dopability Limit (eV from VBM/CBM)",
-                            "Facet",
+                            "limit",
                             "Doping Window (eV at VBM/CBM)",
                         }
                     )
@@ -861,22 +861,22 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
             ]:
                 print(f"Testing formation energy methods with {name} input")  # for debugging
                 if thermo_obj.chempots is None:
-                    facets = [
+                    limits = [
                         None,
                     ]
                     with warnings.catch_warnings(record=True) as w:
-                        _form_en = thermo_obj.get_formation_energy(entry, facet="test pop b...")
+                        _form_en = thermo_obj.get_formation_energy(entry, limit="test pop b...")
                     assert len(w) == 2
                     assert "No chemical potentials supplied" in str(w[0].message)
                     assert (
-                        "You have specified a facet (chemical potential limit) but no chemical potentials "
-                        "(`chempots`) were supplied, so `facet` will be ignored." in str(w[-1].message)
+                        "You have specified a limit (chemical potential limit) but no chemical potentials "
+                        "(`chempots`) were supplied, so `limit` will be ignored." in str(w[-1].message)
                     )
 
-                elif len(thermo_obj.chempots["facets_wrt_el_refs"]) == 2:  # CdTe full chempots
+                elif len(thermo_obj.chempots["limits_wrt_el_refs"]) == 2:  # CdTe full chempots
                     with warnings.catch_warnings(record=True) as w:
                         assert (
-                            not np.isclose(  # if chempots present, uses the first facet which is Cd-rich
+                            not np.isclose(  # if chempots present, uses the first limit which is Cd-rich
                                 thermo_obj.get_formation_energy(entry),
                                 form_en_df_row[8],
                                 atol=1e-3,
@@ -884,34 +884,34 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
                         )
                     assert len(w) == 1
                     assert (
-                        "No facet (chemical potential limit) specified! Using Cd-CdTe for computing "
+                        "No limit (chemical potential limit) specified! Using Cd-CdTe for computing "
                         "the formation energy" in str(w[0].message)
                     )
-                    facets = ["CdTe-Te", "Te-rich"]
+                    limits = ["CdTe-Te", "Te-rich"]
                 else:
-                    facets = list(thermo_obj.chempots["facets_wrt_el_refs"].keys())  # user supplied
+                    limits = list(thermo_obj.chempots["limits_wrt_el_refs"].keys())  # user supplied
 
-                for facet in facets:
+                for limit in limits:
                     if np.isclose(fermi_level, 0.75):  # default CdTe:
[…]
