+++ doped/chemical_potentials.py
@@ -44,7 +44,6 @@ warnings.filterwarnings(
 # TODO: Need to recheck all functionality from old `_chemical_potentials.py` is now present here.
 # TODO: Add chemical potential diagram plotting functionality that we had before
 #  with `plot_cplap_ternary`.
-# TODO: Refactor 'facet' to 'limit' in chempot dicts?
 
 
 def make_molecule_in_a_box(element: str):
@@ -449,7 +448,7 @@ class CompetingPhases:
             # diagram, if their relative energy was downshifted by ``e_above_hull``:
             MP_gga_chempots = get_chempots_from_phase_diagram(bulk_ce, phase_diagram)
 
-            MP_bordering_phases = {phase for facet in MP_gga_chempots for phase in facet.split("-")}
+            MP_bordering_phases = {phase for limit in MP_gga_chempots for phase in limit.split("-")}
             self.entries = [
                 entry for entry in pd_entries if entry.name in MP_bordering_phases or entry.is_element
             ]
@@ -839,7 +838,7 @@ class ExtrinsicCompetingPhases(CompetingPhases):
             if codoping:
                 # TODO: `full_sub_approach` shouldn't necessarily mean `full_phase_diagram =
                 #  True` right? As in can be non-full-phase-diagram intrinsic + extrinsic
-                #  entries, including facets with multiple extrinsic entries but still not the
+                #  entries, including limits with multiple extrinsic entries but still not the
                 #  full phase diagram? - To be updated!
                 # TODO: When `full_phase_diagram` option added to `CompetingPhases`, can remove
                 #  this code block and just use:
@@ -1020,19 +1019,19 @@ class ExtrinsicCompetingPhases(CompetingPhases):
                 )
                 MP_extrinsic_bordering_phases = []
 
-                for facet in MP_extrinsic_gga_chempots:
-                    # if the number of intrinsic competing phases for this facet is equal to the
+                for limit in MP_extrinsic_gga_chempots:
+                    # if the number of intrinsic competing phases for this limit is equal to the
                     # number of species in the bulk composition, then include the extrinsic phase(s)
-                    # for this facet (full_sub_approach = False approach)
+                    # for this limit (full_sub_approach = False approach)
                     MP_intrinsic_bordering_phases = {
-                        phase for phase in facet.split("-") if sub_el not in phase
+                        phase for phase in limit.split("-") if sub_el not in phase
                     }
                     if len(MP_intrinsic_bordering_phases) == len(self.intrinsic_species):
                         # add to list of extrinsic bordering phases, if not already present:
                         MP_extrinsic_bordering_phases.extend(
                             [
                                 phase
-                                for phase in facet.split("-")
+                                for phase in limit.split("-")
                                 if sub_el in phase and phase not in MP_extrinsic_bordering_phases
                             ]
                         )
@@ -1055,19 +1054,19 @@ class ExtrinsicCompetingPhases(CompetingPhases):
                         )
 
                         if new_MP_extrinsic_gga_chempots != MP_extrinsic_gga_chempots:
-                            # new bordering phase, check if not an over-dependent facet:
+                            # new bordering phase, check if not an over-dependent limit:
 
-                            for facet in new_MP_extrinsic_gga_chempots:
-                                if facet not in MP_extrinsic_gga_chempots:
[…]
