### 2024-03-01 f4ccfba
Refactor `facet(s)` to `limit(s)` for chemical potential limits as this is more intuitive for most users
