+++ tests/test_plotting.py
@@ -45,7 +45,7 @@ class DefectPlottingTestCase(unittest.TestCase):
 
     @custom_mpl_image_compare(filename="CdTe_example_defects_plot.png")
     def test_plot_CdTe(self):
-        return self.CdTe_thermo.plot(self.CdTe_chempots, facet="CdTe-Te")
+        return self.CdTe_thermo.plot(self.CdTe_chempots, limit="CdTe-Te")
 
     @custom_mpl_image_compare(filename="CdTe_example_defects_plot_listed_colormap.png")
     def test_plot_CdTe_custom_listed_colormap(self):
@@ -60,12 +60,12 @@ class DefectPlottingTestCase(unittest.TestCase):
 
     @custom_mpl_image_compare(filename="CdTe_example_defects_plot.png")
     def test_plot_CdTe_multiple_figs(self):
-        # when facets not specified, plots all of them (second should be Te-rich here)
+        # when limits not specified, plots all of them (second should be Te-rich here)
         return self.CdTe_thermo.plot(self.CdTe_chempots)[1]
 
     @custom_mpl_image_compare(filename="CdTe_example_defects_plot_Cd_rich.png")
     def test_plot_CdTe_Cd_rich(self):
-        # when facets not specified, plots all of them (first should be Cd-rich here)
+        # when limits not specified, plots all of them (first should be Cd-rich here)
         return self.CdTe_thermo.plot(self.CdTe_chempots)[0]
 
     @custom_mpl_image_compare(filename="YTOS_example_defects_plot.png")
@@ -389,25 +389,25 @@ class DefectPlottingTestCase(unittest.TestCase):
             for defect in [dir for dir in os.listdir(f"{data_dir}/V2O5") if "v_O" in dir]
         }  # charge auto-determined (as neutral)
         thermo = DefectThermodynamics(list(defect_dict.values()))
-        return thermo.plot(chempots, facet="V2O5-O2")
+        return thermo.plot(chempots, limit="V2O5-O2")
 
 
 class DefectThermodynamicsPlotsTestCase(DefectThermodynamicsSetupMixin):
-    def test_plot_facet_no_chempots_error(self):
+    def test_plot_limit_no_chempots_error(self):
         with self.assertRaises(ValueError) as exc:
-            self.CdTe_defect_thermo.plot(facet="Te-rich")
+            self.CdTe_defect_thermo.plot(limit="Te-rich")
         assert (
-            "You have specified a chemical potential facet, but no `chempots` have been supplied, "
-            "so `facet` cannot be used here!"
+            "You have specified a chemical potential limit, but no `chempots` have been supplied, "
+            "so `limit` cannot be used here!"
         ) in str(exc.exception)
 
-    def test_plot_facet_user_chempots_error(self):
+    def test_plot_limit_user_chempots_error(self):
         with self.assertRaises(ValueError) as exc:
-            self.CdTe_defect_thermo.plot(chempots={"Cd": 0.5, "Te": -0.5}, facet="Te-rich")
+            self.CdTe_defect_thermo.plot(chempots={"Cd": 0.5, "Te": -0.5}, limit="Te-rich")
         assert (
-            "You have specified a chemical potential facet, but the supplied chempots are not in the "
-            "doped format (i.e. with `facets` in the chempots dict), and instead correspond to just a "
-            "single phase diagram facet / chemical potential limit, so `facet` cannot be used here!"
[…]
