+++ doped/thermodynamics.py
@@ -25,7 +25,7 @@ from pymatgen.io.vasp.outputs import Vasprun
 from scipy.optimize import brentq
 from scipy.spatial import HalfspaceIntersection
 
-from doped.chemical_potentials import get_X_poor_facet, get_X_rich_facet
+from doped.chemical_potentials import get_X_poor_limit, get_X_rich_limit
 from doped.core import DefectEntry, _no_chempots_warning, _orientational_degeneracy_warning
 from doped.generation import _sort_defect_entries
 from doped.utils.parsing import (
@@ -51,68 +51,68 @@ def bold_print(string: str) -> None:
     print("\033[1m" + string + "\033[0m")
 
 
-def _raise_facet_with_user_chempots_error(no_chempots=True):
+def _raise_limit_with_user_chempots_error(no_chempots=True):
     problem = (
         (
-            "the supplied chempots are not in the doped format (i.e. with `facets` in the chempots dict), "
-            "and instead correspond to just a single phase diagram facet / chemical potential limit"
+            "the supplied chempots are not in the doped format (i.e. with `limits` in the chempots dict), "
+            "and instead correspond to just a single chemical potential limit"
         )
         if no_chempots
         else "no `chempots` have been supplied"
     )
     raise ValueError(
-        f"You have specified a chemical potential facet, but {problem}, so `facet` cannot be used here!"
+        f"You have specified a chemical potential limit, but {problem}, so `limit` cannot be used here!"
     )
 
 
-def _parse_facet(chempots: Dict, facet: Optional[str] = None):
-    if facet is not None:
-        if facet in chempots["facets"]:
-            return facet  # direct match, just return facet name
-        if "facets" not in chempots or "User Chemical Potentials" in chempots["facets"]:
+def _parse_limit(chempots: Dict, limit: Optional[str] = None):
+    if limit is not None:
+        if limit in chempots["limits"]:
+            return limit  # direct match, just return limit name
+        if "limits" not in chempots or "User Chemical Potentials" in chempots["limits"]:
             # user specified chempots
-            _raise_facet_with_user_chempots_error(no_chempots=True)
-        if "No User Chemical Potentials" in chempots["facets"]:
-            _raise_facet_with_user_chempots_error(no_chempots=False)
-        if "rich" in facet:
-            facet = get_X_rich_facet(facet.split("-")[0], chempots)
-        elif "poor" in facet:
-            facet = get_X_poor_facet(facet.split("-")[0], chempots)
+            _raise_limit_with_user_chempots_error(no_chempots=True)
+        if "No User Chemical Potentials" in chempots["limits"]:
+            _raise_limit_with_user_chempots_error(no_chempots=False)
+        if "rich" in limit:
+            limit = get_X_rich_limit(limit.split("-")[0], chempots)
+        elif "poor" in limit:
+            limit = get_X_poor_limit(limit.split("-")[0], chempots)
 
-    return facet
+    return limit
 
 
-def get_rich_poor_facet_dict(chempots: Dict) -> Dict:
+def get_rich_poor_limit_dict(chempots: Dict) -> Dict:
     """
-    Get a dictionary of {"X-rich": facet, "X-poor": facet...} for each element
+    Get a dictionary of {"X-rich": limit, "X-poor": limit...} for each element
     X in the chempots phase diagram.
 
     Args:
         chempots (dict): The chempots dict, in the doped format.
     """
     if (
-        "facets" not in chempots
-        or "User Chemical Potentials" in chempots["facets"]
-        or "No User Chemical Potentials" in chempots["facets"]
+        "limits" not in chempots
+        or "User Chemical Potentials" in chempots["limits"]
+        or "No User Chemical Potentials" in chempots["limits"]
[…]
