+++ doped/analysis.py
@@ -1230,14 +1230,14 @@ class DefectsParser:
             chempots (dict):
                 Dictionary of chemical potentials to use for calculating the defect
                 formation energies. This can have the form of
-                ``{"facets": [{'facet': [chempot_dict]}]}`` (the format generated by
+                ``{"limits": [{'limit': [chempot_dict]}]}`` (the format generated by
                 ``doped``\'s chemical potential parsing functions (see tutorials)) which
-                allows easy analysis over a range of chemical potentials - where facet(s)
+                allows easy analysis over a range of chemical potentials - where limit(s)
                 (chemical potential limit(s)) to analyse/plot can later be chosen using
-                the ``facets`` argument.
+                the ``limits`` argument.
 
                 Alternatively this can be a dictionary of chemical potentials for a
-                single facet (limit), in the format: ``{element symbol: chemical potential}``.
+                single limit (limit), in the format: ``{element symbol: chemical potential}``.
                 If manually specifying chemical potentials this way, you can set the
                 ``el_refs`` option with the DFT reference energies of the elemental phases
                 in order to show the formal (relative) chemical potentials above the
