+++ docs/Dev_ToDo.md
@@ -105,7 +105,6 @@
 - Update Sb2O5 test in `test_analysis.py`
   - Add Sb2O5 as example case in plotting customisation tutorial, nice case for `dist_tol` usage.
 - Plotting lines colour updates.
-- Jupyter lab both notebooks for docstrings! (And other SnB to-dos in Group_V_doping notebook)
 - Add sanity check warning with `DefectsGenerator`, if input structure symmetry is `P1` – you sure about this??
 - Add spin degeneracy auto determination to get symmetries/degeneracies later (if hasn't been parsed with current code for some reason e.g. Kanta...). Should implement this for all degeneracy factors whenever concentration methods used?
 - Deal with cases where "X-rich"/"X-poor" corresponds to more than one limit (pick one and warn user?)
