### 2024-03-01 0c5df7b
Update VASP tests to work without POTCARs
