### 2024-03-01 96e98b7
Update VASP tests to work without POTCARs
Former-commit-id: 0c5df7b0b61f1a517da72cfb94413988ab152bc2
