+++ tests/test_vasp.py
@@ -743,6 +743,9 @@ class DefectRelaxSetTest(unittest.TestCase):
         Test DefectRelaxSet write_all() methods; writing/not writing vasp_gam
         POSCARs.
         """
+        if not _potcars_available():  # need to write files without error for charged defects
+            return
+
         for defect_gen_name in [
             "CdTe_defect_gen",
             "ytos_defect_gen",
@@ -802,34 +805,35 @@ class DefectRelaxSetTest(unittest.TestCase):
         assert poscar.comment == "Cd_i_C3v ~[0.5417,0.5417,0.5417] 0"
 
         assert not os.path.exists("test_dir/CdTe_bulk")
-        drs.write_all("test_dir", bulk=True)
-        assert not os.path.exists("test_dir/CdTe_bulk")
+        if _potcars_available():  # need to write files without error for charged defects
+            drs.write_all("test_dir", bulk=True)
+            assert not os.path.exists("test_dir/CdTe_bulk")
 
-        poscar = Poscar.from_file("CdTe_bulk/vasp_ncl/POSCAR")
-        assert poscar.comment == "Cd27 Te27 - Bulk"
+            poscar = Poscar.from_file("CdTe_bulk/vasp_ncl/POSCAR")
+            assert poscar.comment == "Cd27 Te27 - Bulk"
 
-        drs = DefectRelaxSet(self.CdTe_defect_gen["Cd_i_C3v_0"], poscar_comment="Test pop")
-        drs.write_all("test_dir", unperturbed_poscar=True)
-        poscar = Poscar.from_file("test_dir/vasp_std/POSCAR")
-        assert poscar.comment == "Test pop"
+            drs = DefectRelaxSet(self.CdTe_defect_gen["Cd_i_C3v_0"], poscar_comment="Test pop")
+            drs.write_all("test_dir", unperturbed_poscar=True)
+            poscar = Poscar.from_file("test_dir/vasp_std/POSCAR")
+            assert poscar.comment == "Test pop"
 
-        drs = DefectRelaxSet(self.CdTe_defect_gen["v_Cd_-2"])
-        drs.write_all("test_dir", unperturbed_poscar=True)
-        poscar = Poscar.from_file("test_dir/vasp_std/POSCAR")
-        assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
+            drs = DefectRelaxSet(self.CdTe_defect_gen["v_Cd_-2"])
+            drs.write_all("test_dir", unperturbed_poscar=True)
+            poscar = Poscar.from_file("test_dir/vasp_std/POSCAR")
+            assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
 
-        poscar = Poscar.from_file("test_dir/vasp_nkred_std/POSCAR")
-        assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
+            poscar = Poscar.from_file("test_dir/vasp_nkred_std/POSCAR")
+            assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
 
-        poscar = Poscar.from_file("test_dir/vasp_ncl/POSCAR")
-        assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
+            poscar = Poscar.from_file("test_dir/vasp_ncl/POSCAR")
+            assert poscar.comment == "v_Cd ~[0.3333,0.3333,0.3333] -2"
 
-        assert not os.path.exists("test_dir/CdTe_bulk")
-        drs.write_all("test_dir", bulk=True)
-        assert not os.path.exists("test_dir/CdTe_bulk")
[…]
