### 2024-03-01 36b75c5
Add publication titles
