### 2024-03-01 5276d18
Add publication titles
Former-commit-id: 36b75c57b1fbdb1eb5f20d7408b7a5a86f7ae04f
