### 2024-03-01 dfc078a
Update to match `pymatgen` breaking change from today
