### 2024-03-01 f666158
Make `calc/plot_site_displacements` public functions
