+++ doped/utils/plotting.py
@@ -153,8 +153,10 @@ def format_defect_name(
     wout_charge: bool = False,
 ) -> Optional[str]:
     """
-    Format defect name for plot titles. (i.e. from Cd_i_C3v_0 to $Cd_{i}^{0}$
-    or $Cd_{i_{C3v}}^{0}$). Note this assumes "V_" means vacancy not Vanadium.
+    Format defect name for plot titles.
+
+    (i.e. from Cd_i_C3v_0 to $Cd_{i}^{0}$ or $Cd_{i_{C3v}}^{0}$).
+    Note this assumes "V_" means vacancy not Vanadium.
 
     Args:
         defect_species (:obj:`str`):
