### 2024-03-01 00ab5ec
Make `calc/plot_site_displacements` public functions
Former-commit-id: f6661582f1822e5678066248bcb528829da1df3d
