### 2024-03-01 a4ea4c8
Tests updates and some readme cleanup
