### 2024-03-01 dd11d6b
Update subscripts in pubs list
Former-commit-id: 1103b662f7a02649f05f660263ab0520b8642f29
