### 2024-03-01 ce1bbbb
Updated plotting, fixed p-ratio bug, updated vise check
Former-commit-id: 4316232ec4ebd315acce2d8931e433fdeeb58ea6
