+++ doped/utils/phs.py
@@ -6,9 +6,18 @@ https://github.com/kumagai-group/pydefect
 and vise
 https://github.com/kumagai-group/vise, to avoid the user requiring additional files i.e. PROCAR.
 """
-from typing import Any, List
+import logging
+import os
+import warnings
+from importlib.metadata import version
+from typing import Any, List, Optional
 
+import matplotlib.pyplot as plt
 import numpy as np
+from vise import user_settings
+# suppress pydefect INFO messages
+user_settings.logger.setLevel(logging.CRITICAL)
+
 from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
 from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
 from pydefect.analyzer.make_band_edge_states import make_band_edge_states
@@ -16,8 +25,12 @@ from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
 from pydefect.cli.vasp.make_band_edge_orbital_infos import calc_orbital_character, calc_participation_ratio
 from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
 from pydefect.defaults import defaults
+from pymatgen.electronic_structure.core import Spin
 from pymatgen.io.vasp.outputs import Vasprun
 from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
+
+from doped.utils.plotting import _get_backend, format_defect_name
+
 
 
 def make_band_edge_orbital_infos(
@@ -88,18 +101,34 @@ def get_band_edge_info(DefectParser, bulk_vr, bulk_outcar, defect_vr):
     Returns:
         pydefect EdgeInfo class
     """
-    band_edge_prop = VaspBandEdgeProperties(bulk_vr, bulk_outcar)
+    # Change this to a warning...
+    try:
+        band_edge_prop = VaspBandEdgeProperties(bulk_vr, bulk_outcar)
+        if defect_vr.parameters.get("LNONCOLLINEAR") is True:
+            assert band_edge_prop._ho_band_index(Spin.up) == int(bulk_vr.parameters.get("NELECT")) - 1
+    except AssertionError:
+        v_vise = version("vise")
+        if v_vise <= "0.8.1":
+            warnings.warn(
+                f"You have version {v_vise} of the package `vise`,"
+                f" which does not allow the parsing of non-collinear calculations."
+                f" You can install the updated version of `vise` from the GitHub repo for this"
+                f" functionality. Attempting to load the PHS data has been automatically skipped"
+            )
+            return None, None, None
+
     orbs, s = bulk_vr.projected_eigenvalues, bulk_vr.final_structure
     vbm_info = get_edge_info(band_edge_prop.vbm_info, orbs, s, bulk_vr)
     cbm_info = get_edge_info(band_edge_prop.cbm_info, orbs, s, bulk_vr)
 
+    # Using default values pydefect
     dsinfo = MakeDefectStructureInfo(
         DefectParser.defect_entry.bulk_supercell,
         DefectParser.defect_entry.defect_supercell,
         DefectParser.defect_entry.defect_supercell,
         symprec=0.1,
-        dist_tol=0.05,
-        neighbor_cutoff_factor=2,
+        dist_tol=1.0,
+        neighbor_cutoff_factor=1.3,
     )
 
     band_orb = make_band_edge_orbital_infos(
@@ -113,15 +142,17 @@ def get_band_edge_info(DefectParser, bulk_vr, bulk_outcar, defect_vr):
     return band_orb, vbm_info, cbm_info
 
 
[…]
