### 2024-03-01 e00ff95
Add load_phs_data to DefectsParser
Former-commit-id: 799c5d0e734657f27f03b8931b11bfe1b1d77c27
