### 2024-03-01 b4ad5cd
Improved efficiency load data
Former-commit-id: 02c9819a5428f9e1124ddd06407d9cdd1351340d
