### 2024-03-01 38b6c31
Avoid deprecation warning with `python=3.11`, drop `python=3.8` in line with `pymatgen` and update linting
Former-commit-id: 57218f097b757061e57ded9f0b754f4996b10692
