+++ doped/vasp.py
@@ -1,6 +1,7 @@
 """
 Code to generate VASP defect calculation input files.
 """
+
 import contextlib
 import copy
 import inspect
@@ -11,7 +12,7 @@ from functools import lru_cache
 from importlib import resources
 from multiprocessing import cpu_count
 from multiprocessing.pool import Pool
-from typing import Dict, List, Optional, Tuple, Union, cast
+from typing import Optional, Union, cast
 
 import numpy as np
 from monty.io import zopen
@@ -73,7 +74,7 @@ singleshot_incar_settings = {
 
 
 def _test_potcar_functional_choice(
-    potcar_functional: UserPotcarFunctional = "PBE", symbols: Optional[List] = None
+    potcar_functional: UserPotcarFunctional = "PBE", symbols: Optional[list] = None
 ):
     """
     Check if the potcar functional choice needs to be changed to match those
@@ -106,8 +107,8 @@ def _get_potcar(potcar_symbols, potcar_functional) -> Potcar:
 
 class DopedKpoints(Kpoints):
     """
-    Custom implementation of ``Kpoints`` to handle encoding
-    errors that can happen on some old HPCs/Linux systems.
+    Custom implementation of ``Kpoints`` to handle encoding errors that can
+    happen on some old HPCs/Linux systems.
 
     If an encoding error occurs upon file writing, then changes Γ to
     Gamma and Å to Angstrom in the ``KPOINTS`` comment.
@@ -193,12 +194,12 @@ class DefectDictSet(DictSet):
 
         if user_incar_settings is not None:
             # Load INCAR tag/value check reference file from pymatgen.io.vasp.inputs
-            with resources.open_text("pymatgen.io.vasp", "incar_parameters.json") as json_file:
+            with open(f"{resources.files('pymatgen.io.vasp')}/incar_parameters.json") as json_file:
                 incar_params = json.load(json_file)
 
             for k in user_incar_settings:
                 # check INCAR flags and warn if they don't exist (typos)
-                if k not in incar_params.keys() and "#" not in k:
+                if k not in incar_params and "#" not in k:
                     warnings.warn(  # but only checking keys, not values so we can add comments etc
                         f"Cannot find {k} from your user_incar_settings in the list of INCAR flags",
                         BadIncarWarning,
@@ -340,8 +341,8 @@ class DefectDictSet(DictSet):
     @property
     def nelect(self):
         """
-        Number of electrons (``NELECT``) for the given structure and
-        charge state.
+        Number of electrons (``NELECT``) for the given structure and charge
+        state.
 
         This is equal to the sum of valence electrons (ZVAL) of the
         ``POTCAR``s for each atom in the structure (supercell), minus
@@ -392,8 +393,9 @@ class DefectDictSet(DictSet):
         snb: bool = False,
     ):
         """
-        Writes out all input to a directory. Refactored slightly from ``pymatgen``
-        ``DictSet.write_input()`` to allow checking of user ``POTCAR`` setup.
+        Writes out all input to a directory. Refactored slightly from
+        ``pymatgen`` ``DictSet.write_input()`` to allow checking of user
+        ``POTCAR`` setup.
 
         Args:
             output_path (str): Directory to output the VASP input files.
@@ -507,9 +509,9 @@ class DefectRelaxSet(MSONable):
         **kwargs,
     ):
         r"""
-        The supercell structure and charge state are taken from the ``DefectEntry``
-        attributes, or if a ``Structure`` is provided, then from the
-        ``defect_supercell`` and ``charge_state`` input parameters.
[…]
