### 2024-03-01 0978349
Avoid deprecation warning with `python=3.11`, drop `python=3.8` in line with `pymatgen` and update linting
Former-commit-id: 1ec39915a30116d6c1045c2c6a2de0418c0c528a [formerly 57218f097b757061e57ded9f0b754f4996b10692]
Former-commit-id: 38b6c3104e6d893b12757c479be5abf101ad32bc
