### 2024-03-01 57218f0
Avoid deprecation warning with `python=3.11`, drop `python=3.8` in line with `pymatgen` and update linting
