+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Chemistry",
     "Topic :: Software Development :: Libraries :: Python Modules",
     ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "tabulate",
     "matplotlib>=3.5.2",
@@ -59,9 +59,9 @@ line-length = 107
 
 [tool.ruff]
 line-length = 107
-pydocstyle.convention = "google"
-isort.split-on-trailing-comma = false
-select = [  # from pymatgen
+lint.pydocstyle.convention = "google"
+lint.isort.split-on-trailing-comma = false
+lint.select = [  # from pymatgen
   "B",   # flake8-bugbear
   "C4",  # flake8-comprehensions
   "D",   # pydocstyle
@@ -89,7 +89,7 @@ select = [  # from pymatgen
   "W",   # pycodestyle warning
   "YTT", # flake8-2020
 ]
-ignore = [
+lint.ignore = [
   "B028",    # No explicit stacklevel keyword argument found
   "D101",    # Missing docstring in public class (docstring in init instead)
   "D200",    # One-line docstring should fit on one line with quotes
@@ -100,7 +100,7 @@ ignore = [
   "W605",    # Invalid escape sequence
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["D102"]
 
 [[tool.mypy.overrides]]
