+++ doped/corrections.py
@@ -37,7 +37,7 @@ eFNV) were developed, they should be used here.
 
 import os
 import warnings
-from typing import List, Optional, Union
+from typing import Optional, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -290,7 +290,7 @@ def plot_FNV(plot_data, title=None, ax=None, style_file=None):
             x,
             short_range,
             c="green",
-            label=r"$V_{sr}$ = $\Delta$(Locpot) - $V_{lr}$" + "\n(pre-alignment)",  # noqa: ISC003
+            label=r"$V_{sr}$ = $\Delta$(Locpot) - $V_{lr}$" + "\n(pre-alignment)",
         )
         leg1 = ax.legend(handles=[line1, line2, line3], loc=9)  # middle top legend
         ax.add_artist(leg1)  # so isn't overwritten with later legend call
@@ -318,7 +318,7 @@ def get_kumagai_correction(
     defect_entry,
     dielectric: Optional[Union[float, int, np.ndarray, list]] = None,
     defect_region_radius: Optional[float] = None,
-    excluded_indices: Optional[List[int]] = None,
+    excluded_indices: Optional[list[int]] = None,
     defect_outcar: Optional[Union[str, Outcar]] = None,
     bulk_outcar: Optional[Union[str, Outcar]] = None,
     plot: bool = False,
