### 2024-03-01 00ac201
Fix missing Gamma glyph when plotting
