### 2024-03-01 2a883b9
Fix missing Gamma glyph when plotting
Former-commit-id: 070fdaa01979bcb8f45f5df5d6b4ed33eef5b3fe [formerly 00ac20152421d35dc0a3e5eb0c2f90948b8dfe44]
Former-commit-id: 882ebc4e276a7c1464fbe2953a0dde4e04f1cf22
