+++ doped/utils/phs.py
@@ -1,246 +1,0 @@
-"""
-Helper functions for setting up PHS analysis.
-
-Contains modified versions of functions from pydefect
-https://github.com/kumagai-group/pydefect
-and vise
-https://github.com/kumagai-group/vise, to avoid the user requiring additional files i.e. PROCAR.
-"""
-import logging
-import os
-import warnings
-from importlib.metadata import version
-from typing import Any, List, Optional
-
-import matplotlib.pyplot as plt
-import numpy as np
-from vise import user_settings
-# suppress pydefect INFO messages
-user_settings.logger.setLevel(logging.CRITICAL)
-
-from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
-from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
-from pydefect.analyzer.make_band_edge_states import make_band_edge_states
-from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
-from pydefect.cli.vasp.make_band_edge_orbital_infos import calc_orbital_character, calc_participation_ratio
-from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
-from pydefect.defaults import defaults
-from pymatgen.electronic_structure.core import Spin
-from pymatgen.io.vasp.outputs import Vasprun
-from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
-
-from doped.utils.plotting import _get_backend, format_defect_name
-
-
-
-def make_band_edge_orbital_infos(
-    vasprun: Vasprun, vbm: float, cbm: float, str_info, eigval_shift: float = 0.0
-):
-    """
-    Make BandEdgeOrbitalInfos from vasprun.xml.
-
-    Args:
-        vasprun: Vasprun object
-        vbm: vbm_info
-        cbm: cbm_info
-        str_info: DefectStructureInfo
-        eigval_shift (float): shift eigenvalues by a value
-
-    Returns:
-        BandEdgeOrbitalInfos
-    """
-    eigval_range = defaults.eigval_range
-    kpt_coords = [tuple(coord) for coord in vasprun.actual_kpoints]
-    max_energy_by_spin, min_energy_by_spin = [], []
-    neighbors = str_info.neighbor_atom_indices
-
-    for e in vasprun.eigenvalues.values():
-        max_energy_by_spin.append(np.amax(e[:, :, 0], axis=0))
-        min_energy_by_spin.append(np.amin(e[:, :, 0], axis=0))
-
-    max_energy_by_band = np.amax(np.vstack(max_energy_by_spin), axis=0)
-    min_energy_by_band = np.amin(np.vstack(min_energy_by_spin), axis=0)
-
-    lower_idx = np.argwhere(max_energy_by_band > vbm - eigval_range)[0][0]
-    upper_idx = np.argwhere(min_energy_by_band < cbm + eigval_range)[-1][-1]
-
-    orbs, s = vasprun.projected_eigenvalues, vasprun.final_structure
-    orb_infos: List[Any] = []
-    for spin, eigvals in vasprun.eigenvalues.items():
-        orb_infos.append([])
-        for k_idx in range(len(kpt_coords)):
-            orb_infos[-1].append([])
-            for b_idx in range(lower_idx, upper_idx + 1):
-                e, occ = eigvals[k_idx, b_idx, :]
-                orbitals = calc_orbital_character(orbs, s, spin, k_idx, b_idx)
-                if neighbors:
-                    p_ratio = calc_participation_ratio(orbs, spin, k_idx, b_idx, neighbors)
-                else:
[…]
