### 2024-03-01 48a5e38
Supress warning missing Gamma glyph when plotting
