+++ doped/utils/phs.py
@@ -261,19 +261,6 @@ def get_phs_and_eigenvalue(DefectEntry, filename: Optional[str] = None, ks_label
         ax.axhline(-5, 0, 1, color="black", linewidth=0.5, linestyle="-.", label="Band edges")
         ax.legend(loc="upper right", fontsize=7)
 
-        # gamma_check = "\N{GREEK CAPITAL LETTER GAMMA}"
-        # locs, labels = plt.xticks()
-        #
-        # # Replace the label containing 'gamma' (γ) with the word 'gamma'
-        # labels = [label.get_text() for label in labels]
-        #
-        # # Replace the label containing 'gamma' (γ) with the word 'gamma'
-        # for i, label in enumerate(labels):
-        #     if gamma_check in label:  # Check if the label contains 'gamma'
-        #         labels[i] = r'$\Gamma$'  # Replace the label
-        #
-        # plt.xticks(locs, labels)
-
     if filename:
         emp.plt.savefig(filename, bbox_inches="tight", transparent=True, backend=_get_backend(filename))
 
