### 2024-03-01 14197d6
Supress warning missing Gamma glyph when plotting
Former-commit-id: 48a5e382b60f20b7120cf480ab9da0107469c817
