### 2024-03-01 be48d14
Update file parsing to avoid hidden files and add tests (flagged by @adair-nicolson 🙌)
