### 2024-03-01 0abaf63
Update file parsing to avoid hidden files and add tests (flagged by @adair-nicolson 🙌)
Former-commit-id: be48d14ebcbaadb93a0960bc1f6c533ea9d73df0
