### 2024-03-01 a392859
Update chempot file parsing to avoid hidden files and give more informative error messages, and add tests
Former-commit-id: 9825cbcb9979d2f62939260e80b86f087afa1352
