### 2024-03-01 9825cbc
Update chempot file parsing to avoid hidden files and give more informative error messages, and add tests
