### 2024-03-02 e696bbb
Update to tips and tricks
Former-commit-id: 0b007169a314e0cf3000a65ad2c6531ba688fa34
