### 2024-03-02 7db1fcf
Added new tests for wrong vise version and test for no projected orbitals. Updated old tests to include load_phs_data=False to avoid clash with new warnings
Former-commit-id: 37832998824256d4ba3dba740beb0b98b9e04068
