+++ tests/test_analysis.py
@@ -300,7 +300,7 @@ class DefectsParsingTestCase(unittest.TestCase):
                 output_path=self.CdTe_EXAMPLE_DIR,
                 dielectric=9.13,
                 json_filename="CdTe_example_defect_dict.json",
-                load_phs_data = False
+                load_phs_data = False,
             )  # for testing in test_thermodynamics.py
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(default_dp, w)
@@ -325,7 +325,8 @@ class DefectsParsingTestCase(unittest.TestCase):
     def test_DefectsParser_CdTe_without_multiprocessing(self):
         # test same behaviour without multiprocessing:
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, processes=1)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, processes=1,
+                               load_phs_data=False)
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w)
 
@@ -333,7 +334,8 @@ class DefectsParsingTestCase(unittest.TestCase):
         # check using filterwarnings works as expected:
         warnings.filterwarnings("ignore", "Multiple")
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13,
+                               load_phs_data=False)
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, multiple_outcars_warning=False)
         warnings.filterwarnings("default", "Multiple")
@@ -342,14 +344,16 @@ class DefectsParsingTestCase(unittest.TestCase):
         # test with reduced dist_tol:
         # Int_Te_3_Unperturbed merged with Int_Te_3 with default dist_tol = 1.5, now no longer merged
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13,
+                               load_phs_data=False)
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, dist_tol=0.1)
 
     def test_DefectsParser_CdTe_no_dielectric_json(self):
         # test no dielectric and no JSON:
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, json_filename=False)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, json_filename=False,
+                               load_phs_data=False)
         print([warn.message for warn in w])  # for debugging
         assert any(
             "The dielectric constant (`dielectric`) is needed to compute finite-size charge "
@@ -370,6 +374,7 @@ class DefectsParsingTestCase(unittest.TestCase):
                 bulk_band_gap_path=self.CdTe_BULK_DATA_DIR,
                 processes=4,
                 json_filename="test_pop.json",
+                load_phs_data=False,
             )
         print([warn.message for warn in w])  # for debugging
         assert any(
[…]
