### 2024-03-02 4627055
Round `0.0` to `0` by default with chempots, add note about kpoints convergence to tutorial, and show how to quickly generate converged KPOINTS in each folder
