### 2024-03-02 18e015e
Round `0.0` to `0` by default with chempots, add note about kpoints convergence to tutorial, and show how to quickly generate converged KPOINTS in each folder
Former-commit-id: 46270559fe392493444086859a336d4a6c6e71db
