### 2024-03-02 94bfdc2
Add sanity check warning with `DefectsGenerator`, if input structure symmetry is `P1` – you sure about this?? And test
