### 2024-03-02 46538e7
Add sanity check warning with `DefectsGenerator`, if input structure symmetry is `P1` – you sure about this?? And test
Former-commit-id: 94bfdc215650901140dfe096037de751b2df375e
