### 2024-03-02 3993b0e
Add Sb2O5 tests and plotting customisation example (merge all inequivalent sites)
Former-commit-id: a34add76fa73393b60efb44fcefc0e302be6f9b3
