+++ tests/test_thermodynamics.py
@@ -19,6 +19,7 @@ import pandas as pd
 import pytest
 from monty.serialization import dumpfn, loadfn
 
+from doped.generation import _sort_defect_entries
 from doped.thermodynamics import DefectThermodynamics, scissor_dos
 
 # for pytest-mpl:
@@ -96,6 +97,7 @@ class DefectThermodynamicsSetupMixin(unittest.TestCase):
         self.V2O5_defect_dict = deepcopy(self.orig_V2O5_defect_dict)
         self.MgO_defect_thermo = deepcopy(self.orig_MgO_defect_thermo)
         self.MgO_defect_dict = deepcopy(self.orig_MgO_defect_dict)
+        self.Sb2O5_defect_thermo = deepcopy(self.orig_Sb2O5_defect_thermo)
 
     @classmethod
     def setUpClass(cls):
@@ -157,6 +159,9 @@ class DefectThermodynamicsSetupMixin(unittest.TestCase):
         cls.orig_MgO_defect_dict = loadfn(os.path.join(cls.MgO_EXAMPLE_DIR, "MgO_defect_dict.json"))
         cls.MgO_chempots = loadfn(os.path.join(cls.EXAMPLE_DIR, "competing_phases/mgo_chempots.json"))
 
+        cls.Sb2O5_chempots = loadfn(os.path.join(data_dir, "Sb2O5/Sb2O5_chempots.json"))
+        cls.orig_Sb2O5_defect_thermo = loadfn(os.path.join(data_dir, "Sb2O5/Sb2O5_thermo.json"))
+
 
 class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
     def _compare_defect_thermo_and_dict(self, defect_thermo, defect_dict):
@@ -200,10 +205,12 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
 
         # CdTe, YTOS, Sb2Se3, Sb2Si2Te6, V2O5, MgO values:
         assert any(
-            np.isclose(defect_thermo.vbm, i, atol=1e-2) for i in [1.65, 3.26, 4.19, 6.60, 0.90, 3.1293]
+            np.isclose(defect_thermo.vbm, i, atol=1e-2)
+            for i in [1.65, 3.26, 4.19, 6.60, 0.90, 3.1293, 4.0002]
         )
         assert any(
-            np.isclose(defect_thermo.band_gap, i, atol=1e-2) for i in [1.5, 0.7, 1.47, 0.44, 2.22, 4.7218]
+            np.isclose(defect_thermo.band_gap, i, atol=1e-2)
+            for i in [1.5, 0.7, 1.47, 0.44, 2.22, 4.7218, 3.1259]
         )  # note YTOS is GGA calcs so band gap underestimated
 
         assert defect_thermo.check_compatibility == check_compatibility
@@ -218,10 +225,18 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
         # CdTe, YTOS, Sb2Se3, Sb2Si2Te6, V2O5 values:
         assert any(
             os.path.exists(f"{i}_defect_thermodynamics.json")
-            for i in ["CdTe", "Y2Ti2S2O5", "Sb2Se3", "SiSbTe3", "V2O5", "MgO"]
+            for i in ["CdTe", "Y2Ti2S2O5", "Sb2Se3", "SiSbTe3", "V2O5", "MgO", "Sb2O5"]
         )
-        assert defect_thermo.bulk_formula in ["CdTe", "Y2Ti2S2O5", "Sb2Se3", "SiSbTe3", "V2O5", "MgO"]
[…]
