### 2024-03-02 a34add7
Add Sb2O5 tests and plotting customisation example (merge all inequivalent sites)
