+++ tests/test_analysis.py
@@ -2249,81 +2249,3 @@ class ReorderedParsingTestCase(unittest.TestCase):
         assert np.isclose(
             sum(parsed_v_cd_m2_orig.corrections.values()), sum(parsed_v_cd_m2_alt2.corrections.values())
         )
-
-
-# class AnalysisFunctionsTestCase(unittest.TestCase):
-#     """
-#     Test post-processing analysis functions.
-#     """
-#
-#     def setUp(self):
-#         self.module_path = os.path.dirname(os.path.abspath(__file__))
-#         self.data_dir = os.path.join(os.path.dirname(__file__), "data")
-#         self.sb2o5_chempots = loadfn(f"{self.data_dir}/Sb2O5/Sb2O5_chempots.json")
-#         self.sb2o5_thermo = loadfn(f"{self.data_dir}/Sb2O5/sb2o5_thermo.json")
-#
-#     def tearDown(self):
-#         if_present_rm("test.csv")
-#
-#     def test_get_formation_energies(
-#         self,
-#     ):  # TODO: Get Ke to reparse with new defects thermo code and add here
-#         def _check_formation_energy_table(
-#             formation_energy_table_df, fermi_level=0, thermo=self.sb2o5_thermo
-#         ):
-#             defect_entry_names = [defect_entry.name for defect_entry in thermo.entries]
-#             assert sorted(formation_energy_table_df["Defect"].tolist()) == sorted(defect_entry_names)
-#
-#             # for each row, assert sum of formation energy terms equals formation energy column
-#             np.isclose(
-#                 np.asarray(sum(formation_energy_table_df.iloc[:, i] for i in range(2, 8))),
-#                 np.asarray(formation_energy_table_df.iloc[:, 8]),
-#                 atol=2e-3,
-#             )
-#
-#             assert np.isclose(
-#                 formation_energy_table_df.iloc[:, 1] * fermi_level,
-#                 formation_energy_table_df.iloc[:, 4],
-#                 atol=2e-3,
-#             ).all()
-#
-#         formation_energy_table_df = self.sb2o5_thermo.get_formation_energies(
-#             self.sb2o5_chempots, limits=["Sb2O5-SbO2"], fermi_level=3
-#         )
-#         _check_formation_energy_table(formation_energy_table_df, fermi_level=3)
-#
-#         formation_energy_table_df = self.sb2o5_thermo.get_formation_energies(  # test default w/ E_F = 0
-#             self.sb2o5_chempots,
-#             limits=["Sb2O5-O2"],
-#         )
-#         _check_formation_energy_table(formation_energy_table_df, fermi_level=0)
-#
-#         formation_energy_table_df_manual_chempots = (
-#             self.sb2o5_thermo.get_formation_energies(  # test default with E_F = 0
-#                 chempots=self.sb2o5_chempots["limits_wrt_el_refs"]["Sb2O5-O2"],
-#                 el_refs=self.sb2o5_chempots["elemental_refs"],
-#             )
-#         )
-#         _check_formation_energy_table(formation_energy_table_df_manual_chempots, fermi_level=0)
-#
-#         # check manual and auto chempots the same:
-#         assert formation_energy_table_df_manual_chempots.equals(formation_energy_table_df)
-#
-#         # assert runs fine without chempots:
-#         formation_energy_table_df = self.sb2o5_thermo.get_formation_energies()
-#         _check_formation_energy_table(formation_energy_table_df)
-#
-#         # assert runs fine with only raw chempots:
-#         formation_energy_table_df = self.sb2o5_thermo.get_formation_energies(
-#             chempots=self.sb2o5_chempots["limits"]["Sb2O5-O2"]
-#         )
-#         _check_formation_energy_table(formation_energy_table_df)
[…]
