+++ doped/core.py
@@ -403,7 +403,7 @@ class DefectEntry(thermo.DefectEntry):
         For example, with layered materials, the defect charge is often localised
         to one layer, so we may want to adjust ``defect_region_radius`` and/or
         ``excluded_indices`` to ensure that only sites in other layers are used for
-        the sampling region (plateau) - see example on doped docs.
+        the sampling region (plateau) - see example on doped docs Tips page.
 
         Args:
             dielectric (float or int or 3x1 matrix or 3x3 matrix):
