### 2024-03-03 6a22ea9
Small updates; add `NKRED` to `INCAR` mismatch checks (@utf 🙌), and exclude bulk path from potential defect folders
Former-commit-id: 2322dd861aef5117718c6ae43917ae2cd17f84ed
