+++ doped/utils/parsing.py
@@ -656,6 +656,7 @@ def _compare_incar_tags(
             "PREC": "Normal",  # default Normal
             "PRECFOCK": "Normal",  # default Normal
             "LDAU": False,  # default False
+            "NKRED": None,  # default None
         }
 
     def _compare_incar_vals(val1, val2):
