### 2024-03-03 2322dd8
Small updates; add `NKRED` to `INCAR` mismatch checks (@utf 🙌), and exclude bulk path from potential defect folders
