### 2024-03-03 1be7ec1
Add ZnS periodicity-breaking supercell data and tests
Former-commit-id: d10630764d3d3df27e1a250e32b79370c60964ad
