### 2024-03-03 d106307
Add ZnS periodicity-breaking supercell data and tests
