### 2024-03-03 8e1d800
Add `defect_from_structure` robustness tests with `test_defect_from_structures_rattled`
Former-commit-id: e634a836f04b8cb461953df5bb7faf6d7b728341
