### 2024-03-04 2b24f98
Fix legend in plots
