### 2024-03-09 1e1116a
Add `simple_spin_degeneracy_from_charge` and refactor some utility functions to `doped.utils.parsing`
Former-commit-id: 43219e455df2a02876355fe23e11e7f031385466
