+++ doped/utils/parsing.py
@@ -11,6 +11,7 @@ from typing import Optional, Union
 import numpy as np
 from monty.serialization import loadfn
 from pymatgen.analysis.defects.core import DefectType
+from pymatgen.core.periodic_table import Element
 from pymatgen.core.structure import PeriodicSite, Structure
 from pymatgen.io.vasp.inputs import UnknownPotcarWarning
 from pymatgen.io.vasp.outputs import Locpot, Outcar, Vasprun
@@ -1042,8 +1043,8 @@ def _get_defect_supercell_bulk_site_coords(defect_entry: DefectEntry, relaxed=Tr
     sc_defect_frac_coords = defect_entry.sc_defect_frac_coords
     site = None
 
-    if not relaxed and hasattr(defect_entry, "calculation_metadata") and defect_entry.calculation_metadata:
-        site = defect_entry.calculation_metadata.get("bulk_site")
+    if not relaxed:
+        site = _get_defect_supercell_site(defect_entry, relaxed=False)
     if sc_defect_frac_coords is None and site is None:
         site = _get_defect_supercell_site(defect_entry)
     if site is not None:
@@ -1052,7 +1053,88 @@ def _get_defect_supercell_bulk_site_coords(defect_entry: DefectEntry, relaxed=Tr
     return sc_defect_frac_coords
 
 
-def _get_defect_supercell_site(defect_entry: DefectEntry):
+def _get_defect_supercell_site(defect_entry: DefectEntry, relaxed=True):
+    site = None
+    if not relaxed:
+        if hasattr(defect_entry, "calculation_metadata") and defect_entry.calculation_metadata:
+            site = defect_entry.calculation_metadata.get("bulk_site")
+        if site is None:  # need to reparse info:
+            from doped.analysis import defect_from_structures
+
+            bulk_supercell = _get_bulk_supercell(defect_entry)
+            defect_supercell = _get_defect_supercell(defect_entry)
+
+            (
+                _defect,
+                _defect_site,  # _relaxed_ defect site in supercell (if substitution/interstitial)
+                defect_site_in_bulk,  # bulk site for vacancies/substitutions, relaxed defect site
+                # w/interstitials
+                defect_site_index,
+                bulk_site_index,
+                guessed_initial_defect_structure,
+                unrelaxed_defect_structure,
+                _bulk_voronoi_node_dict,
+            ) = defect_from_structures(bulk_supercell, defect_supercell, return_all_info=True)
+
+            # update any missing calculation_metadata:
+            defect_entry.calculation_metadata["guessed_initial_defect_structure"] = (
+                defect_entry.calculation_metadata.get(
+                    "guessed_initial_defect_structure", guessed_initial_defect_structure
+                )
+            )
+            defect_entry.calculation_metadata["defect_site_index"] = defect_entry.calculation_metadata.get(
+                "defect_site_index", defect_site_index
+            )
+            defect_entry.calculation_metadata["bulk_site_index"] = defect_entry.calculation_metadata.get(
+                "bulk_site_index", bulk_site_index
+            )
+            defect_entry.calculation_metadata["unrelaxed_defect_structure"] = (
+                defect_entry.calculation_metadata.get(
+                    "unrelaxed_defect_structure", unrelaxed_defect_structure
+                )
+            )
+
+            # add displacement from (guessed) initial site to final defect site:
+            if defect_site_index is not None:  # not a vacancy
+                guessed_initial_site = guessed_initial_defect_structure[defect_site_index]
+                final_site = defect_supercell[defect_site_index]
+                guessed_displacement = final_site.distance(guessed_initial_site)
+                defect_entry.calculation_metadata["guessed_initial_defect_site"] = (
[…]
