### 2024-03-09 43219e4
Add `simple_spin_degeneracy_from_charge` and refactor some utility functions to `doped.utils.parsing`
