+++ doped/analysis.py
@@ -20,7 +20,6 @@ from monty.json import MontyDecoder
 from monty.serialization import dumpfn, loadfn
 from pymatgen.analysis.defects import core
 from pymatgen.analysis.structure_matcher import ElementComparator, StructureMatcher
-from pymatgen.core.composition import Element
 from pymatgen.core.sites import PeriodicSite
 from pymatgen.ext.matproj import MPRester
 from pymatgen.io.vasp.inputs import Poscar
@@ -35,13 +34,18 @@ from doped.utils.parsing import (
     _compare_incar_tags,
     _compare_kpoints,
     _compare_potcar_symbols,
+    _defect_charge_from_vasprun,
+    _defect_spin_degeneracy_from_vasprun,
+    _get_bulk_locpot_dict,
+    _get_bulk_site_potentials,
     _get_defect_supercell_bulk_site_coords,
     _get_output_files_and_check_if_multiple,
+    _multiple_files_warning,
+    _vasp_file_parsing_action_dict,
     check_atom_mapping_far_from_defect,
     get_defect_site_idxs_and_unrelaxed_structure,
     get_defect_type_and_composition_diff,
     get_locpot,
-    get_neutral_nelect_from_vasprun,
     get_orientational_degeneracy,
     get_outcar,
     get_vasprun,
@@ -346,90 +350,6 @@ def defect_name_from_structures(bulk_structure, defect_structure):
     return get_defect_name_from_defect(defect)
 
 
-def _defect_spin_degeneracy_from_vasprun(defect_vr: Vasprun, charge_state: int = 0) -> int:
-    """
-    Get the defect spin degeneracy from the vasprun output, assuming either
-    singlet (S=0) or doublet (S=1/2) behaviour.
-
-    Even-electron defects are assumed to have a singlet ground state, and odd-
-    electron defects are assumed to have a doublet ground state.
-    """
-    total_Z = int(
-        sum(Element(elt).Z * num for elt, num in defect_vr.final_structure.composition.as_dict().items())
-    )
-    return int((total_Z + charge_state) % 2 + 1)
-
-
-def _defect_charge_from_vasprun(bulk_vr: Vasprun, defect_vr: Vasprun, charge_state: Optional[int]):
-    """
-    Determine the defect charge state from the defect and bulk vaspruns, and
-    compare to the manually-set charge state if provided.
-    """
-    auto_charge = None
-
-    try:
-        if defect_vr.incar.get("NELECT") is None:
-            auto_charge = 0  # neutral defect if NELECT not specified
-
-        else:
-            defect_nelect = defect_vr.parameters.get("NELECT")
-            bulk_nelect = get_neutral_nelect_from_vasprun(bulk_vr)
-            neutral_defect_nelect = get_neutral_nelect_from_vasprun(defect_vr)
-
-            if bulk_vr.parameters.get("NELECT", False):
-                assert bulk_nelect == bulk_vr.parameters["NELECT"]
-
-            auto_charge = -1 * (defect_nelect - neutral_defect_nelect)
-
-            if auto_charge is None or abs(auto_charge) >= 10:
-                neutral_defect_nelect = get_neutral_nelect_from_vasprun(defect_vr, skip_potcar_init=True)
-                try:
-                    auto_charge = -1 * (defect_nelect - neutral_defect_nelect)
-
-                except Exception as e:
-                    auto_charge = None
-                    if charge_state is None:
-                        raise RuntimeError(
-                            "Defect charge cannot be automatically determined as POTCARs have not been "
-                            "setup with pymatgen (see Step 2 at "
-                            "https://github.com/SMTG-Bham/doped#installation). Please specify defect "
-                            "charge manually using the `charge_state` argument, or set up POTCARs with "
-                            "pymatgen."
[…]
