### 2024-03-09 f131ea9
Add `_parse_and_set_degeneracies` function to attempt (re-)parsing of defect degeneracy factors if not present (e.g. if generated with older versions, manually, with `pymatgen` etc), and ignore overflow warnings
Former-commit-id: 67a5e1420ebd0925cdd240f7bedcc277dc2587e1
