### 2024-03-09 47a3d34
Use `defect_entry._parse_and_set_degeneracies()` in thermodynamics analysis functions, and add tests for defect entries without symmetry/degeneracy info
