+++ doped/thermodynamics.py
@@ -36,11 +36,10 @@ from doped.utils.parsing import (
     _get_bulk_supercell,
     _get_defect_supercell_site,
     get_neutral_nelect_from_vasprun,
-    get_orientational_degeneracy,
     get_vasprun,
 )
 from doped.utils.plotting import _rename_key_and_dicts, _TLD_plot
-from doped.utils.symmetry import _get_all_equiv_sites, _get_sga, point_symmetry_from_defect_entry
+from doped.utils.symmetry import _get_all_equiv_sites, _get_sga
 
 
 def bold_print(string: str) -> None:
@@ -498,14 +497,14 @@ class DefectThermodynamics(MSONable):
                     band_gap_vals.append(defect_entry.calculation_metadata["gap"])
 
             # get the max difference in VBM & band_gap vals:
-            if max(vbm_vals) - min(vbm_vals) > 0.05 and self.vbm is None:
+            if vbm_vals and max(vbm_vals) - min(vbm_vals) > 0.05 and self.vbm is None:
                 _raise_VBM_band_gap_value_error(vbm_vals, type="VBM")
-            elif self.vbm is None:
+            elif vbm_vals and self.vbm is None:
                 self.vbm = vbm_vals[0]
 
-            if max(band_gap_vals) - min(band_gap_vals) > 0.05 and self.band_gap is None:
+            if band_gap_vals and max(band_gap_vals) - min(band_gap_vals) > 0.05 and self.band_gap is None:
                 _raise_VBM_band_gap_value_error(band_gap_vals, type="band_gap")
-            elif self.band_gap is None:
+            elif band_gap_vals and self.band_gap is None:
                 self.band_gap = band_gap_vals[0]
 
         for i, name in [(self.vbm, "VBM eigenvalue"), (self.band_gap, "band gap value")]:
@@ -667,7 +666,9 @@ class DefectThermodynamics(MSONable):
         """
         # determine defect charge transition levels:
         midgap_formation_energies = [  # without chemical potentials
-            entry.formation_energy(fermi_level=0.5 * self.band_gap, vbm=self.vbm)
+            entry.formation_energy(
+                fermi_level=0.5 * self.band_gap, vbm=entry.calculation_metadata.get("vbm", self.vbm)
+            )
             for entry in self.defect_entries
         ]
         # set range to {min E_form - 30, max E_form +30} eV for y (formation energy), and
@@ -777,11 +778,15 @@ class DefectThermodynamics(MSONable):
                 # confirm formation energies dominant for one defect over other identical defects
                 name_set = [entry.name for entry in sorted_defect_entries]
                 vb_list = [
-                    entry.formation_energy(fermi_level=limits[0][0], vbm=self.vbm)
+                    entry.formation_energy(
+                        fermi_level=limits[0][0], vbm=entry.calculation_metadata.get("vbm", self.vbm)
+                    )
                     for entry in sorted_defect_entries
                 ]
                 cb_list = [
-                    entry.formation_energy(fermi_level=limits[0][1], vbm=self.vbm)
+                    entry.formation_energy(
+                        fermi_level=limits[0][1], vbm=entry.calculation_metadata.get("vbm", self.vbm)
+                    )
                     for entry in sorted_defect_entries
                 ]
 
@@ -1224,7 +1229,11 @@ class DefectThermodynamics(MSONable):
 
         for defect_entry in self.defect_entries:
             formation_energy = defect_entry.formation_energy(
-                chempots=chempots, limit=limit, el_refs=el_refs, fermi_level=fermi_level, vbm=self.vbm
+                chempots=chempots,
+                limit=limit,
+                el_refs=el_refs,
+                fermi_level=fermi_level,
[…]
