### 2024-03-09 f2bd925
Update `README` with JOSS figures
