### 2024-03-09 e91e09f
Update `README` with JOSS figures
Former-commit-id: f7b0ad982f641a18963b49b6a17da7b28be9ab81 [formerly f2bd9256c5dd7f3a70fac1639f7a44e88370f273]
Former-commit-id: 9c2631b7770e7c30e1509ad5c41b08a4a4396c09
