+++ README.md
@@ -1,86 +1,0 @@
-[![Build status](https://github.com/SMTG-Bham/doped/actions/workflows/test.yml/badge.svg)](https://github.com/SMTG-Bham/doped/actions)
-[![Documentation Status](https://readthedocs.org/projects/doped/badge/?version=latest&style=flat)](https://doped.readthedocs.io/en/latest/)
-[![PyPI](https://img.shields.io/pypi/v/doped)](https://pypi.org/project/doped)
-[![Conda Version](https://img.shields.io/conda/vn/conda-forge/doped.svg)](https://anaconda.org/conda-forge/doped)
-[![Downloads](https://img.shields.io/pypi/dm/doped)](https://pypi.org/project/doped)
-
-<a href="https://doped.readthedocs.io/en/latest/"><img align="right" width="275" src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_v2_logo.png" alt="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping." title="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping."></a>`doped` is a python package for
-managing solid-state defect calculations, with functionality to
-generate defect structures and relevant competing phases (for chemical potentials), interface with
-[`ShakeNBreak`](https://shakenbreak.readthedocs.io) for
-[defect structure-searching](https://www.nature.com/articles/s41524-023-00973-1), write VASP input files for defect
-supercell calculations, and automatically parse and analyse the results.
-
-Tutorials showing the code functionality and usage are provided on the [docs](https://doped.readthedocs.io/en/latest/) site.
-
-### Example Outputs:
-Chemical potential/stability region plots and defect formation energy (a.k.a. transition level) diagrams:
-
-<a href="https://doped.readthedocs.io/en/latest/chemical_potentials_tutorial.html#analysing-and-visualising-the-chemical-potential-limits"><img align="left" width="365" src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_chempot_plotting.png"></a> <a href="https://doped.readthedocs.io/en/latest/dope_parsing_example.html#defect-formation-energy-transition-level-diagrams"><img align="right" width="385" src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_TLD_plot.png"></a>
-<br><br><br><br><br><br><br><br><br><br><br>
-
-
-## Installation
-```bash
-pip install doped  # install doped and dependencies
-```
-
-Alternatively if desired, `doped` can also be installed from `conda` with:
-
-```bash
-  conda install -c conda-forge doped
-```
-
-If you haven't done so already, you will need to set up your VASP `POTCAR` files and `Materials Project` API with `pymatgen` using the `.pmgrc.yaml` file, in order for `doped` to automatically generate VASP input files for defect calculations and determine competing phases for chemical potentials.
-See the docs [Installation](https://doped.readthedocs.io/en/latest/Installation.html) page for details on this.
-
-
-
-## `ShakeNBreak`
[…]
