### 2024-03-09 effe173
Update docs index page likewise
