+++ docs/index.rst
@@ -1,5 +1,5 @@
-doped
-=====
+``doped``
+=========
 
 .. image:: https://github.com/SMTG-Bham/doped/actions/workflows/test.yml/badge.svg
    :target: https://github.com/SMTG-Bham/doped/actions
@@ -14,40 +14,53 @@ doped
 
 .. raw:: html
 
-   <img src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_v2_logo.png" align="right" width="250" alt="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping." title="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping.">
+   <img src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_v2_logo.png" align="right" width="200" alt="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping." title="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping.">
 
-``doped`` is a Python package for managing solid-state defect calculations, with functionality to
-generate defect structures and relevant competing phases (for chemical potentials), interface with
-`ShakeNBreak`_ for defect structure-searching, write VASP input files for defect supercell calculations,
-and automatically parse and analyse the results.
+``doped`` is a Python software for the generation, pre-/post-processing and analysis of defect supercell
+calculations, implementing the defect simulation workflow in an efficient, reproducible, user-friendly yet
+powerful and fully-customisable manner.
 
-Example Outputs:
------------------
-Chemical potential/stability region plots and defect formation energy (a.k.a. transition level) diagrams:
+Tutorials showing the code functionality and usage are provided on the :ref:`Tutorials` page, and an
+overview of the key advances of the package is given in the
+`JOSS paper <https://github.com/openjournals/joss-reviews/issues/6433>`__.
 
-.. image:: doped_chempot_plotting.png
-   :align: left
-   :width: 305
-.. image:: doped_TLD_plot.png
-   :align: right
-   :width: 325
+.. update JOSS paper link when done!
 
 .. raw:: html
 
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
+    <a href="https://doi.org/10.21105/joss.06433"><img class="center" width="800" src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/JOSS/doped_JOSS_workflow_figure.png"></a>
 
-Tutorials showing the code functionality and usage are provided on the :ref:`Tutorials` page.
+Key Features
+============
+All features and functionality are fully-customisable:
 
+- **Supercell Generation**: Generate an optimal supercell, maximising periodic image separation for the minimum number of atoms (computational cost).
+- **Defect Generation**: Generate defect supercells and guess likely charge states based on chemical intuition.
+- **Calculation I/O**: Automatically write inputs & parse supercell calculations (``VASP`` & other DFT/force-field codes).
+- **Chemical Potentials**: Determine relevant competing phases for chemical potential limits, with automated calculation setup, parsing and analysis.
+- **Defect Analysis**: Automatically parse calculation outputs to compute defect formation energies, finite-size corrections (FNV & eFNV), symmetries, degeneracies, transition levels, etc.
[…]
