+++ docs/Dev_ToDo.md
@@ -22,11 +22,6 @@
 ## Housekeeping
 - Tutorials general structure clean-up
 - Clean up repo, removing old unnecessary git blobs
-- Clean `README` with bullet-point summary of key features, and sidebar like `SnB`. Add correction plots and other example outputs (see MRS poster for this).
-- Code tidy up:
-  - Test coverage?
-  - Generate docs locally and check warnings
-  - Add type hints for all functions.
 
 - Docs:
   - Update note at end of thermo tutorial to link to py-sc-fermi/doped interface.
