+++ docs/index.rst
@@ -1,189 +1,0 @@
-doped
-=====
-
-.. image:: https://github.com/SMTG-Bham/doped/actions/workflows/test.yml/badge.svg
-   :target: https://github.com/SMTG-Bham/doped/actions
-.. image:: https://readthedocs.org/projects/doped/badge/?version=latest&style=flat
-   :target: https://doped.readthedocs.io/en/latest/
-.. image:: https://img.shields.io/pypi/v/doped
-   :target: https://pypi.org/project/doped
-.. image:: https://img.shields.io/conda/vn/conda-forge/doped.svg
-   :target: https://anaconda.org/conda-forge/doped
-.. image:: https://img.shields.io/pypi/dm/doped
-   :target: https://pypi.org/project/doped
-
-.. raw:: html
-
-   <img src="https://raw.githubusercontent.com/SMTG-Bham/doped/main/docs/doped_v2_logo.png" align="right" width="250" alt="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping." title="Schematic of a doped (defect-containing) crystal, inspired by the biological analogy to (semiconductor) doping.">
-
-``doped`` is a Python package for managing solid-state defect calculations, with functionality to
-generate defect structures and relevant competing phases (for chemical potentials), interface with
-`ShakeNBreak`_ for defect structure-searching, write VASP input files for defect supercell calculations,
-and automatically parse and analyse the results.
-
-Example Outputs:
------------------
-Chemical potential/stability region plots and defect formation energy (a.k.a. transition level) diagrams:
-
-.. image:: doped_chempot_plotting.png
-   :align: left
-   :width: 305
-.. image:: doped_TLD_plot.png
-   :align: right
-   :width: 325
-
-.. raw:: html
-
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-   <br>
-
-Tutorials showing the code functionality and usage are provided on the :ref:`Tutorials` page.
-
-``doped`` can be installed via PyPI (``pip install doped``) or ``conda`` if preferred, and further
-instructions for setting up ``POTCAR`` files with ``pymatgen`` (needed for input file generation), if not
-already done, are provided on the :ref:`Installation` page.
-
-
-``ShakeNBreak``
-================
-As shown in the tutorials, it is highly recommended to use the `ShakeNBreak`_ approach when calculating
-point defects in solids, to ensure you have identified the ground-state structures of your defects. As
-detailed in the `theory paper`_, skipping this step can result in drastically incorrect formation
-energies, transition levels, carrier capture (basically any property associated with defects). This
-approach is followed in the :ref:`tutorials <Tutorials>`, with a more in-depth explanation and tutorial
-given on
-the
-`ShakeNBreak`_ docs.
-
-.. _theory paper: https://www.nature.com/articles/s41524-023-00973-1
-.. _tutorials: https://www.nature.com/articles/s41524-023-00973-1 # TODO!!!
-
-Summary GIF:
-
-.. image:: https://raw.githubusercontent.com/SMTG-Bham/ShakeNBreak/main/docs/SnB_Supercell_Schematic_PES_2sec_Compressed.gif
-
-``SnB`` CLI Usage:
-
[…]
