### 2024-03-11 ec5f86d
Update `pydefect` and `vise` requirements
Former-commit-id: 0811ce604cd0a544ded513cc23e0be342a25f89d [formerly 6870b2beb386ae5ff6bab0ed859fbe4bdd41f72b]
Former-commit-id: 9d4f229a5614758a842ffa03d65d785a9ecb3208
