+++ docs/requirements.txt
@@ -9,7 +9,8 @@ monty
 shakenbreak>=3.3.1
 packaging
 pandas>=1.1.0
-pydefect
+pydefect>=0.8.1
+vise>=0.8.1
 
 # optional dependencies:
 py-sc-fermi
