+++ docs/docs_requirements.txt
@@ -21,4 +21,5 @@ monty
 shakenbreak>=3.3.1
 packaging
 pandas>=1.1.0
-pydefect
+pydefect>=0.8.1
+vise>=0.8.1
