+++ pyproject.toml
@@ -25,8 +25,9 @@ dependencies = [
     "pymatgen-analysis-defects>=2023.8.22",
     "shakenbreak>=3.3.1",
     "pandas",
-    "pydefect",  # for Kumagai (eFNV) correction
+    "pydefect>=0.8.1",  # for Kumagai (eFNV) correction & PHS analysis
     "filelock",
+    "vise>=0.8.1",  # for PHS analysis
 ]
 
 [project.urls]
