+++ doped/utils/phs.py
@@ -1,23 +1,17 @@
 """
 Helper functions for setting up PHS analysis.
 
-Contains modified versions of functions from pydefect
-https://github.com/kumagai-group/pydefect
-and vise
-https://github.com/kumagai-group/vise, to avoid the user requiring additional files i.e. PROCAR.
+Contains modified versions of functions from pydefect (https://github.com/kumagai-group/pydefect)
+and vise (https://github.com/kumagai-group/vise), to avoid requiring additional files (i.e. ``PROCAR``s).
 """
+
 import logging
 import os
-import warnings
 from importlib.metadata import version
-from typing import Any, List, Optional
+from typing import Any, Optional
 
 import matplotlib.pyplot as plt
 import numpy as np
-from vise import user_settings
-# suppress pydefect INFO messages
-user_settings.logger.setLevel(logging.CRITICAL)
-
 from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
 from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
 from pydefect.analyzer.make_band_edge_states import make_band_edge_states
@@ -26,22 +20,29 @@ from pydefect.cli.vasp.make_band_edge_orbital_infos import calc_orbital_characte
 from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
 from pydefect.defaults import defaults
 from pymatgen.electronic_structure.core import Spin
-from pymatgen.io.vasp.outputs import Vasprun
+from pymatgen.io.vasp.outputs import Outcar, Vasprun
+from vise import user_settings
 from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
 
 from doped.utils.plotting import _get_backend, format_defect_name
+
+# suppress pydefect INFO messages
+user_settings.logger.setLevel(logging.CRITICAL)
+
 
 def make_band_edge_orbital_infos(
     defect_vr: Vasprun, vbm: float, cbm: float, str_info, eigval_shift: float = 0.0
 ):
     """
-    Make BandEdgeOrbitalInfos from vasprun.xml. Modified from my pydefects to
-    use projected orbitals stored in vasprun.xml.
+    Make BandEdgeOrbitalInfos from vasprun.xml.
+
+    Modified from my pydefect to use projected
+    orbitals stored in vasprun.xml.
 
     Args:
         defect_vr: defect Vasprun object
         vbm: vbm_info from get_edge_info
-        cbm: cbm_info fomr get_edge_info
+        cbm: cbm_info from get_edge_info
         str_info: pydefect DefectStructureInfo
         eigval_shift (float): Shift eigenvalues by E-E_VBM to set VBM at 0 eV
 
@@ -64,7 +65,7 @@ def make_band_edge_orbital_infos(
     upper_idx = np.argwhere(min_energy_by_band < cbm + eigval_range)[-1][-1]
 
     orbs, s = defect_vr.projected_eigenvalues, defect_vr.final_structure
-    orb_infos: List[Any] = []
+    orb_infos: list[Any] = []
     for spin, eigvals in defect_vr.eigenvalues.items():
         orb_infos.append([])
         for k_idx in range(len(kpt_coords)):
@@ -88,47 +89,51 @@ def make_band_edge_orbital_infos(
     )
 
 
-def get_band_edge_info(DefectParser, bulk_vr, bulk_outcar, defect_vr):
+def get_band_edge_info(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar):
     """
-    Load metadata required for performing phs identification.
+    Load metadata required for performing PHS identification.
+
[…]
