### 2024-03-11 321d226
Refactor Cu2SiSe3 data to example data, and some minor updates
