+++ tests/test_analysis.py
@@ -70,7 +70,7 @@ class DefectsParsingTestCase(unittest.TestCase):
         self.Sb2Se3_DATA_DIR = os.path.join(self.module_path, "data/Sb2Se3")
         self.Sb2Se3_dielectric = np.array([[85.64, 0, 0], [0.0, 128.18, 0], [0, 0, 15.00]])
 
-        self.Cu2SiSe3_DATA_DIR = os.path.join(self.module_path, "data/Cu2SiSe3")
+        self.Cu2SiSe3_EXAMPLE_DIR = os.path.join(self.module_path, "../examples/Cu2SiSe3")
 
         self.Sb2Si2Te6_dielectric = [44.12, 44.12, 17.82]
         self.Sb2Si2Te6_DATA_DIR = os.path.join(self.EXAMPLE_DIR, "Sb2Si2Te6")
@@ -714,14 +714,14 @@ class DefectsParsingTestCase(unittest.TestCase):
         """
         # Test the loading of Cu2SiSe3 vacancy
         defect = DefectParser.from_paths(
-            f"{self.Cu2SiSe3_DATA_DIR}/v_Cu_0/vasp_std",
-            f"{self.Cu2SiSe3_DATA_DIR}/bulk/vasp_std",
+            f"{self.Cu2SiSe3_EXAMPLE_DIR}/v_Cu_0/vasp_std",
+            f"{self.Cu2SiSe3_EXAMPLE_DIR}/bulk/vasp_std",
             skip_corrections=True,
             load_phs_data=True,
         ).defect_entry
 
         bes, fig = defect.get_perturbed_host_state()  # Test plotting KS
-        with open(f"{self.Cu2SiSe3_DATA_DIR}/Cu2SiSe3_band_edge_states.json") as f:
+        with open(f"{self.Cu2SiSe3_EXAMPLE_DIR}/Cu2SiSe3_band_edge_states.json") as f:
             expected = json.load(f)
         assert bes.as_dict() == expected
 
