### 2024-03-11 38c7956
Add `py-sc-fermi` interface tutorial to docs
