### 2024-03-11 5edcfcc
Update `fermi_solver.py`; run import checks and warning only when `py-sc-fermi` functions needed, code/docstrings cleanup, formatting
