+++ doped/interface/fermi_solver.py
@@ -1,28 +1,28 @@
-from typing import Optional, Union, List
-from warnings import warn
-from itertools import product
+import functools
 from copy import deepcopy
-import numpy as np
-from joblib import Parallel, delayed
-from typing import Dict, Any
+from itertools import product
+from typing import TYPE_CHECKING, Any, Optional, Union
+from warnings import warn
 
-from scipy.spatial import ConvexHull, Delaunay
-from scipy.interpolate import griddata
+import numpy as np
 import pandas as pd
+from joblib import Parallel, delayed
 from monty.json import MSONable
 from pymatgen.electronic_structure.dos import FermiDos
 from pymatgen.io.vasp import Vasprun
+from scipy.interpolate import griddata
+from scipy.spatial import ConvexHull, Delaunay
 
-from doped.thermodynamics import DefectThermodynamics
 from doped.utils.parsing import get_neutral_nelect_from_vasprun
 
-try:
-    from py_sc_fermi.defect_system import DefectSystem
-    from py_sc_fermi.defect_species import DefectSpecies
-    from py_sc_fermi.defect_charge_state import DefectChargeState
-    from py_sc_fermi.dos import DOS
-except ImportError:
-    warn("py-sc-fermi not installed, will only be able to use doped backend")
+if TYPE_CHECKING:
+    import contextlib
+
+    from doped.thermodynamics import DefectThermodynamics
+
+    with contextlib.suppress(ImportError):
+        from py_sc_fermi.defect_species import DefectSpecies
+        from py_sc_fermi.defect_system import DefectSystem
 
 
 def _get_label_and_charge(name: str) -> tuple:
@@ -41,33 +41,39 @@ def _get_label_and_charge(name: str) -> tuple:
     return label, int(charge)
 
 
-class FermiSolver(MSONable):
-    """class to manage the generation of DefectSystems from DefectPhaseDiagrams"""
+# TODO: Add link to Beth & Jiayi paper in py-sc-fermi tutorial, showing the utility of the effective
+#  dopant concentration analysis
 
-    def __init__(self, defect_thermodynamics: DefectThermodynamics, bulk_dos_vr: str):
+
+class FermiSolver(MSONable):
+    """
+    Class to manage the calculation of self-consistent Fermi levels and defect/
+    carrier concentrations from ``DefectThermodynamics`` objects.
+    """
+
+    def __init__(self, defect_thermodynamics: "DefectThermodynamics", bulk_dos_vr: str):
         self.defect_thermodynamics = defect_thermodynamics
         self.bulk_dos = bulk_dos_vr
+        self._not_implemented_message = "This method is implemented in the derived class, use FermiSolverDoped or FermiSolverPyScFermi instead."
 
     def equilibrium_solve(self) -> None:
-        """not implemented in the base class, implemented in the derived class"""
-        raise NotImplementedError(
-            """This method is implemented in the derived class, 
-            use FermiSolverDoped or FermiSolverPyScFermi instead."""
-        )
+        """
+        Not implemented in the base class, implemented in the derived class.
+        """
+        raise NotImplementedError(self._not_implemented_message)
 
     def pseudo_equilibrium_solve(self) -> None:
-        """not implemented in the base class, implemented in the derived class"""
-        raise NotImplementedError(
-            """This method is implemented in the derived class, 
-            use FermiSolverDoped or FermiSolverPyScFermi instead."""
-        )
+        """
+        Not implemented in the base class, implemented in the derived class.
+        """
+        raise NotImplementedError(self._not_implemented_message)
 
     def scan_temperature(
         self,
         chempots: dict[str, float],
-        temperature_range: Union[float, List[float]],
-        annealing_temperature_range: Optional[Union[float, List[float]]] = None,
[…]
