### 2024-03-11 c9b8c98
Resolve `TODO` in `thermodynamics`
