+++ doped/core.py
@@ -39,12 +39,12 @@ _orientational_degeneracy_warning = (
 @dataclass
 class DefectEntry(thermo.DefectEntry):
     """
-    Subclass of pymatgen.analysis.defects.thermo.DefectEntry with additional
-    attributes used by doped.
+    Subclass of ``pymatgen.analysis.defects.thermo.DefectEntry`` with
+    additional attributes used by doped.
 
     Core Attributes:
         defect:
-            doped/pymatgen defect object corresponding to the defect in the entry.
+            ``doped``/``pymatgen`` defect object corresponding to the defect in the entry.
         charge_state:
             Charge state of the defect.
         sc_entry:
@@ -80,8 +80,8 @@ class DefectEntry(thermo.DefectEntry):
 
     Generation Attributes:
         name:
-            The doped-generated name of the defect entry. See docstrings of
-            DefectsGenerator for the doped naming algorithm.
+            The ``doped``-generated name of the defect entry. See docstrings
+            of ``DefectsGenerator`` for the doped naming algorithm.
         conventional_structure:
             Conventional cell structure of the host according to the Bilbao
             Crystallographic Server (BCS) definition, used to determine defect
@@ -101,14 +101,14 @@ class DefectEntry(thermo.DefectEntry):
             A log of the input & computed values used to determine charge state
             probabilities.
         defect_supercell:
-            pymatgen Structure object of the defect supercell.
+            ``pymatgen`` ``Structure`` object of the defect supercell.
         defect_supercell_site:
-            pymatgen PeriodicSite object of the defect in the defect supercell.
+            ``pymatgen`` ``PeriodicSite`` object of the defect in the defect supercell.
         equivalent_supercell_sites:
-            List of pymatgen PeriodicSite objects of symmetry-equivalent defect
+            List of ``pymatgen`` ``PeriodicSite`` objects of symmetry-equivalent defect
             sites in the defect supercell.
         bulk_supercell:
-            pymatgen Structure object of the bulk (pristine, defect-free) supercell.
+            ``pymatgen`` ``Structure`` object of the bulk (pristine, defect-free) supercell.
     """
 
     # core attributes:
@@ -184,13 +184,13 @@ class DefectEntry(thermo.DefectEntry):
 
     def to_json(self, filename: Optional[str] = None):
         """
-        Save the DefectEntry object to a json file, which can be reloaded with
-        the DefectEntry.from_json() class method.
+        Save the ``DefectEntry`` object to a json file, which can be reloaded
+        with the ``DefectEntry.from_json()`` class method.
 
         Args:
             filename (str):
                 Filename to save json file as. If None, the filename will
-                be set as "{DefectEntry.name}.json".
+                be set as ``"{DefectEntry.name}.json"``.
         """
         if filename is None:
             filename = f"{self.name}.json"
@@ -200,20 +200,20 @@ class DefectEntry(thermo.DefectEntry):
     @classmethod
     def from_json(cls, filename: str):
         """
-        Load a DefectEntry object from a json file.
+        Load a ``DefectEntry`` object from a json file.
 
         Args:
             filename (str):
-                Filename of json file to load DefectEntry from.
+                Filename of json file to load ``DefectEntry`` from.
 
         Returns:
-            DefectEntry object
+            ``DefectEntry`` object
         """
         return loadfn(filename)
 
     def as_dict(self) -> dict:
         """
-        Returns a dictionary representation of the DefectEntry object.
+        Returns a dictionary representation of the ``DefectEntry`` object.
         """
         # ignore warning about oxidation states not summing to Structure charge:
[…]
