### 2024-03-12 #6433 @kavanase
@yuan-gist just fyi, I think for the editorialbot to generate your checklist, you need to do the `@editorialbot generate my checklist` command that you did above, but without any code formatting
