### 2024-03-12 7e4dd57
Add `plot` and `style_file` options, efficiency/formatting updates, and add test
Former-commit-id: 7a615000674ce3a3b678d9f69073003ef34b69de
