### 2024-03-12 7a61500
Add `plot` and `style_file` options, efficiency/formatting updates, and add test
