+++ doped/utils/phs.py
@@ -21,10 +21,11 @@ from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
 from pydefect.defaults import defaults
 from pymatgen.electronic_structure.core import Spin
 from pymatgen.io.vasp.outputs import Outcar, Vasprun
+from shakenbreak.plotting import _install_custom_font
 from vise import user_settings
 from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
 
-from doped.utils.plotting import _get_backend, format_defect_name
+from doped.utils.plotting import _get_backend
 
 # suppress pydefect INFO messages
 user_settings.logger.setLevel(logging.CRITICAL)
@@ -34,16 +35,16 @@ def make_band_edge_orbital_infos(
     defect_vr: Vasprun, vbm: float, cbm: float, str_info, eigval_shift: float = 0.0
 ):
     """
-    Make BandEdgeOrbitalInfos from vasprun.xml.
+    Make ``BandEdgeOrbitalInfos`` from ``vasprun.xml``.
 
-    Modified from my pydefect to use projected
-    orbitals stored in vasprun.xml.
+    Modified from ``pydefect`` to use projected
+    orbitals stored in ``vasprun.xml``.
 
     Args:
-        defect_vr: defect Vasprun object
-        vbm: vbm_info from get_edge_info
-        cbm: cbm_info from get_edge_info
-        str_info: pydefect DefectStructureInfo
+        defect_vr: defect ``Vasprun`` object
+        vbm: VBM eigenvalue in eV
+        cbm: CBM eigenvalue in eV
+        str_info: ``pydefect`` ``DefectStructureInfo``
         eigval_shift (float): Shift eigenvalues by E-E_VBM to set VBM at 0 eV
 
     Returns:
@@ -147,20 +148,35 @@ def get_band_edge_info(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar
     return band_orb, vbm_info, cbm_info
 
 
-def get_phs_and_eigenvalue(DefectEntry, filename: Optional[str] = None, ks_labels: bool = False):
+def get_phs_and_eigenvalue(
+    DefectEntry,
+    filename: Optional[str] = None,
+    plot: bool = True,
+    ks_labels: bool = False,
+    style_file: Optional[str] = None,
+):
     """
-    Get PHS info and eigenvalue plot for a given DefectEntry.
+    Get PHS info and eigenvalue plot (if ``plot = True``; default) for a given
+    ``DefectEntry``.
 
     Args:
-        DefectEntry: DefectEntry object
+        DefectEntry: ``DefectEntry`` object
         filename (str):
             Filename to save the Kumagai site potential plots to.
-            If None, plots are not saved.
+            If None (default), plots are not saved.
         ks_labels (bool):
             Add the band index to the KS levels.
+            (Default: False)
+        plot (bool):
+            Whether to plot the single-particle eigenvalues.
+            (Default: True)
+        style_file (str):
+            Path to a ``mplstyle`` file to use for the plot. If None
+            (default), uses the ``doped`` displacement plot style
+            (``doped/utils/displacement.mplstyle``).
 
     Returns:
-        pydefect PerfectBandEdgeState class
+        ``pydefect`` ``PerfectBandEdgeState`` class
     """
     band_orb = DefectEntry.calculation_metadata["phs_data"]["band_orb"]
     vbm_info = DefectEntry.calculation_metadata["phs_data"]["vbm_info"]
@@ -169,11 +185,15 @@ def get_phs_and_eigenvalue(DefectEntry, filename: Optional[str] = None, ks_label
[…]
