### 2024-03-13 9a54f8d
Suggested: Refactor `fix_defect_species` to `fix_charge_states` as this is more intuitive?
Former-commit-id: 0371f33e42a6dc3535497af950702334b36980af
