### 2024-03-13 a40a2d7
Add monkey patch to ensure defect site always included in neighbour list. Changed ordering of plot legend
