+++ doped/utils/phs.py
@@ -7,6 +7,7 @@ and vise (https://github.com/kumagai-group/vise), to avoid requiring additional 
 
 import logging
 import os
+from collections import defaultdict
 from importlib.metadata import version
 from typing import Any, Optional
 
@@ -19,6 +20,7 @@ from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
 from pydefect.cli.vasp.make_band_edge_orbital_infos import calc_orbital_character, calc_participation_ratio
 from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
 from pydefect.defaults import defaults
+from pydefect.util.structure_tools import Coordination, Distances
 from pymatgen.electronic_structure.core import Spin
 from pymatgen.io.vasp.outputs import Outcar, Vasprun
 from shakenbreak.plotting import _install_custom_font
@@ -128,6 +130,30 @@ def get_band_edge_info(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar
     vbm_info = get_edge_info(band_edge_prop.vbm_info, orbs, s, bulk_vr)
     cbm_info = get_edge_info(band_edge_prop.cbm_info, orbs, s, bulk_vr)
 
+    # Money patch to set include_on_site = True
+    _orig_method = Distances.coordination
+
+    def coordination(self, include_on_site=True, cutoff_factor=None) -> "Coordination":
+        cutoff_factor = cutoff_factor or defaults.cutoff_distance_factor
+        cutoff = self.shortest_distance * cutoff_factor
+        elements = [element.specie.name for element in self.structure]
+        e_d = zip(elements, self.distances(remove_self=False))
+
+        unsorted_distances = defaultdict(list)
+        neighboring_atom_indices = []
+        for i, (element, distance) in enumerate(e_d):
+            if distance < cutoff and include_on_site:
+                unsorted_distances[element].append(round(distance, 2))
+                neighboring_atom_indices.append(i)
+
+        distance_dict = {}
+        for element, distances in unsorted_distances.items():
+            distance_dict[element] = sorted(distances)
+
+        return Coordination(distance_dict, round(cutoff, 3), neighboring_atom_indices)
+
+    Distances.coordination = coordination
+
     dsinfo = MakeDefectStructureInfo(  # Using default values suggested by pydefect
         bulk_vr.final_structure,
         defect_vr.final_structure,
@@ -136,6 +162,9 @@ def get_band_edge_info(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar
         dist_tol=1.0,
         neighbor_cutoff_factor=1.3,  # Neighbors are sites within min_dist * neighbor_cutoff_factor
     )
+
+    # Undo monkey patch in case used in other parts of the code:
+    Distances.coordination = _orig_method
 
     band_orb = make_band_edge_orbital_infos(
         defect_vr,
@@ -257,8 +286,8 @@ def get_phs_and_eigenvalue(
 
         ax.set_ylim([ymin - 0.25, ymax + 0.75])
         # add a point at 0,-5 with the color range and label unoccopied states
-        ax.scatter(0, -5, label="Occupied", color=(0.22, 0.325, 0.643))
         ax.scatter(0, -5, label="Unoccupied", color=(0.98, 0.639, 0.086))
+        ax.scatter(0, -5, label="Occupied", color=(0.22, 0.325, 0.643))
         if partial:
             ax.scatter(0, -5, label="Partially Occupied", color=(0, 0.5, 0))
         ax.axhline(-5, 0, 1, color="black", linewidth=0.5, linestyle="-.", label="Band Edges")
