### 2024-03-13 b206448
Redefine `element_changes` for `Defect` for fast formation energy / concentration calculations
Former-commit-id: eaf96cea509da0e5aff0816a13741d4e7c512bd0
