### 2024-03-13 eaf96ce
Redefine `element_changes` for `Defect` for fast formation energy / concentration calculations
