### 2024-03-13 b028056
Make `_get_chempot_term` far more efficient (realised was bottleneck in `doped` concentration calcs, due to inefficient `defect.element_changes` in `pymatgen`; avoided now)
Former-commit-id: 616126a6afd160978a4a24947264801bd873b641
