### 2024-03-13 9901b68
Add `effective_dopant_concentration` parameter to `doped` thermodynamics functions
