+++ doped/thermodynamics.py
@@ -1138,6 +1138,7 @@ class DefectThermodynamics(MSONable):
         per_charge: bool = True,
         per_site: bool = False,
         skip_formatting: bool = False,
+        skip_formation_energy: bool = False,
     ) -> pd.DataFrame:
         r"""
         Compute the `equilibrium` concentrations (in cm^-3) for all
@@ -1215,6 +1216,10 @@ class DefectThermodynamics(MSONable):
             skip_formatting (bool):
                 Whether to skip formatting the defect charge states and concentrations as
                 strings (and keep as ints and floats instead). (default: False)
+            skip_formation_energy (bool):
+                Whether to include the defect formation energies in the output ``DataFrame``.
+                Mainly intended for internal use, to speed up thermodynamics calculations.
+                (default: False)
 
         Returns:
             pandas DataFrame of defect concentrations (and formation energies) for each
@@ -1228,13 +1233,14 @@ class DefectThermodynamics(MSONable):
         energy_concentration_list = []
 
         for defect_entry in self.defect_entries:
-            formation_energy = defect_entry.formation_energy(
-                chempots=chempots,
-                limit=limit,
-                el_refs=el_refs,
-                fermi_level=fermi_level,
-                vbm=defect_entry.calculation_metadata.get("vbm", self.vbm),
-            )
+            if not skip_formation_energy:
+                formation_energy = defect_entry.formation_energy(
+                    chempots=chempots,
+                    limit=limit,
+                    el_refs=el_refs,
+                    fermi_level=fermi_level,
+                    vbm=defect_entry.calculation_metadata.get("vbm", self.vbm),
+                )
             concentration = defect_entry.equilibrium_concentration(
                 chempots=chempots,
                 limit=limit,
@@ -1253,7 +1259,9 @@ class DefectThermodynamics(MSONable):
                         if skip_formatting
                         else f"{'+' if defect_entry.charge_state > 0 else ''}{defect_entry.charge_state}"
                     ),
-                    "Formation Energy (eV)": round(formation_energy, 3),
+                    "Formation Energy (eV)": (
+                        "N/A" if skip_formation_energy else round(formation_energy, 3)
+                    ),
                     "Raw Concentration": concentration,
                     "Concentration (per site)" if per_site else "Concentration (cm^-3)": (
                         concentration if skip_formatting else f"{concentration:.3e}"
@@ -1310,6 +1318,39 @@ class DefectThermodynamics(MSONable):
             )
         return FermiDos(bulk_dos_vr.complete_dos, nelecs=get_neutral_nelect_from_vasprun(bulk_dos_vr))
 
+    def _add_effective_dopant_concentration(
+        self, conc_df: pd.DataFrame, effective_dopant_concentration: Optional[float] = None
+    ):
+        """
+        Add the effective dopant concentration to the concentration
+        ``DataFrame``.
+        """
+        if effective_dopant_concentration is not None:
+            eff_dopant_df = pd.DataFrame(
+                {
+                    "Defect": "Effective Dopant",
+                    "Charge": np.sign(effective_dopant_concentration),
+                    "Formation Energy (eV)": "N/A",
+                    "Concentration (cm^-3)": np.abs(effective_dopant_concentration),
+                    "Charge State Population": "100.0%",
+                },
+                index=[0],
+            )
+            for col in conc_df.columns:
+                if col not in eff_dopant_df.columns:
+                    eff_dopant_df[col] = "N/A"  # e.g. concentration per site, if per_site=True
+
+            if "Charge" not in conc_df.columns:
+                eff_dopant_df = eff_dopant_df.drop(
+                    columns=["Charge", "Formation Energy (eV)", "Charge State Population"]
+                )
[…]
