### 2024-03-13 b1a5853
Add `effective_dopant_concentration` parameter to `FermiSolverDoped` functions, and fix bug in `generate_annealed_defect_system`
Former-commit-id: c5a25e1f93cb751827bdbe7dd80099ea9beef14d
