### 2024-03-13 4924ada
Use environment rather than `set-output` commands to avoid GitHub deprecation
