+++ .github/workflows/release.yml
@@ -40,30 +40,30 @@ jobs:
           fi
 
       - name: Build packages
-        if: steps.check_version.outputs.version_changed == 'true'
+        if: env.version_changed == 'true'
         run: |
           python3 -m pip install --upgrade build
           python3 -m build
 
       - name: Publish
-        if: steps.check_version.outputs.version_changed == 'true'
+        if: env.version_changed == 'true'
         uses: pypa/gh-action-pypi-publish@master
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
 
       - name: Write release info
-        if: steps.check_version.outputs.version_changed == 'true'
+        if: env.version_changed == 'true'
         run: |
           awk 'BEGIN {p = 0} {a = 0 }; /^v\d*.\d*.\d*./ { p += 1; a = 1}; p + a == 1 { print } ' CHANGELOG.rst | sed -e '1,1d' | sed  -e '/./,$!d' -e :a -e '/^\n*$/{$d;N;ba' -e '}' > release_info.txt
 
       - name: Get Release Tag
-        if: steps.check_version.outputs.version_changed == 'true'
+        if: env.version_changed == 'true'
         run: |
           VERSION=$(grep version pyproject.toml | awk '{print $NF}' | tr -d '"' | head -1)
           echo "version=${VERSION}" >> $GITHUB_ENV
 
       - name: Release
-        if: steps.check_version.outputs.version_changed == 'true'
+        if: env.version_changed == 'true'
         uses: actions/create-release@v1
         id: create_release
 
