### 2024-03-14 470d377
Update parsing code so that bulk vasprun only loaded once
Former-commit-id: 15cd98f12eecae5f4082a1bfa40dfc10e364b0b2
