### 2024-03-14 44686cc
Add cherry-picked commit from `shallow_defect_analysis` for bulk vasprun loading by @kavanase
Former-commit-id: e172c151eedd3f06d986142c4c2f8dea079bae25
