### 2024-03-14 7043157
Make vasprun data parsing more efficient and reduce memory load
Former-commit-id: a6d0277a8f4c411ccfdd183305f1430f097407c8
