### 2024-03-14 ec621d3
Add monkey-patched `Vasprun._parse_projected_eigen()` to make parsing more efficient with `load_phs_data=True`
