### 2024-03-14 95972ec
Fix `spglib` to `2.0.2` for the moment, as it avoids unnecessary debug warnings at C level (hard to catch without `subprocess` slowing things down...). Issue tracked here: https://github.com/spglib/spglib/issues/338
Former-commit-id: 457088291b8fbf9e15647634c3f530e28b8951c7 [formerly f79c02629baf4b4d4392a2f1096552d569834570]
Former-commit-id: 4c023b54d459bab36f7c1b5a3b6b75a7466374ad
