+++ pyproject.toml
@@ -22,6 +22,8 @@ dependencies = [
     "matplotlib>=3.5.2",
     "numpy>=1.21.0",
     "pymatgen>=2023.11.12",  # limiting factor: POTCAR_STATS_PATH available from pymatgen.io.vasp.inputs
+    "spglib==2.0.2",  # dependent of pymatgen, but explicitly set to 2.0.2 here to avoid unnecessary warnings
+    # see https://github.com/spglib/spglib/issues/338; remove this limit when issue resolved
     "pymatgen-analysis-defects>=2023.8.22",
     "shakenbreak>=3.3.1",
     "pandas",
