### 2024-03-14 f79c026
Fix `spglib` to `2.0.2` for the moment, as it avoids unnecessary debug warnings at C level (hard to catch without `subprocess` slowing things down...). Issue tracked here: https://github.com/spglib/spglib/issues/338
