### 2024-03-14 SMTG-Bham/doped#55 @yuan-gist
JOSS review issues #1
**I will ad a number of issues seen from Review by tomorrow. There are just minor issues.**

In this issue, I used `doped` to parse my old calculations (trying to use doped to get degeneracy factors for defect densities):
`CaO_bulk  v_Ca_0  v_Ca_+1  v_Ca_-1  v_Ca_-2`

The jupyter-notebook is found: https://github.com/yuan-gist/doped_review/blob/main/doped.ipynb

The calculations are found: the file sizes are too large, can I put on Google drive and share via email?

The issue is that it seems ` v_Ca_+1 ` is not parsed. I also tested other cases. it seems only one type charge states can be parsed.
