### 2024-03-14 #6433 @yuan-gist
I will continuously update my review comments here:

 [Using doped to parse calculations not prepared by doped](https://github.com/SMTG-Bham/doped/issues/55)
