### 2024-03-14 ba8df59
Handle oxi state guessing in parsing similar to generation, trying without composition reduction, then with, both with a 5 second timeout and just setting as 'Undetermined' if still not parsed (not necessary for parsing anyway, and can hang with large complex structures). Flagged by @utf 🙌
