### 2024-03-14 e2cced1
Handle oxi state guessing in parsing similar to generation, trying without composition reduction, then with, both with a 5 second timeout and just setting as 'Undetermined' if still not parsed (not necessary for parsing anyway, and can hang with large complex structures). Flagged by @utf 🙌
Former-commit-id: ba8df59ca74ca2d724b2e9105507604717477176
