### 2024-03-14 SMTG-Bham/doped#55 @kavanase
Hi @yuan-gist,
Thanks for raising, that's odd! I haven't seen this issue before (and in general `DefectsParser` should warn if any defects haven't been successfully parsed and why.
Yes if you could share the output files to reproduce that would be great.
Emailing a Google drive / Dropbox /OneDrive link is perfect.
Just to note, doped can parse gzipped files (`vasprun.xml.gz` etc) to save space (so could possibly also just upload to your GitHub repo as this, as the max file size when uploaded by the CLI is 100 Mb which I guess should be enough?
Thanks!
