### 2024-03-15 SMTG-Bham/doped#55 @kavanase
Hi @yuan-gist, email now received!
Looking at the folder structure you sent, I can spot the issue.
So in all your defect folders except `v_Ca_+1`, the defect output is in the `vasp_gam` subfolder. As you see when you run `doped`'s `DefectsParser`, it automatically detects `vasp_gam` as the appropriate subfolder (with the most amount of VASP output files present across the defect directories – but of course this can also be manually controlled with the `subfolder` parameter), and so it only parsed the `v_Cd_X/vasp_gam` folders, and so does not detect `v_Cd_+1` as a defect folder.

If you make the same `vasp_gam` folder for `v_Ca_+1` and move the output file to this directory, it will be correctly parsed. Alternatively you can use no subfolder for each defect folder (or any other sub-folder name, as long as it's consistent with all defect folders), and `doped` will automatically detect this and parse as appropriate. Hope this answers your query! 😃
