### 2024-03-15 SMTG-Bham/doped#55 @kavanase
Note that for the bulk, it can be with or without a subfolder (as you have here)
