### 2024-03-16 8309792
Update oxi-state handling; guess for bulk supercell once when parsing, implement timeout functions compatible with/without multiprocessing, early catch if extremely complex and will hang anyway...
