+++ doped/core.py
@@ -8,8 +8,8 @@ import inspect
 import warnings
 from dataclasses import asdict, dataclass, field
 from functools import reduce
-from multiprocessing import TimeoutError
-from multiprocessing.pool import ThreadPool
+from itertools import combinations_with_replacement
+from multiprocessing import Process, Queue, current_process
 from typing import Any, Optional, Union
 
 import numpy as np
@@ -909,6 +909,17 @@ def _get_dft_chempots(chempots, el_refs, limit):
 def _guess_and_set_struct_oxi_states(structure, try_without_max_sites=False, queue=None):
     """
     Tries to guess (and set) the oxidation states of the input structure.
+
+    Args:
+        structure (Structure): The structure for which to guess the oxidation states.
+        try_without_max_sites (bool):
+            Whether to try to guess the oxidation states
+            without using the ``max_sites=-1`` argument (``True``)(which attempts
+            to use the reduced composition for guessing oxi states) or not (``False``).
+        queue (Queue):
+            A multiprocessing queue to put the guessed structure in, if provided.
+            Only really intended for use internally in ``doped``, during defect
+            generation.
     """
     if try_without_max_sites:
         with contextlib.suppress(Exception):
@@ -939,6 +950,96 @@ def _guess_and_set_struct_oxi_states(structure, try_without_max_sites=False, que
 
     if queue is not None:
         queue.put(structure)
+
+
+def _guess_and_set_oxi_states_with_timeout(structure, timeout_1=10, timeout_2=15, queue=None) -> bool:
+    """
+    Tries to guess (and set) the oxidation states of the input structure, with
+    a timeout catch for cases where the structure is complex and oxi state
+    guessing will take a very very long time.
+
+    Tries first without using the ``max_sites=-1`` argument with ``pymatgen``'s
+    oxidation state guessing functions (which attempts to use the reduced
+    composition for guessing oxi states, but can be a little less reliable for tricky
+    cases), and if that times out, tries without ``max_sites=-1``.
+
+    Args:
+        structure (Structure): The structure for which to guess the oxidation states.
+        queue (Queue):
+            A multiprocessing queue to put the guessed structure in, if provided.
+            Only really intended for use internally in ``doped``, during defect
+            generation.
+        timeout_1 (float):
+            Timeout in seconds for the first attempt to guess the oxidation states
+            (with ``max_sites=-1``). Default is 10 seconds.
+        timeout_2 (float):
+            Timeout in seconds for the second attempt to guess the oxidation states
+            (without ``max_sites=-1``). Default is 15 seconds.
+    """
+    if queue is None:
+        queued_struct = False
+        queue = Queue()
+    else:
+        queued_struct = True
+
+    guess_oxi_process_wout_max_sites = Process(
+        target=_guess_and_set_struct_oxi_states, args=(structure, True, queue)
+    )  # try without max sites first, if fails, try with max sites
+    guess_oxi_process_wout_max_sites.start()
+    guess_oxi_process_wout_max_sites.join(timeout=timeout_1)
+
+    if guess_oxi_process_wout_max_sites.is_alive():  # still running, revert to using max sites
+        guess_oxi_process_wout_max_sites.terminate()
+        guess_oxi_process_wout_max_sites.join()
+
+        guess_oxi_process = Process(
+            target=_guess_and_set_struct_oxi_states,
+            args=(structure, False, queue),
+        )
+        guess_oxi_process.start()
+        guess_oxi_process.join(timeout=timeout_2)  # wait for pymatgen to guess oxi states,
+        # otherwise revert to all Defect oxi states being set to 0
+
+        if guess_oxi_process.is_alive():
+            guess_oxi_process.terminate()
+            guess_oxi_process.join()
+
[…]
