### 2024-03-16 a5699a7
Update oxi-state handling; guess for bulk supercell once when parsing, implement timeout functions compatible with/without multiprocessing, early catch if extremely complex and will hang anyway...
Former-commit-id: 83097924f139bfad705c6b867fd88515a00b8c95
