+++ doped/analysis.py
@@ -11,7 +11,7 @@ import contextlib
 import inspect
 import os
 import warnings
-from multiprocessing import Pool, cpu_count
+from multiprocessing import Pool, Queue, cpu_count
 from typing import Optional, Union
 
 import numpy as np
@@ -27,7 +27,7 @@ from pymatgen.io.vasp.outputs import Vasprun
 from tqdm import tqdm
 
 from doped import _ignore_pmg_warnings
-from doped.core import DefectEntry
+from doped.core import DefectEntry, _guess_and_set_oxi_states_with_timeout, _rough_oxi_state_cost_from_comp
 from doped.generation import get_defect_name_from_defect, get_defect_name_from_entry, name_defect_entries
 from doped.thermodynamics import DefectThermodynamics
 from doped.utils.parsing import (
@@ -160,7 +160,7 @@ def check_and_set_defect_entry_name(
 
 
 def defect_from_structures(
-    bulk_supercell, defect_supercell, return_all_info=False, bulk_voronoi_node_dict=None
+    bulk_supercell, defect_supercell, return_all_info=False, bulk_voronoi_node_dict=None, oxi_state=None
 ):
     """
     Auto-determines the defect type and defect site from the supplied bulk and
@@ -187,6 +187,9 @@ def defect_from_structures(
         bulk_voronoi_node_dict (dict):
             Dictionary of bulk supercell Voronoi node information, for
             expedited site-matching. If None, will be re-calculated.
+        oxi_state (int, float, str):
+            Oxidation state of the defect site. If not provided, will be
+            automatically determined from the defect structure.
 
     Returns:
         defect (Defect):
@@ -309,6 +312,7 @@ def defect_from_structures(
         "@class": def_type.capitalize(),
         "structure": bulk_supercell,
         "site": defect_site_in_bulk,
+        "oxi_state": oxi_state,
     }  # note that we now define the Defect in the bulk supercell, rather than the primitive structure
     # as done during generation. Future work could try mapping the relaxed defect site back to the
     # primitive cell, however interstitials will be very tricky for this...
@@ -342,7 +346,8 @@ def defect_name_from_structures(bulk_structure, defect_structure):
     Returns:
         str: Defect name.
     """
-    defect = defect_from_structures(bulk_structure, defect_structure)
+    # set oxi_state to avoid wasting time trying to auto-determine when unnecessary here
+    defect = defect_from_structures(bulk_structure, defect_structure, oxi_state=0)
 
     # note that if the symm_op approach fails for any reason here, the defect-supercell expansion
     # approach will only be valid if the defect structure is a diagonal expansion of the primitive...
@@ -626,6 +631,21 @@ class DefectsParser:
         self.bulk_vr = get_vasprun(
             bulk_vr_path,
         )  # parsing projected eigenvalues makes Vasprun loading much slower...
+
+        # try parsing the bulk oxidation states first, for later assigning defect "oxi_state"s (i.e.
+        # fully ionised charge states):
+        if _rough_oxi_state_cost_from_comp(self.bulk_vr.final_structure.composition) > 1e6:
+            self._bulk_oxi_states: Union[dict, bool] = False  # will take very long to guess oxi_state
+        else:
+            queue: Queue = Queue()
+            self._bulk_oxi_states = _guess_and_set_oxi_states_with_timeout(
+                self.bulk_vr.final_structure, queue=queue
+            )
+            if self._bulk_oxi_states:
+                self.bulk_vr.final_structure = queue.get()  # oxi-state decorated structure
+                self._bulk_oxi_states = {
+                    el.symbol: el.oxi_state for el in self.bulk_vr.final_structure.composition.elements
+                }
 
[…]
