### 2024-03-20 904c13d
defualt multiplicity factor handling
