### 2024-03-20 e63701c
Make chempots parsing backwards compatible
