### 2024-03-20 01d1607
Add import error catches for `vise`/`pydefect` as they're not auto-installed via `conda` (not available), and add installation notes for this
Former-commit-id: a2337db3e9d7e16c540cac114d36fecf93b9ef0a
