### 2024-03-20 a2337db
Add import error catches for `vise`/`pydefect` as they're not auto-installed via `conda` (not available), and add installation notes for this
