### 2024-03-21 6a485ac
Add PROCAR parsing as default behaviour
Former-commit-id: f3d44957a6a87dba2f6b048933ab8c2674dd30f7
