+++ doped/analysis.py
@@ -23,7 +23,7 @@ from pymatgen.analysis.structure_matcher import ElementComparator, StructureMatc
 from pymatgen.core.sites import PeriodicSite
 from pymatgen.ext.matproj import MPRester
 from pymatgen.io.vasp.inputs import Poscar
-from pymatgen.io.vasp.outputs import Vasprun
+from pymatgen.io.vasp.outputs import Procar, Vasprun
 from tqdm import tqdm
 
 from doped import _ignore_pmg_warnings
@@ -50,7 +50,7 @@ from doped.utils.parsing import (
     get_outcar,
     get_vasprun,
 )
-from doped.utils.phs import get_band_edge_info
+from doped.utils.phs import get_band_edge_info, get_band_edge_info_procar
 from doped.utils.plotting import format_defect_name
 from doped.utils.symmetry import (
     _frac_coords_sort_func,
@@ -523,9 +523,12 @@ class DefectsParser:
                 {Chemical Formula} is the chemical formula of the host material.
                 If False, no json file is saved.
             load_phs_data (bool):
-                Automatically determines the band edge states of the defect to determine
-                if the defect is a PHS. Also returns single-particle levels and their
-                occupation. cite: https://doi.org/10.1103/PhysRevMaterials.5.123803
+                Load the projected eigenvalues and sets up a ``pydefect`` ``BandEdgeOrbitalInfos``
+                object with the required information for determining the band edge states
+                using ``DefectEntry.get_perturbed_host_state()``. Will initially try to load
+                from ``PROCAR`` files, otherwise will fall back on loading project eigenvalues
+                from ``vasprun.xml``.This can cause a significant increase in the parsing
+                time (30%-70% longer), so set to False if not needed.
                 Default = True.
 
         Attributes:
@@ -636,9 +639,16 @@ class DefectsParser:
                 bulk_vr_path,
                 dir_type="bulk",
             )
-        self.bulk_vr = get_vasprun(
-            bulk_vr_path, parse_projected_eigen=load_phs_data
-        )  # parsing projected eigenvalues makes Vasprun loading much slower...
+
+        # Checking if PROCAR available to avoid slow loading
+        bulk_procar_path, multiple = _get_output_files_and_check_if_multiple("PROCAR", self.bulk_path)
+        if "PROCAR" in bulk_procar_path:
+            self.bulk_vr = get_vasprun(bulk_vr_path, parse_projected_eigen=False)
+            self.bulk_procar = Procar(bulk_procar_path)
+        else:
+            self.bulk_vr = get_vasprun(bulk_vr_path, parse_projected_eigen=load_phs_data)
+            self.bulk_procar = None
+            # parsing projected eigenvalues makes Vasprun loading much slower...
 
         # try parsing the bulk oxidation states first, for later assigning defect "oxi_state"s (i.e.
         # fully ionised charge states):
@@ -1116,6 +1126,7 @@ class DefectsParser:
                 defect_path=os.path.join(self.output_path, defect_folder, self.subfolder),
                 bulk_path=self.bulk_path,
                 bulk_vr=self.bulk_vr,
+                bulk_procar=self.bulk_procar,
                 dielectric=self.dielectric,
                 skip_corrections=self.skip_corrections,
                 error_tolerance=self.error_tolerance,
@@ -1146,7 +1157,7 @@ class DefectsParser:
                     dp.defect_entry.calculation_metadata
                 )["bulk_site_potentials"]
 
-            if dp.kwargs.get("bulk_outcar", None) is not None:  # ADAIR CHECK
+            if dp.kwargs.get("bulk_outcar", None) is not None:
[…]
