### 2024-03-21 a4e5a42
Fix phs parsing in cases where Elements replaced with Species due to auto oxidation detection
