### 2024-03-22 415ce05
Remove some code redundancy
Former-commit-id: 0062f3b3f8373144ab51fbb35704f89e5bee2cf5
