### 2024-03-22 0b647f0
Update doc string, fix parsing cases, update tests
