+++ tests/test_analysis.py
@@ -71,6 +71,7 @@ class DefectsParsingTestCase(unittest.TestCase):
         self.Sb2Se3_dielectric = np.array([[85.64, 0, 0], [0.0, 128.18, 0], [0, 0, 15.00]])
 
         self.Cu2SiSe3_EXAMPLE_DIR = os.path.join(self.module_path, "../examples/Cu2SiSe3")
+        self.MgO_EXAMPLE_DIR = os.path.join(self.module_path, "../examples/MgO")
 
         self.Sb2Si2Te6_dielectric = [44.12, 44.12, 17.82]
         self.Sb2Si2Te6_DATA_DIR = os.path.join(self.EXAMPLE_DIR, "Sb2Si2Te6")
@@ -597,6 +598,7 @@ class DefectsParsingTestCase(unittest.TestCase):
                 output_path=self.YTOS_EXAMPLE_DIR,
                 dielectric=self.ytos_dielectric,
                 json_filename="YTOS_example_defect_dict.json",
+                load_phs_data=False,
             )  # for testing in test_thermodynamics.py
         print([warn.message for warn in w])  # for debugging
         assert not w  # hidden files ignored
@@ -622,6 +624,7 @@ class DefectsParsingTestCase(unittest.TestCase):
                 output_path=self.YTOS_EXAMPLE_DIR,
                 dielectric=self.ytos_dielectric,
                 json_filename="YTOS_example_defect_dict.json",
+                load_phs_data=False,
             )  # for testing in test_thermodynamics.py
         print([warn.message for warn in w])  # for debugging
         assert not w  # hidden files ignored
@@ -717,31 +720,33 @@ class DefectsParsingTestCase(unittest.TestCase):
         """
         Test PHS functions.
         """
-        # Test the loading of Cu2SiSe3 vacancy
+        # Test loading of MgO using vasprun.xml
         defect_entry = DefectParser.from_paths(
-            f"{self.Cu2SiSe3_EXAMPLE_DIR}/v_Cu_0/vasp_std",
-            f"{self.Cu2SiSe3_EXAMPLE_DIR}/bulk/vasp_std",
+            f"{self.MgO_EXAMPLE_DIR}/Defects/Mg_O_+1/vasp_std",
+            f"{self.MgO_EXAMPLE_DIR}/Defects/MgO_bulk/vasp_std",
             skip_corrections=True,
             load_phs_data=True,
         ).defect_entry
 
         bes, fig = defect_entry.get_perturbed_host_state()  # Test plotting KS
+        with open(f"{self.MgO_EXAMPLE_DIR}/Defects/Mg_O_1_band_edge_states.json") as f:
+            expected = json.load(f)
+        assert bes.as_dict() == expected
+
+        # Test loading using ``PROCAR`` and ``DefectsParser``
+        dp = DefectsParser(f"{self.Cu2SiSe3_EXAMPLE_DIR}", skip_corrections=True)
+
+        bes, fig = dp.defect_dict["v_Cu_0"].get_perturbed_host_state()  # Test plotting KS
         with open(f"{self.Cu2SiSe3_EXAMPLE_DIR}/Cu2SiSe3_vac_band_edge_states.json") as f:
             expected = json.load(f)
         assert bes.as_dict() == expected
 
-        bes2 = defect_entry.get_perturbed_host_state(plot=False)  # Test getting BES and not plot
+        bes2 = dp.defect_dict["v_Cu_0"].get_perturbed_host_state(
+            plot=False,
+        )  # Test getting BES and not plot
         assert bes2.as_dict() == bes.as_dict()
 
-        # Test the loading of Cu2SiSe3 interstitial
[…]
