### 2024-03-22 fc70d05
Refactor PHS parsing, update tests to avoid floating-point error
