+++ doped/utils/phs.py
@@ -9,6 +9,7 @@ import logging
 import os
 from collections import defaultdict
 from importlib.metadata import version
+from itertools import zip_longest
 from typing import Any, Optional
 
 import matplotlib.pyplot as plt
@@ -19,6 +20,7 @@ from vise import user_settings
 
 user_settings.logger.setLevel(logging.CRITICAL)
 
+import pydefect.analyzer.make_band_edge_states
 import pydefect.cli.vasp.make_band_edge_orbital_infos as make_bes
 from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
 from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
@@ -32,7 +34,7 @@ from pydefect.defaults import defaults
 from pydefect.util.structure_tools import Coordination, Distances
 from pymatgen.core import Element, Species
 from pymatgen.electronic_structure.core import Spin
-from pymatgen.io.vasp.outputs import Outcar, Vasprun
+from pymatgen.io.vasp.outputs import Outcar, Procar, Vasprun
 from shakenbreak.plotting import _install_custom_font
 from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
 
@@ -80,7 +82,7 @@ def _distances(self, remove_self=True, specie=None) -> list[float]:
     return result
 
 
-def make_band_edge_orbital_infos_vr(
+def _make_band_edge_orbital_infos_vr(
     defect_vr: Vasprun, vbm: float, cbm: float, str_info, eigval_shift: float = 0.0
 ):
     """
@@ -139,29 +141,49 @@ def make_band_edge_orbital_infos_vr(
     )
 
 
-def get_band_edge_info(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar):
+# elif defect_procar and bulk_procar:
+# band_orb, vbm_info, cbm_info = get_band_edge_info_procar(
+#     bulk_procar, bulk_vr, bulk_outcar_phs, defect_vr, defect_procar
+# )
+# else:
+# band_orb, vbm_info, cbm_info = get_band_edge_info(defect_vr, bulk_vr, bulk_outcar_phs)
+
+
+def get_band_edge_info(
+    bulk_vr: Vasprun,
+    bulk_outcar: Outcar,
+    defect_vr: Vasprun,
+    bulk_procar: Optional[Procar] = None,
+    defect_procar: Optional[Procar] = None,
+):
     """
     Load metadata required for performing PHS identification.
 
     See https://doped.readthedocs.io/en/latest/Tips.html#perturbed-host-states
 
     Args:
-        defect_vr (Vasprun):
-            Vasprun object of the defect supercell calculation
         bulk_vr (Vasprun):
             Vasprun object of the bulk supercell calculation
         bulk_outcar (Outcar):
-            Outcar object of the bulk supercell calculation
+            Outcar object of
+        defect_vr (Vasprun):
+            Vasprun object of the defect supercell calculation
+        bulk_procar (Procar):
+            Procar object of the bulk supercell calculation
+        defect_procar (Procar):
+            Procar object of the defect supercell calculation
     Returns:
-        pydefect EdgeInfo class
+        ``pydefect`` ``EdgeInfo`` class
     """
     # Check in the correct version of vise installed if a non-collinear calculation is parsed.
     # TODO: Remove this check when ``vise 0.8.2`` is released on PyPi.
     try:
-        band_edge_prop = VaspBandEdgeProperties(bulk_vr, bulk_outcar)
+        if bulk_procar is None:
+            band_edge_prop = VaspBandEdgeProperties(bulk_vr, bulk_outcar)
+        else:
[…]
