### 2024-03-22 f63cc54
Refactors ``get_band_edge_info``
