+++ doped/utils/phs.py
@@ -141,14 +141,6 @@ def _make_band_edge_orbital_infos_vr(
     )
 
 
-# elif defect_procar and bulk_procar:
-# band_orb, vbm_info, cbm_info = get_band_edge_info_procar(
-#     bulk_procar, bulk_vr, bulk_outcar_phs, defect_vr, defect_procar
-# )
-# else:
-# band_orb, vbm_info, cbm_info = get_band_edge_info(defect_vr, bulk_vr, bulk_outcar_phs)
-
-
 def get_band_edge_info(
     bulk_vr: Vasprun,
     bulk_outcar: Outcar,
@@ -239,135 +231,6 @@ def get_band_edge_info(
         )
 
     return band_orb, vbm_info, cbm_info
-
-
-# def get_band_edge_info_vr(defect_vr: Vasprun, bulk_vr: Vasprun, bulk_outcar: Outcar):
-#     """
-#     Load metadata required for performing PHS identification.
-#
-#     See https://doped.readthedocs.io/en/latest/Tips.html#perturbed-host-states
-#
-#     Args:
-#         defect_vr (Vasprun):
-#             Vasprun object of the defect supercell calculation
-#         bulk_vr (Vasprun):
-#             Vasprun object of the bulk supercell calculation
-#         bulk_outcar (Outcar):
-#             Outcar object of the bulk supercell calculation
-#     Returns:
-#         pydefect EdgeInfo class
-#     """
-#     # Check in the correct version of vise installed if a non-collinear calculation is parsed.
-#     # TODO: Remove this check when ``vise 0.8.2`` is released on PyPi.
-#     try:
-#         band_edge_prop = VaspBandEdgeProperties(bulk_vr, bulk_outcar)
-#         if defect_vr.parameters.get("LNONCOLLINEAR") is True:
-#             assert band_edge_prop._ho_band_index(Spin.up) == int(bulk_vr.parameters.get("NELECT")) - 1
-#
-#     except AssertionError as exc:
-#         v_vise = version("vise")
-#         if v_vise <= "0.8.1":
-#             raise RuntimeError(
-#                 f"You have version {v_vise} of the package `vise`, which does not allow the parsing of "
-#                 f"non-collinear (SOC) calculations. You can install the updated version of `vise` from "
-#                 f"the GitHub repo for this functionality."
-#             ) from exc
-#
-#         raise exc
-#
-#     orbs, s = bulk_vr.projected_eigenvalues, bulk_vr.final_structure
-#     vbm_info = get_edge_info(band_edge_prop.vbm_info, orbs, s, bulk_vr)
-#     cbm_info = get_edge_info(band_edge_prop.cbm_info, orbs, s, bulk_vr)
-#
-#     # Money patch to set include_on_site = True
-#     _orig_method_coor = Distances.coordination
-#     Distances.coordination = _coordination
-#     _orig_method_dist = Distances.distances
-#     Distances.distances = _distances
-#
-#     dsinfo = MakeDefectStructureInfo(  # Using default values suggested by pydefect
-#         bulk_vr.final_structure,
-#         defect_vr.final_structure,
-#         defect_vr.final_structure,
-#         symprec=0.1,
-#         dist_tol=1.0,
-#         neighbor_cutoff_factor=1.3,  # Neighbors are sites within min_dist * neighbor_cutoff_factor
-#     )
-#
-#     # Undo monkey patch in case used in other parts of the code:
-#     Distances.coordination = _orig_method_coor
-#     Distances.distances = _orig_method_dist
-#
-#     band_orb = _make_band_edge_orbital_infos_vr(
-#         defect_vr,
-#         vbm_info.orbital_info.energy,
-#         cbm_info.orbital_info.energy,
-#         eigval_shift=-vbm_info.orbital_info.energy,
[…]
