### 2024-03-22 4bef3a7
pymatgen.io.vasp.outputs.Procar can't parse non-collinear, so have to use vasprun.xml for those calculations to get PHS data. Rough fix, need to polish
Former-commit-id: e09e9741d387d2e5ac0081d7c3dda6c52c6ff4d8
