### 2024-03-22 72fee8f
Add `DopedDictSet` class as base `DictSet` class, with expedited I/O (`POTCAR` hashing, robust `KPOINTS`/`POTCAR`/`POSCAR` handling (formatting & sorting), etc)
Former-commit-id: c62bee68e62b4a22428e5e03f6d61a89aabc310b
