+++ doped/vasp.py
@@ -4,7 +4,6 @@ Code to generate VASP defect calculation input files.
 
 import contextlib
 import copy
-import inspect
 import json
 import os
 import warnings
@@ -24,7 +23,7 @@ from pymatgen.io.vasp.inputs import BadIncarWarning, Kpoints, Poscar, Potcar
 from pymatgen.io.vasp.sets import DictSet, UserPotcarFunctional
 from tqdm import tqdm
 
-from doped import _ignore_pmg_warnings
+from doped import _doped_obj_properties_methods, _ignore_pmg_warnings
 from doped.core import DefectEntry
 from doped.generation import (
     DefectsGenerator,
@@ -53,7 +52,7 @@ def deep_dict_update(d: dict, u: dict) -> dict:
 
 
 MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
-default_potcar_dict = loadfn(os.path.join(MODULE_DIR, "VASP_sets/PotcarSet.yaml"))
+default_potcar_dict = loadfn(os.path.join(MODULE_DIR, "VASP_sets/PotcarSet.yaml"))["POTCAR"]
 default_relax_set = loadfn(os.path.join(MODULE_DIR, "VASP_sets/RelaxSet.yaml"))
 default_HSE_set = loadfn(os.path.join(MODULE_DIR, "VASP_sets/HSESet.yaml"))
 default_defect_set = loadfn(os.path.join(MODULE_DIR, "VASP_sets/DefectSet.yaml"))
@@ -61,9 +60,6 @@ default_defect_relax_set = copy.deepcopy(default_relax_set)
 default_defect_relax_set = deep_dict_update(
     default_defect_relax_set, default_defect_set
 )  # defect set is just INCAR settings
-default_defect_relax_set = deep_dict_update(
-    default_defect_relax_set, default_potcar_dict
-)  # only POTCAR settings, not set in other *Set.yamls
 singleshot_incar_settings = {
     "EDIFF": 1e-6,  # tight EDIFF for final energy and converged DOS
     "EDIFFG": None,  # no ionic relaxation, remove to avoid confusion
@@ -129,9 +125,194 @@ class DopedKpoints(Kpoints):
             return super().__repr__()
 
 
-class DefectDictSet(DictSet):
+class DopedDictSet(DictSet):
     """
-    Extension to pymatgen DictSet object for VASP defect calculations.
+    Modified version of ``pymatgen`` ``DictSet``, to have more robust
+    ``POTCAR`` handling, expedited I/O (particularly for ``POTCAR`` generation,
+    which can be slow when generating many folders), ensure ``POSCAR`` atom
+    sorting, avoid encoding issues with ``KPOINTS`` comments etc.
+    """
+
+    def __init__(
+        self,
+        structure: Structure,
+        user_incar_settings: Optional[dict] = None,
+        user_kpoints_settings: Optional[Union[dict, Kpoints]] = None,
+        user_potcar_functional: UserPotcarFunctional = "PBE",
+        user_potcar_settings: Optional[dict] = None,
+        auto_kpar: bool = True,
+        poscar_comment: Optional[str] = None,
+        **kwargs,
+    ):
+        r"""
+        Args:
+            structure (Structure): ``pymatgen`` ``Structure`` object
+            user_incar_settings (dict):
+                Dictionary of user INCAR settings (AEXX, NCORE etc.) to override
+                default ``INCAR`` settings. Note that any flags that aren't
+                numbers or ``True/False`` need to be input as strings with
+                quotation marks (e.g. ``{"ALGO": "All"}``).
+                (default: None)
+            user_kpoints_settings (dict or Kpoints):
+                Dictionary of user ``KPOINTS`` settings (in ``pymatgen`` ``DictSet``
[…]
