### 2024-03-22 d5ccac2
Use `DopedDictSet` for IO in competing phases calculations, much faster and less redundant
Former-commit-id: b64a0f3cfa6c31b2a7bd6d798f2eb6c49f1fb328 [formerly fb378ca91339c4fc061f45787f5511efa14edfde]
Former-commit-id: e7ffbc7e649730fa06b5d999328f44721249066f
