### 2024-03-22 fb378ca
Use `DopedDictSet` for IO in competing phases calculations, much faster and less redundant
