+++ doped/chemical_potentials.py
@@ -19,20 +19,12 @@ from pymatgen.entries.computed_entries import ComputedStructureEntry
 from pymatgen.ext.matproj import MPRester
 from pymatgen.io.vasp.inputs import Kpoints
 from pymatgen.io.vasp.outputs import Vasprun
-from pymatgen.io.vasp.sets import DictSet
 
 from doped import _ignore_pmg_warnings
 from doped.utils.parsing import _get_output_files_and_check_if_multiple
-from doped.vasp import (
-    MODULE_DIR,
-    _test_potcar_functional_choice,
-    deep_dict_update,
-    default_HSE_set,
-    default_potcar_dict,
-    default_relax_set,
-)
+from doped.vasp import MODULE_DIR, DopedDictSet, default_HSE_set, default_relax_set
 
-pbesol_convrg_set = loadfn(os.path.join(MODULE_DIR, "VASP_sets/PBEsol_ConvergenceSet.yaml"))
+pbesol_convrg_set = loadfn(os.path.join(MODULE_DIR, "VASP_sets/PBEsol_ConvergenceSet.yaml"))  # just INCAR
 
 # globally ignore:
 _ignore_pmg_warnings()
@@ -287,10 +279,11 @@ class CompetingPhases:
     # analyze_GGA_chempots code for example.
     # TODO: Add note to notebook that if your bulk phase is lower energy than its version on the MP
     # (e.g. distorted perovskite), then you should use this for your bulk competing phase calculation.
-    # TODO: DictSet initialisation and writing here can be a little slow. Could be made more
-    #  efficient by just editing the kpoints in each loop (not reinitialising) for the convergence test
-    #  generation, and using the POTCAR hashing function from `doped.vasp`. Also should dynamically set
-    #  KPAR based on the (default) number of KPOINTS for each material
+
+    # TODO: Is the Materials Project entry structure always the primitive structure? Should check,
+    #  and if not then use something like this:
+    #  sga = SpacegroupAnalyzer(e.structure)
+    #  struct = sym.get_primitive_standard_structure() -> output this structure
 
     def __init__(self, composition, e_above_hull=0.1, api_key=None, full_phase_diagram=False):
         """
@@ -479,7 +472,6 @@ class CompetingPhases:
             )
         )
 
-    # TODO: Similar refactor to work mainly off config dict object here as well (as vasp_input)?
     # TODO: Return dict of DictSet objects for this and vasp_std_setup() functions, as well as
     #  write_files option, for ready integration with high-throughput workflows
     def convergence_setup(
@@ -523,17 +515,8 @@ class CompetingPhases:
         min_nm, max_nm, step_nm = kpoints_nonmetals
         min_m, max_m, step_m = kpoints_metals
 
-        potcar_dict = copy.deepcopy(default_potcar_dict)
-        if user_potcar_settings:
-            potcar_dict["POTCAR"].update(user_potcar_settings)
-        if user_potcar_functional:
-            potcar_dict["POTCAR_FUNCTIONAL"] = user_potcar_functional
-        # test potcar choice:
-        if not kwargs.get("potcar_spec", False):
-            potcar_dict["POTCAR_FUNCTIONAL"] = _test_potcar_functional_choice(
-                potcar_dict["POTCAR_FUNCTIONAL"]
-            )
-        pbesol_convrg_set.update(potcar_dict)
+        base_user_incar_settings = copy.deepcopy(pbesol_convrg_set["INCAR"])
+        base_user_incar_settings.update(user_incar_settings or {})  # user_incar_settings override defaults
 
         # separate metals and non-metals
         self.nonmetals = []
@@ -546,23 +529,22 @@ class CompetingPhases:
                 self.nonmetals.append(e)
             else:
[…]
