### 2024-03-22 e7ffbc7
Use `DopedDictSet` for IO in competing phases calculations, much faster and less redundant
Former-commit-id: fb378ca91339c4fc061f45787f5511efa14edfde
