### 2024-03-22 c1633af
Add virtual environment installation instructions (JOSS review R1 Q2); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: f113ab02ca8d4cf1da9124e667265cdef86fcdcf [formerly eda6e3d29ac328bc82f23c3f55f7bbfc9a53a7e9]
Former-commit-id: c6f4da1ce68636b5543bee25087698317663a3b3
