### 2024-03-22 eda6e3d
Add virtual environment installation instructions (JOSS review R1 Q2); https://github.com/openjournals/joss-reviews/issues/6433
