+++ docs/Installation.rst
@@ -18,6 +18,16 @@ If you are installing ``doped`` via ``conda`` and you plan on using the eFNV (Ku
 correction, you will need to install the ``pydefect`` package with ``pip`` as shown or otherwise, as it is
 not available on ``conda``.
 
+It may be desirable to install ``doped`` in a virtual envionment (e.g. if you encounter package dependency
+conflict warnings during installation etc). You can do this with ``conda`` with:
+
+.. code-block:: bash
+
+   conda create -n doped python=3.11
+   conda activate doped
+   pip install doped
+
+
 Setup ``POTCAR``\s and Materials Project API
 --------------------------------------------
 
