### 2024-03-22 #6433 @kavanase
Here I'm responding point-by-point to track these issues. I'll update as I go, and then bump this thread when complete.

- [ ] "[Using doped to parse calculations not prepared by doped](https://github.com/SMTG-Bham/doped/issues/55)"
Was just an issue with the user format, now resolved.

- [ ] "For installation, I suggest mentioning pip install doped under conda virtual environment"
Done! https://github.com/SMTG-Bham/doped/commit/eda6e3d29ac328bc82f23c3f55f7bbfc9a53a7e9; see https://doped.readthedocs.io/en/latest/Installation.html

- [ ] The cells about snb_path and figs["v_Cd_0" break in [generation_tutorial.ipynb](https://github.com/SMTG-Bham/doped/blob/main/examples/generation_tutorial.ipynb)

- [ ] I suggest adding a minimum example of specifying supercell size and shape, for more user control. For example, starting from a silicon conventional cell (in cubic shape), one may want to simply expand it to
 while keeping the cubic shape. Achieving a minimum image distance is nice, yet sometimes one may want to use a certain supercell shape to do supercell extrapolation. generation_tutorial.ipynb

- [ ] For the functions introduced in the tutorials, I suggest (in future update) specifying the arguments, even if they are the default values. For example, write_vasp_files(output_path=’./’), which could help users learn to use doped.

- [ ] There seems to be some renames (limits → facets) that are not updated in the tutorials parsing_tutorial.ipynb and thermodynamics_tutorial.ipynb

- [ ] For postprocessing and analysis, I suggest the authors add a short section in the tutorials summarizing what json/csv files will be generated, for which step.

- [ ] Please update renames csv_fname → csv_path in chemical_potentials_tutorial.ipynb

- [ ] Error in in advanced_analysis_tutorial.ipynb: bulk_path = "CdTe/CdTe_bulk/vasp_gam/, but FileNotFoundError: [Errno 2] No such file or directory: 'CdTe/CdTe_bulk/vasp_gam/

- [ ] Error in plotting_customisation_tutorial.ipynb te_cd_entry = [entry for entry in CdTe_thermo.defect_entries if entry.name == "Te_Cd_1"][0], but IndexError: list index out of range

- [ ] For ShakeNbreak, it might be better to replace the folder name Rattled with another word that describes more intuitively the POSCAR under this folder.

- [ ] For setting the spin multiplicity in INCAR, should the authors remind the users there are two options: NUPDOWN and MAGMOM?

- [ ] For competing phases calculations, how is the k-point density look like in doped? How does it compare with the Materials Project Relaxset k-point density?

- [ ] For defect structure generation, the multiplicity from doped is different from the one from pycdt, despite I give both codes the same conventional cell (not primitive cell). For example, I give both codes a CaO conventional cell (NaCl structure, cubic, 8 atoms). doped gives multiplicity=1 for V_Ca, while pycdt gives multiplicity=4. Could the authors help clarify?

- [ ] The paper is clear and very readable. There could be three minor revisions. First, in Figure 2(b), there are "prob." and "probabilities", which could be unified. Second, in Figure 2(b), what is the "ICDS prob."? Third, I suggest the authors cite two recent work related to high-throughput defect calculations: npj Computational Materials (2023) 9:72 ; https://doi.org/10.1038/s41524-023-01015-6 and Yuan et al., Discovery of the Zintl-phosphide BaCd2P2 as a long carrier lifetime and stable solar absorber, Joule (2024), https://doi.org/10.1016/j.joule.2024.02.01715.

- [ ] The authors mentioned high-throughput compatibility. As far as I checked, there is no tutorial for this. From my high-throughput defect computation experience using the PyCDT (with atomate1) and the pymatgen-defect-analysis (with atomate2), I believe it will be involved to compile the doped with high-throughput computation, as this involves management of job workflow, data, and database. This functionality could be strengthened in future development.

- [ ] In future development, the authors could consider adding a function plotting the defect states (i.e., single-particle defect levels) in the band gap. Such plots are widely seen nowadays.

- [ ] In future development, generating defect complexes could also be a feature. Moreover, designing defects/defect complexes with specific symmetry could be interesting.

- [ ] In future development, could the functions like get_symmetries_and_degeneracies be more independent? These functions are very useful for analyzing old calculations not prepared by doped.
