### 2024-03-22 c7c0fa3
Add note about needing to clone SnB repo for example data (JOSS review R1 Q3); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: faf6e873d07419d2d71256f1ee9858af260ef43e
