### 2024-03-22 9507178
Add note about needing to clone SnB repo for example data (JOSS review R1 Q3); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 3ec2ad45769c6292c76432f2e19ada376cf067dd [formerly faf6e873d07419d2d71256f1ee9858af260ef43e]
Former-commit-id: c7c0fa3fec74080388436ef26a548ed7ee1f9ef7
