### 2024-03-22 faf6e87
Add note about needing to clone SnB repo for example data (JOSS review R1 Q3); https://github.com/openjournals/joss-reviews/issues/6433
