### 2024-03-22 c3ded5c
Add custom supercell generation section to advanced analysis docs, and link in generation docs (JOSS review R1 Q4); https://github.com/openjournals/joss-reviews/issues/6433
