### 2024-03-22 10dd47e
Add custom supercell generation section to advanced analysis docs, and link in generation docs (JOSS review R1 Q4); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 7c0aa0bdf0977230a0a9d9b472243f70ab4b3105 [formerly c3ded5ce56b0a9f46bba69eed31f938b77cb4c03]
Former-commit-id: 96351d4558b3fc82380506b4a1acab24b55b0e72
