### 2024-03-22 96351d4
Add custom supercell generation section to advanced analysis docs, and link in generation docs (JOSS review R1 Q4); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: c3ded5ce56b0a9f46bba69eed31f938b77cb4c03
