### 2024-03-22 4c879c2
Add more explicit parameter names to tutorials (JOSS review R1 Q5); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: b90083c7f16b90ae273f5be500655eaabd63695e [formerly c29cd1b9068d733f8259dd9bda975bc6b0e38d7f]
Former-commit-id: 9f6cc7580d68fd7ed27771f92df17e484b83ab60
