### 2024-03-22 c29cd1b
Add more explicit parameter names to tutorials (JOSS review R1 Q5); https://github.com/openjournals/joss-reviews/issues/6433
