### 2024-03-22 a973ddb
Fix typo (JOSS review R1 Q10); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: a10f381284645cb485e28a08eb7ef9be018f325a [formerly 325eac28878253034a0756697bd18ccfa3c4fbad]
Former-commit-id: 45eba3e629e7123bcc96637db512c9371010bd09
