### 2024-03-22 45eba3e
Fix typo (JOSS review R1 Q10); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 325eac28878253034a0756697bd18ccfa3c4fbad
