### 2024-03-22 325eac2
Fix typo (JOSS review R1 Q10); https://github.com/openjournals/joss-reviews/issues/6433
