### 2024-03-23 f41ad1d
More chemical potentials code cleanup
