### 2024-03-23 a53a8b0
Add `Serialization & Data Provenance` section to docs tips page (JOSS review R1 Q7); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 6caee5bfc9667a5e56cdca4898b057201a8114e2
