### 2024-03-23 bc63e95
Add `Serialization & Data Provenance` section to docs tips page (JOSS review R1 Q7); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 4cc507d852452a773334f69955aa9249c0c32b24 [formerly 6caee5bfc9667a5e56cdca4898b057201a8114e2]
Former-commit-id: a53a8b0c8757f0c69b9d1fff671c82ed356f0aba
