### 2024-03-23 6caee5b
Add `Serialization & Data Provenance` section to docs tips page (JOSS review R1 Q7); https://github.com/openjournals/joss-reviews/issues/6433
