### 2024-03-23 bf50910
Chempots code cleanup
