+++ doped/generation.py
@@ -1989,8 +1989,8 @@ class DefectsGenerator(MSONable):
 
     def to_json(self, filename: Optional[str] = None):
         """
-        Save the DefectsGenerator object as a json file, which can be reloaded
-        with the DefectsGenerator.from_json() class method.
+        Save the ``DefectsGenerator`` object as a json file, which can be
+        reloaded with the ``DefectsGenerator.from_json()`` class method.
 
         Args:
             filename (str): Filename to save json file as. If None, the filename will be
