### 2024-03-23 65058ed
Add tips section on 'Spin Polarisation' (JOSS review R1 Q12); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: d52091ceb2649ebdfa381417a1a50b41e2f75547
