+++ docs/Tips.rst
@@ -12,7 +12,7 @@ their specific needs/system.
     (based on the literature and our experience), there is no substitute for the user's own judgement.
     Defect behaviour can be incredibly system-dependent, and so it is `always` important to question and
     consider the choices and approximations made in the workflow (such as supercell choice, charge state
-    ranges, interstitial site pruning, `MAGMOM` initialisation etc.) in the context of your specific
+    ranges, interstitial site pruning, ``MAGMOM`` initialisation etc.) in the context of your specific
     host system.
 
 
@@ -91,7 +91,7 @@ underlying calculation and/or extreme forces.
       `this part <https://shakenbreak.readthedocs.io/en/latest/Tips.html#bulk-phase-transformations>`_
       of the ``SnB`` docs.
 
-    - **Alternatively (if you have already performed `SnB` structure-seaerching), convergence of the forces can be aided by:**
+    - **Alternatively (if you have already performed ``SnB`` structure-seaerching), convergence of the forces can be aided by:**
     - Switching the ionic relaxation algorithm back and forth (i.e. change :code:`IBRION` to :code:`1` or
       :code:`3` and back).
     - Reducing the ionic step width (e.g. change :code:`POTIM` to :code:`0.02` in the :code:`INCAR`)
@@ -131,8 +131,8 @@ when parsing the intrinsic defects, the -3 charge antimony vacancy (``v_Sb-3``) 
 .. code-block::
 
         Estimated error in the Kumagai (eFNV) charge correction for defect v_Sb_-3 is 0.067 eV (i.e. which is
-        greater than the `error_tolerance`: 0.050 eV). You may want to check the accuracy of the correction by
-        plotting the site potential differences (using `defect_entry.get_kumagai_correction()` with `plot=True`).
+        greater than the ``error_tolerance``: 0.050 eV). You may want to check the accuracy of the correction by
+        plotting the site potential differences (using ``defect_entry.get_kumagai_correction()`` with ``plot=True``).
         Large errors are often due to unstable or shallow defect charge states (which can't be accurately modelled
         with the supercell approach). If this error is not acceptable, you may need to use a larger supercell
         for more accurate energies.
@@ -195,6 +195,99 @@ Below are the two resulting charge correction plots (using ``defect_region_radiu
     :width: 320px
     :align: right
 
+
+Spin Polarisation
+-----------------
+Proper accounting of spin polarisation and multiplicity is crucial for accurate defect calculations and
+analysis. For defect species with odd numbers of electrons (and thus being open-shell), they will adopt
+non-zero integer spin states, while defect species with even numbers of electrons can be either
+closed-shell (spin-paired) or open-shell (spin-active), depending on the defect species and its electronic
+structure. As such, defect calculations should typically be performed with spin polarisation allowed in all
+cases (i.e. with ``ISPIN = 2`` in VASP).
+
+.. tip::
+
+    If we have (nearly) converged the geometry relaxation for an even-electron defect species and there is
+    no non-zero magnetic moments on any site (given by the ``magnetization`` output in the ``OUTCAR`` file)
+    – and so adopting a closed-shell electronic structure, then we can set ``ISPIN = 1`` (turning off
+    spin polarisation) for subsequent calculations to reduce the computational cost.
+
+    The ``snb-mag --verbose`` CLI command from ``ShakeNBreak`` can be used to automatically check the
+    magnetisation of a VASP defect calculation in this way (and is automatically used by ``snb-run`` to
+    set ``ISPIN = 1`` for continued ``ShakeNBreak`` relaxations of any closed-shell defect calculations,
+    if it is being used to manage the structure-searching calculations).
+
+    .. code-block::
+
[…]
