### 2024-03-23 d52091c
Add tips section on 'Spin Polarisation' (JOSS review R1 Q12); https://github.com/openjournals/joss-reviews/issues/6433
