### 2024-03-23 304dd2c
Add tips section on 'Spin Polarisation' (JOSS review R1 Q12); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 743c54749d8938b605f6a0a3289752e1e214d4b8 [formerly d52091ceb2649ebdfa381417a1a50b41e2f75547]
Former-commit-id: 65058ed518d0038faed2e929b4c860e79c30ae1b
