### 2024-03-25 fad9ae5
Make point symmetry determination more robust and make `symprec` a customisable parameter
Former-commit-id: 05223035b452b99535b10940e4c1b9ceb0f09bd9
