### 2024-03-25 0522303
Make point symmetry determination more robust and make `symprec` a customisable parameter
