+++ doped/utils/symmetry.py
@@ -2,6 +2,7 @@
 Utility code and functions for symmetry analysis of structures and defects.
 """
 
+import contextlib
 import os
 import warnings
 from typing import Optional
@@ -573,8 +574,8 @@ def get_BCS_conventional_structure(structure, pbar=None, return_wyckoff_dict=Fal
     """
     Get the conventional crystal structure of the input structure, according to
     the Bilbao Crystallographic Server (BCS) definition. Also returns the
-    transformation matrix from the spglib (SpaceGroupAnalyzer) conventional
-    structure definition to the BCS definition.
+    transformation matrix from the ``spglib`` (``SpaceGroupAnalyzer``)
+    conventional structure definition to the BCS definition.
 
     Args:
         structure (Structure): pymatgen Structure object for this to
@@ -585,7 +586,7 @@ def get_BCS_conventional_structure(structure, pbar=None, return_wyckoff_dict=Fal
     number.
 
     Returns:
-        pymatgen Structure object and spglib -> BCS conv cell transformation matrix.
+        pymatgen Structure object and ``spglib`` -> BCS conv cell transformation matrix.
     """
     struc_wout_oxi = structure.copy()
     struc_wout_oxi.remove_oxidation_states()
@@ -733,7 +734,7 @@ def get_wyckoff_dict_from_sgn(sgn):
     (https://gitlab.com/ase/ase/-/merge_requests/1035) based on the tabulated
     datasets in https://github.com/xtalopt/randSpg (also found at
     https://github.com/spglib/spglib/blob/develop/database/Wyckoff.csv).
-    By default, doped uses the Wyckoff functionality of spglib (along with
+    By default, doped uses the Wyckoff functionality of ``spglib`` (along with
     symmetry operations in pymatgen) when possible however.
     """
     datafile = _get_wyckoff_datafile()
@@ -1036,15 +1037,15 @@ def point_symmetry_from_defect(defect, symm_ops=None, symprec=0.01):
     Get the defect site point symmetry from a Defect object.
 
     Note that this is intended only to be used for unrelaxed,
-    as-generated Defect objects (rather than parsed defects).
+    as-generated ``Defect`` objects (rather than parsed defects).
 
     Args:
-        defect (Defect): Defect object.
+        defect (Defect): ``Defect`` object.
         symm_ops (list):
-            List of symmetry operations of defect.structure, to avoid
+            List of symmetry operations of ``defect.structure``, to avoid
             re-calculating. Default is None (recalculates).
         symprec (float):
-            Symmetry tolerance for spglib. Default is 0.01.
+            Symmetry tolerance for ``spglib``. Default is 0.01.
 
     Returns:
         str: Defect point symmetry.
@@ -1124,10 +1125,10 @@ def point_symmetry_from_defect_entry(
             structure (if relaxed=False) or defect_entry.defect_supercell (if
             relaxed=True), to avoid re-calculating. Default is None (recalculates).
         symprec (float):
-            Symmetry tolerance for spglib. Default is 0.01 for unrelaxed structures,
-            0.2 for relaxed (to account for residual structural noise). You may
-            want to adjust for your system (e.g. if there are very slight
-            octahedral distortions etc).
+            Symmetry tolerance for ``spglib``. Default is 0.01 for unrelaxed structures,
+            0.1 for relaxed (to account for residual structural noise, matching that
+            used by the ``Materials Project``). You may want to adjust for your system
+            (e.g. if there are very slight octahedral distortions etc.).
         relaxed (bool):
             If False, determines the site symmetry using the defect site `in the
             unrelaxed bulk supercell` (i.e. the bulk site symmetry), otherwise
[…]
