### 2024-03-25 1fc0150
Add more explicit point symmetry tests
Former-commit-id: a5cfea0ec4a0c3e68b12d623987a4198d6015966
