### 2024-03-25 a5cfea0
Add more explicit point symmetry tests
