### 2024-03-25 08cc21c
Test setting `symprec` in various functions
