### 2024-03-25 0beed74
Test setting `symprec` in various functions
Former-commit-id: 08cc21c0155ccaaaefd93ab3bf851b3e169166c7
