### 2024-03-25 8d57f65
Add a note on the default `symprec` to docs
