### 2024-03-25 f9aa0cc
Paper updates (JOSS review R1 Q15); https://github.com/openjournals/joss-reviews/issues/6433
