+++ docs/JOSS/paper.md
@@ -77,7 +77,7 @@ Here we report ``doped``, a Python package for the generation, pre-/post-process
 The materials science sub-field of computational defect modelling has seen considerable growth in recent years, driven by the crucial importance of these species in functional materials and the major advances in computational methodologies and resources facilitating their accurate simulation.
 Software which enables researchers to efficiently and accurately perform these calculations, while allowing for in-depth target analyses of the resultant data, is thus of significant value to the community.
 Indeed there are many critical stages in the computational workflow for defects, which when performed manually not only consume significant researcher time and effort but also leave room for human error – particularly for newcomers to the field.
-Moreover, there are growing efforts to perform high-throughput investigations of defects in solids [@xiong_high-throughput_2023], necessitating robust, user-friendly and efficient software implementing this calculation workflow.
+Moreover, there are growing efforts to perform high-throughput investigations of defects in solids [@xiong_high-throughput_2023,@yuan_discovery_2024,@broberg_high-throughput_2023], necessitating robust, user-friendly and efficient software implementing this calculation workflow.
 
 [//]: # (By expediting the defect simulation workflow and providing efficient analysis tools, ``doped`` aims to... facilitate the investigation of defects in solids, and to enable the efficient and reproducible calculation of solid-state defect properties.)
 
@@ -100,28 +100,28 @@ While each of these codes have their strengths, they do not include the full sui
 The design philosophy of `doped` has been to implement the defect simulation workflow in an efficient, reproducible, user-friendly yet powerful and fully-customisable manner, combining reasonable defaults with full user control for each parameter in the workflow.
 As depicted in \autoref{fig_workflow}, the core functionality of `doped` is the generation of defect supercells and competing phases, writing calculation input files, parsing calculation outputs and analysing/plotting defect-related properties. This functionality and recommended usage of `doped` is demonstrated in the [tutorials](https://doped.readthedocs.io/en/latest/Tutorials.html) on the [documentation website](https://doped.readthedocs.io/en/latest/).
 
-![**a.** Average minimum periodic image distance, normalised by the ideal image distance at that volume (i.e. for a perfect close-packed face-centred cubic (FCC) cell), versus the number of primitive unit cells, for the supercell generation algorithms of `doped`, `ASE` and `pymatgen`. "SC" = simple cubic and "HCP" = hexagonal close-packed. **b.** Average performance of charge state guessing routines from `doped` compared to alternative approaches, in terms of false positives and false negatives. Asterisk indicates that the `pyCDT` false _negatives_ are underestimated as the majority of this test set used the guessed charge state ranges from `pyCDT`. "Ox." = oxidation & "prob." = probabilities. Example **(c)** Kumagai-Oba (extended Freysoldt-Neugebauer-Van-de-Walle; "eFNV") finite-size correction plot, **(d)** defect formation energy diagram, **(e)** chemical potential / stability region, **(f)** Fermi level vs. annealing temperature, **(g)** defect/carrier concentrations vs. annealing temperature and **(h)** Fermi level / carrier concentration heatmap plots from `doped`. Data and code to reproduce these plots is provided in the [`docs/JOSS`](https://github.com/SMTG-Bham/doped/blob/main/docs/JOSS) subfolder of the `doped` GitHub repository. \label{fig1}](doped_JOSS_figure.png)
[…]
