+++ docs/JOSS/paper.bib
@@ -826,3 +826,37 @@
   file = {/Users/kavanase/Zotero/storage/6TZVYDAH/Wang et al_2024_Upper efficiency limit of Sb2Se3 solar cells.pdf}
 }
 
+@article{yuan_discovery_2024,
+  title = {Discovery of the {{Zintl-phosphide BaCd2P2}} as a Long Carrier Lifetime and Stable Solar Absorber},
+  author = {Yuan, Zhenkun and Dahliah, Diana and Hasan, Muhammad Rubaiat and Kassa, Gideon and Pike, Andrew and Quadir, Shaham and Claes, Romain and Chandler, Cierra and Xiong, Yihuang and Kyveryga, Victoria and Yox, Philip and Rignanese, Gian-Marco and Dabo, Ismaila and Zakutayev, Andriy and Fenning, David P. and Reid, Obadiah G. and Bauers, Sage and Liu, Jifeng and Kovnir, Kirill and Hautier, Geoffroy},
+  year = {2024},
+  month = mar,
+  journal = {Joule},
+  issn = {2542-4351},
+  doi = {10.1016/j.joule.2024.02.017},
+  urldate = {2024-03-25},
+  abstract = {Thin-film photovoltaics (PV) offers a path to decarbonize global energy production. Unfortunately, existing thin-film solar absorbers have major issues associated with either elemental abundance, stability, or performance. Entirely new and disruptive materials platforms are rarely discovered, and their search is traditionally slow and serendipitous. Here, we report a first-principles high-throughput (HT) computational screening for new solar absorbers among 40,000 known inorganic materials. Next to band gap and carrier effective masses, we also use computed intrinsic defects as they can limit the carrier lifetime. We identify the Zintl-phosphide BaCd2P2 as a potential high-efficiency solar absorber. Follow-up experiments confirm the promises of BaCd2P2, highlighting an optimal band gap, bright photoluminescence (PL), and long carrier lifetime, even in unoptimized powder samples. Importantly, BaCd2P2 contains no critical elements and is stable in air and water. Our work demonstrates how computational screening combined with experiments can accelerate the search for photovoltaic materials.},
+  keywords = {carrier lifetime,first-principles calculations,high-throughput computational screening,intrinsic defects,nonradiative recombination,photoluminescence,synthesis,thin-film solar absorbers,time resolved microwave conductivity},
+}
+
+@article{broberg_high-throughput_2023,
+  title = {High-Throughput Calculations of Charged Point Defect Properties with Semi-Local Density Functional Theory---Performance Benchmarks for Materials Screening Applications},
+  author = {Broberg, Danny and Bystrom, Kyle and Srivastava, Shivani and Dahliah, Diana and Williamson, Benjamin A. D. and Weston, Leigh and Scanlon, David O. and Rignanese, Gian-Marco and Dwaraknath, Shyam and Varley, Joel and Persson, Kristin A. and Asta, Mark and Hautier, Geoffroy},
+  year = {2023},
+  month = may,
+  journal = {npj Computational Materials},
+  volume = {9},
+  number = {1},
+  pages = {1--12},
+  publisher = {Nature Publishing Group},
+  issn = {2057-3960},
+  doi = {10.1038/s41524-023-01015-6},
+  urldate = {2023-10-18},
[…]
