### 2024-03-25 c8fdfcc
Paper updates (JOSS review R1 Q15); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: b98a8fd0330d2d35bbd9031397a704a89294af6a [formerly f9aa0cc3d343cac7ce33b0590d45c920654123ab]
Former-commit-id: 301db69e086d206540e5ebe6748a8015656e068e
