### 2024-03-25 301db69
Paper updates (JOSS review R1 Q15); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: f9aa0cc3d343cac7ce33b0590d45c920654123ab
