### 2024-03-25 758fcdf
Avoid unnecessary oxi state guessing in `_get_unrelaxed_structure`
Former-commit-id: 2cafbbb3db6ff2196a0853fbe63fdde4cf5c8a37
