### 2024-03-25 c5020ae
Fix vise check for parsing PROCAR with SOC
