### 2024-03-25 cdaa265
Add `point_symmetry` function to pull relaxed & unrelaxed point symmetries directly from structures; (JOSS review R1 Q20); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 72b67c965625e4cf101543d461d873a1d9be8eb4
