+++ doped/utils/symmetry.py
@@ -12,6 +12,7 @@ from pymatgen.analysis.defects.core import DefectType
 from pymatgen.analysis.structure_matcher import ElementComparator, StructureMatcher
 from pymatgen.core.operations import SymmOp
 from pymatgen.core.structure import Lattice, PeriodicSite, Structure
+from pymatgen.entries.computed_entries import ComputedStructureEntry
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from pymatgen.transformations.standard_transformations import SupercellTransformation
 from pymatgen.util.coord import pbc_diff
@@ -652,7 +653,7 @@ def get_conv_cell_site(defect_entry):
     structure definition.
 
     Args:
-        defect_entry: DefectEntry object.
+        defect_entry: ``DefectEntry`` object.
     """
     bulk_prim_structure = defect_entry.defect.structure.copy()
     bulk_prim_structure.remove_oxidation_states()  # adding oxidation states adds the
@@ -1076,7 +1077,7 @@ def point_symmetry_from_defect_entry(
     return_periodicity_breaking: bool = False,
 ):
     r"""
-    Get the defect site point symmetry from a DefectEntry object.
+    Get the defect site point symmetry from a ``DefectEntry`` object.
 
     Note: If ``relaxed = True`` (default), then this tries to use the
     defect_entry.defect_supercell to determine the site symmetry. This will
@@ -1098,7 +1099,7 @@ def point_symmetry_from_defect_entry(
                   get_defect_name_from_entry(defect_entry), "\n")
 
     And if the point symmetries match in each case, then using this function on your
-    parsed `relaxed` DefectEntry objects should correctly determine the final relaxed
+    parsed `relaxed` ``DefectEntry`` objects should correctly determine the final relaxed
     defect symmetry - otherwise periodicity-breaking prevents this.
 
     If periodicity-breaking prevents auto-symmetry determination, you can manually
@@ -1115,15 +1116,15 @@ def point_symmetry_from_defect_entry(
     while for interstitials it is the point symmetry of the `final relaxed` interstitial
     site when placed in the (unrelaxed) bulk structure.
     The degeneracy factor is used in the calculation of defect/carrier concentrations
-    and Fermi level behaviour (see e.g. doi.org/10.1039/D2FD00043A &
-    doi.org/10.1039/D3CS00432E).
+    and Fermi level behaviour (see e.g. https://doi.org/10.1039/D2FD00043A &
+    https://doi.org/10.1039/D3CS00432E).
 
     Args:
-        defect_entry (DefectEntry): DefectEntry object.
+        defect_entry (DefectEntry): ``DefectEntry`` object.
         symm_ops (list):
             List of symmetry operations of either the defect_entry.bulk_supercell
-            structure (if relaxed=False) or defect_entry.defect_supercell (if
-            relaxed=True), to avoid re-calculating. Default is None (recalculates).
+            structure (if ``relaxed=False``) or defect_entry.defect_supercell (if
+            ``relaxed=True``), to avoid re-calculating. Default is None (recalculates).
         symprec (float):
             Symmetry tolerance for ``spglib``. Default is 0.01 for unrelaxed structures,
             0.1 for relaxed (to account for residual structural noise, matching that
@@ -1377,6 +1378,147 @@ def _check_relaxed_defect_symmetry_determination(
         return True
 
     return False  # return False if symmetry couldn't be checked
+
+
+def point_symmetry(
+    structure: Structure,
+    bulk_structure: Optional[Structure] = None,
+    symm_ops: Optional[list] = None,
+    symprec: Optional[float] = None,
+    relaxed: bool = True,
+    verbose: bool = True,
+    return_periodicity_breaking: bool = False,
+):
+    r"""
[…]
