+++ doped/thermodynamics.py
@@ -317,10 +317,10 @@ def group_defects_by_name(entry_list: list[DefectEntry]) -> dict[str, list[Defec
     ``{defect name without charge: [DefectEntry]}``, where the values are lists
     of DefectEntry objects with the same defect name (excluding charge state).
 
-    The DefectEntry.name attributes are used to get the defect names.
+    The ``DefectEntry.name`` attributes are used to get the defect names.
     These should be in the format:
     "{defect_name}_{optional_site_info}_{charge_state}".
-    If the DefectEntry.name attribute is not defined or does not end with the
+    If the ``DefectEntry.name`` attribute is not defined or does not end with the
     charge state, then the entry will be renamed with the doped default name
     for the `unrelaxed` defect (i.e. using the point symmetry of the defect
     site in the bulk cell).
@@ -389,7 +389,7 @@ class DefectThermodynamics(MSONable):
         can also be initialised with a list or dict of DefectEntry objects (e.g.
         from DefectsParser.defect_dict).
 
-        Note that the DefectEntry.name attributes are used to label the defects in
+        Note that the ``DefectEntry.name`` attributes are used to label the defects in
         plots.
 
         Args:
@@ -397,7 +397,7 @@ class DefectThermodynamics(MSONable):
                 A list or dict of DefectEntry objects. Note that ``DefectEntry.name``
                 attributes are used for grouping and plotting purposes! These should
                 be in the format "{defect_name}_{optional_site_info}_{charge_state}".
-                If the DefectEntry.name attribute is not defined or does not end with
+                If the ``DefectEntry.name`` attribute is not defined or does not end with
                 the charge state, then the entry will be renamed with the doped
                 default name.
             chempots (dict):
@@ -962,7 +962,7 @@ class DefectThermodynamics(MSONable):
                 DefectThermodynamics.defect_entries list. Note that ``DefectEntry.name``
                 attributes are used for grouping and plotting purposes! These should
                 be in the format "{defect_name}_{optional_site_info}_{charge_state}".
-                If the DefectEntry.name attribute is not defined or does not end with
+                If the ``DefectEntry.name`` attribute is not defined or does not end with
                 the charge state, then the entry will be renamed with the doped
                 default name.
             check_compatibility (bool):
@@ -1171,10 +1171,11 @@ class DefectThermodynamics(MSONable):
         docstring for more information).
 
         The degeneracy/multiplicity factor "g" is an important parameter in the defect
-        concentration equation (see discussion in doi.org/10.1039/D2FD00043A and
-        doi.org/10.1039/D3CS00432E), affecting the final concentration by up to 2 orders
-        of magnitude. This factor is taken from the product of the
-        defect_entry.defect.multiplicity and defect_entry.degeneracy_factors attributes.
+        concentration equation (see discussion in https://doi.org/10.1039/D2FD00043A
+        and https://doi.org/10.1039/D3CS00432E), affecting the final concentration by
+        up to ~2 orders of magnitude. This factor is taken from the product of the
+        ``defect_entry.defect.multiplicity`` and ``defect_entry.degeneracy_factors``
+        attributes.
 
         Args:
             chempots (dict):
@@ -1361,10 +1362,10 @@ class DefectThermodynamics(MSONable):
 
         The degeneracy/multiplicity factor "g" is an important parameter in the defect
         concentration equation and thus Fermi level calculation (see discussion in
-        doi.org/10.1039/D2FD00043A and doi.org/10.1039/D3CS00432E), affecting the
[…]
