### 2024-03-25 dff4c6b
Add `point_symmetry` function to pull relaxed & unrelaxed point symmetries directly from structures; (JOSS review R1 Q20); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 28c0e1de1344b9ecbaee1c550b4d6c4f69d0189c [formerly 72b67c965625e4cf101543d461d873a1d9be8eb4]
Former-commit-id: cdaa2653906e8cb4faf80ee37e22fa11d3f3ae35
