### 2024-03-25 bfaa804
Add `point_symmetry` tests
