+++ tests/test_thermodynamics.py
@@ -21,6 +21,7 @@ from monty.serialization import dumpfn, loadfn
 
 from doped.generation import _sort_defect_entries
 from doped.thermodynamics import DefectThermodynamics, scissor_dos
+from doped.utils.symmetry import _get_sga, point_symmetry
 
 # for pytest-mpl:
 module_path = os.path.dirname(os.path.abspath(__file__))
@@ -1139,7 +1140,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
                         )
 
         for i, row in enumerate(cdte_form_en_lists):
-            assert list(form_en_df.iloc[i]) == row
+            assert list(form_en_df.iloc[i])[:-1] == row[:-1]  # compare everything except path
             _check_formation_energy_methods(row, self.CdTe_defect_thermo, 0.7493)  # default mid-gap value
 
         with pytest.raises(ValueError) as exc:
@@ -1171,7 +1172,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
             else:
                 print(f"No VBM warnings for {new_entry.name}")  # v_Cd_0
                 assert "No chemical potentials supplied" in str(w[0].message)
-                assert not any("VBM eigenvalue was not set" in str(i.message) for i in w)
+                assert all("VBM eigenvalue was not set" not in str(i.message) for i in w)
 
         non_formatted_form_en_df, output, w = _run_func_and_capture_stdout_warnings(
             self.CdTe_defect_thermo.get_formation_energies, skip_formatting=True
@@ -1180,7 +1181,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
         self._check_chempot_w_and_fermi_message(w, output)
         for i, row in enumerate(cdte_form_en_lists):
             row[1] = int(row[1])
-            assert list(non_formatted_form_en_df.iloc[i]) == row  # and all other terms the same
+            assert list(non_formatted_form_en_df.iloc[i])[:-1] == row[:-1]  # and all other terms the same
             _check_formation_energy_methods(row, self.CdTe_defect_thermo, 0.7493)  # default mid-gap value
 
         # with chempots:
@@ -1325,7 +1326,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
         self._check_form_en_df(te_rich_df, fermi_level=0.7493, defect_thermo=self.CdTe_defect_thermo)
 
         for i, row in enumerate(cdte_te_rich_form_en_lists):
-            assert list(te_rich_df.iloc[i]) == row
+            assert list(te_rich_df.iloc[i])[:-1] == row[:-1]  # compare everything except path
             _check_formation_energy_methods(row, self.CdTe_defect_thermo, 0.7493)  # default mid-gap
             # value, chempots also now attached
 
@@ -1346,7 +1347,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
             self._check_no_w_fermi_message_and_new_matches_ref_df(output, w, df, non_formatted_te_rich_df)
         for i, row in enumerate(cdte_te_rich_form_en_lists):
             row[1] = int(row[1])
-            assert list(non_formatted_te_rich_df.iloc[i]) == row
+            assert list(non_formatted_te_rich_df.iloc[i])[:-1] == row[:-1]
             _check_formation_energy_methods(row, self.CdTe_defect_thermo, 0.7493)
 
         # hard test random case with random chempots, el_refs and fermi level
@@ -1374,7 +1375,7 @@ class DefectThermodynamicsTestCase(DefectThermodynamicsSetupMixin):
[…]
