### 2024-03-25 cd44611
Add `point_symmetry` tests
Former-commit-id: bfaa8043c373f1aa4e6ca2e756aada7bbcb7f18d
