### 2024-03-25 393286a
Add tip to thermo tutorial about multiplicity being defined wrt primitive cell (JOSS review R1 Q14); https://github.com/openjournals/joss-reviews/issues/6433
