### 2024-03-25 ca930e1
Add tip to thermo tutorial about multiplicity being defined wrt primitive cell (JOSS review R1 Q14); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: efd3aba446a6703f0fb93c527b9339b5a2542d77 [formerly 393286ad0f80c9ddb0b65910345f301c352ce015]
Former-commit-id: 6262be6fd7eddc9b8063f4ea678107f13301b25b
