### 2024-03-25 6262be6
Add tip to thermo tutorial about multiplicity being defined wrt primitive cell (JOSS review R1 Q14); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 393286ad0f80c9ddb0b65910345f301c352ce015
