### 2024-03-25 7d15569
Add mini sections to tips docs and advanced analysis tutorial about `point_symmetry` direct with structures (JOSS review R1 Q20); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 6b76fe370935387f49c1aef66b455903b6b53bb0
