### 2024-03-25 eabaaee
Add mini sections to tips docs and advanced analysis tutorial about `point_symmetry` direct with structures (JOSS review R1 Q20); https://github.com/openjournals/joss-reviews/issues/6433
Former-commit-id: 740e3faa8457e646aeca9ea70185631f50177875 [formerly 6b76fe370935387f49c1aef66b455903b6b53bb0]
Former-commit-id: 7d15569c5f9f6216b5b48873615129ff2eae7ccf
