### 2024-03-25 6b76fe3
Add mini sections to tips docs and advanced analysis tutorial about `point_symmetry` direct with structures (JOSS review R1 Q20); https://github.com/openjournals/joss-reviews/issues/6433
