### 2024-03-26 04ea5f0
Suggested refactoring ("phs" -> "eigenvalues" etc) and some formatting
