+++ doped/analysis.py
@@ -51,7 +51,7 @@ from doped.utils.parsing import (
     get_outcar,
     get_vasprun,
 )
-from doped.utils.phs import get_band_edge_info
+from doped.utils.eigenvalues import get_band_edge_info
 from doped.utils.plotting import format_defect_name
 from doped.utils.symmetry import (
     _frac_coords_sort_func,
@@ -446,7 +446,7 @@ class DefectsParser:
         bulk_band_gap_path: Optional[str] = None,
         processes: Optional[int] = None,
         json_filename: Optional[Union[str, bool]] = None,
-        load_phs_data: Optional[bool] = True,
+        parse_projected_eigen: Optional[bool] = True,
         **kwargs,
     ):
         r"""
@@ -485,7 +485,7 @@ class DefectsParser:
                 Ionic + static contributions to the dielectric constant, in the same xyz
                 Cartesian basis as the supercell calculations. If not provided, charge
                 corrections cannot be computed and so ``skip_corrections`` will be set to
-                true.
+                ``True``.
             subfolder (str):
                 Name of subfolder(s) within each defect calculation folder (in the
                 ``output_path`` directory) containing the VASP calculation files to
@@ -511,10 +511,10 @@ class DefectsParser:
                 occur at reciprocal space points not included in the bulk supercell
                 calculation, you should use this tag to point to a bulk bandstructure
                 calculation instead. Alternatively, you can edit/add the "gap" and "vbm"
-                entries in DefectParser.defect_entry.calculation_metadata to match the
+                entries in ``DefectParser.defect_entry.calculation_metadata`` to match the
                 correct (eigen)values.
-                If None, will calculate "gap"/"vbm" using the outputs at:
-                DefectParser.defect_entry.calculation_metadata["bulk_path"]
+                If ``None`` (default), will calculate "gap"/"vbm" using the outputs at:
+                ``DefectParser.defect_entry.calculation_metadata["bulk_path"]``
             processes (int):
                 Number of processes to use for multiprocessing for expedited parsing.
                 If not set, defaults to one less than the number of CPUs available.
@@ -523,16 +523,15 @@ class DefectsParser:
                 to in ``output_path``, to avoid having to re-parse defects when later analysing
                 further and aiding calculation provenance. Can be reloaded using the ``loadfn``
                 function from ``monty.serialization`` (and then input to ``DefectThermodynamics``
-                etc). If None (default), set as ``{Host Chemical Formula}_defect_dict.json``.
-                If False, no json file is saved.
-            load_phs_data (bool):
-                Load the projected eigenvalues and sets up a ``pydefect`` ``BandEdgeOrbitalInfos``
-                object with the required information for determining the band edge states
-                using ``DefectEntry.get_perturbed_host_state()``. Will initially try to load
-                from ``PROCAR`` files, otherwise will fall back on loading project eigenvalues
-                from ``vasprun.xml``.This can cause a significant increase in the parsing
-                time (30%-70% longer), so set to False if not needed.
-                Default = True.
+                etc.). If ``None`` (default), set as ``{Host Chemical Formula}_defect_dict.json``.
+                If ``False``, no json file is saved.
+            parse_projected_eigen (bool):
+                Whether to parse the projected eigenvalues & orbitals from the bulk and defect
+                calculations, allowing ``DefectEntry.get_eigenvalue_analysis()`` to be used.
+                Will initially try to load from ``PROCAR`` files in the bulk/defect directories
+                if present, otherwise will attempt to load from ``vasprun.xml(.gz)`` (~50%
[…]
