+++ tests/test_analysis.py
@@ -39,7 +39,7 @@ from doped.utils.symmetry import point_symmetry
 mpl.use("Agg")  # don't show interactive plots if testing from CLI locally
 
 
-# TODO: Remove `load_phs_data = False` when new `vise` version released
+# TODO: Remove `parse_projected_eigen = False` when new `vise` version released
 
 
 def if_present_rm(path):
@@ -324,7 +324,7 @@ class DefectsParsingTestCase(unittest.TestCase):
                 output_path=self.CdTe_EXAMPLE_DIR,
                 dielectric=9.13,
                 json_filename="CdTe_example_defect_dict.json",
-                load_phs_data=False,
+                parse_projected_eigen=False,
             )  # for testing in test_thermodynamics.py
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(default_dp, w)
@@ -350,7 +350,7 @@ class DefectsParsingTestCase(unittest.TestCase):
         # test same behaviour without multiprocessing:
         with warnings.catch_warnings(record=True) as w:
             dp = DefectsParser(
-                output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, processes=1, load_phs_data=False
+                output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, processes=1, parse_projected_eigen=False
             )
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w)
@@ -364,7 +364,7 @@ class DefectsParsingTestCase(unittest.TestCase):
         # check using filterwarnings works as expected:
         warnings.filterwarnings("ignore", "Multiple")
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, load_phs_data=False)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, parse_projected_eigen=False)
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, multiple_outcars_warning=False)
         warnings.filterwarnings("default", "Multiple")
@@ -377,14 +377,14 @@ class DefectsParsingTestCase(unittest.TestCase):
         # test with reduced dist_tol:
         # Int_Te_3_Unperturbed merged with Int_Te_3 with default dist_tol = 1.5, now no longer merged
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, load_phs_data=False)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, parse_projected_eigen=False)
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, dist_tol=0.1)
 
     def test_DefectsParser_CdTe_no_dielectric_json(self):
         # test no dielectric and no JSON:
         with warnings.catch_warnings(record=True) as w:
-            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, json_filename=False, load_phs_data=False)
+            dp = DefectsParser(output_path=self.CdTe_EXAMPLE_DIR, json_filename=False, parse_projected_eigen=False)
         print([warn.message for warn in w])  # for debugging
         assert any(
             "The dielectric constant (`dielectric`) is needed to compute finite-size charge "
@@ -405,7 +405,7 @@ class DefectsParsingTestCase(unittest.TestCase):
[…]

[…]
