### 2024-03-26 bcfb270
Suggested refactoring ("phs" -> "eigenvalues" etc) and some formatting
Former-commit-id: 04ea5f0362c477bef15ba2a679b0b84869e03788
