+++ docs/JOSS/paper.md
@@ -121,7 +121,7 @@ Some key advances of `doped` include:
   `doped` aims to improve the efficiency of this step by querying the [Materials Project](https://materialsproject.org) database (containing both experimentally-measured and theoretically-predicted crystal structures), and pulling only compounds which _could border the host material_ within a user-specified error tolerance for the semi-local DFT database energies (0.1 eV/atom by default), along with the elemental reference phases. The necessary _k_-point convergence step for these compounds is also implemented in a semi-automated fashion to expedite this process.
   - With the parsed chemical potentials in `doped`, the user can easily select various X-poor/rich chemical conditions, or scan over a range of chemical potentials (growth conditions) as shown in \autoref{fig1}e,h.
 
-- **Automated Symmetry & Degeneracy Handling:** `doped` automatically determines the point symmetry of both initial (un-relaxed) and final (relaxed) defect configurations, and computes the corresponding orientational (and spin) degeneracy factors. This is a key pre-factor in the defect concentration equation:
+- **Automated Symmetry & Degeneracy Handling:** `doped` automatically determines the point symmetry of both initial (un-relaxed) and final (relaxed) defect configurations, and computes the corresponding orientational (and spin) degeneracy factors. This functionality is also offered in the form of [standalone functions](https://doped.readthedocs.io/en/latest/advanced_analysis_tutorial.html#point-symmetry-analysis) which do not require the defect calculations to have been generated/parsed with `doped`. This is a key pre-factor in the defect concentration equation:
 
   \begin{equation}
   N_D = gN_s \exp(-E_f/k_BT)
@@ -141,13 +141,13 @@ Some key advances of `doped` include:
 
 - **[`ShakeNBreak`](https://shakenbreak.readthedocs.io):** `doped` is natively interfaced with our defect structure-searching code `ShakeNBreak` [@mosquera-lois_shakenbreak_2022], seamlessly incorporating this phase in the defect calculation workflow. This step can optionally be skipped or an alternative structure-searching approach readily implemented.
 
-Some additional features of `doped` include directional-dependent site displacement (local strain) analysis, deterministic & informative defect naming, molecule generation for gaseous competing phases, multiprocessing for expedited generation & parsing, shallow defect analysis (via `pydefect` [@Kumagai2021]), Wyckoff site analysis (including arbitrary/interstitial sites), controllable defect site placement to aid visualisation and more.
+Some additional features of `doped` include directional-dependent site displacement (local strain) analysis, deterministic & informative defect naming, molecule generation for gaseous competing phases, multiprocessing for expedited generation & parsing, shallow defect analysis (via `pydefect` [@Kumagai2021]), Wyckoff site analysis (including _arbitrary/interstitial_ sites), controllable defect site placement to aid visualisation and more.
 
 The defect generation and thermodynamic analysis components of `doped` are agnostic to the underlying software used for the defect supercell calculations.
 Direct calculation I/O is fully-supported for `VASP` [@vasp], while input defect structure files can be generated for several widely-used DFT codes, including `FHI-aims` [@fhi_aims], `CP2K` [@cp2k], `Quantum Espresso` [@espresso] and `CASTEP` [@castep] via the `pymatgen` `Structure` object. Full support for calculation I/O with other DFT codes may be added in the future if there is sufficient demand.
[…]
