+++ docs/JOSS/paper.bib
@@ -203,20 +203,19 @@
   file = {/Users/kavanase/Zotero/storage/NG4MIPJX/Larsen et al_2017_The atomic simulation environment—a Python library for working with atoms.pdf}
 }
 
-@misc{li_computational_2023,
-  title = {Computational {{Prediction}} of an {{Antimony-based}} n-Type {{Transparent Conducting Oxide}}: {{F-doped Sb2O5}}},
-  shorttitle = {Computational {{Prediction}} of an {{Antimony-based}} n-Type {{Transparent Conducting Oxide}}},
+@article{li_computational_2024,
+  title = {Computational {{Prediction}} of an {{Antimony-Based}} n-{{Type Transparent Conducting Oxide}}: {{F-Doped Sb2O5}}},
+  shorttitle = {Computational {{Prediction}} of an {{Antimony-Based}} n-{{Type Transparent Conducting Oxide}}},
   author = {Li, Ke and Willis, Joe and Kavanagh, Se{\'a}n R. and Scanlon, David O.},
-  year = {2023},
-  month = dec,
-  publisher = {{ChemRxiv}},
-  doi = {10.26434/chemrxiv-2023-8l7pb},
-  urldate = {2024-01-18},
-  abstract = {Transparent conducting oxides (TCOs) possess a unique combination of optical transparency and electrical conductivity, making them indispensable in optoelectronic applications. However, the heavy dependence on a small number of established materials limits the range of devices they can support. The discovery and development of additional wide bandgap oxides that can be doped to display metallic-like conductivity is therefore necessary. In this work, we use hybrid density functional theory to identify a binary Sb(V) system, Sb2O5, as a promising TCO with high conductivity and transparency when doped with fluorine. We have conducted a full point defect analysis, finding F-doped Sb2O5 to exhibit degenerate n-type transparent conducting behavior. The inherently large electron affinity found in antimony oxides also widens its application in organic solar cells. Following our previous work on zinc antimonate, this work provides additional support for designing Sb(V)-based oxides as cost-effective transparent conducting oxides for a broader range of applications.},
-  archiveprefix = {ChemRxiv},
-  langid = {english},
-  keywords = {Optoelectronics,Point defects,Transparent Conducting Oxides},
-  file = {/Users/kavanase/Zotero/storage/GE7QAA9E/Li et al. - 2023 - Computational Prediction of an Antimony-based n-ty.pdf}
+  year = {2024},
+  month = mar,
+  journal = {Chemistry of Materials},
+  publisher = {American Chemical Society},
+  issn = {0897-4756},
+  doi = {10.1021/acs.chemmater.3c03257},
+  urldate = {2024-03-25},
+  abstract = {Transparent conducting oxides (TCOs) possess a unique combination of optical transparency and electrical conductivity, making them indispensable in optoelectronic applications. However, their heavy dependence on a small number of established materials limits the range of devices that they can support. The discovery and development of additional wide bandgap oxides that can be doped to exhibit metallic-like conductivity are therefore necessary. In this work, we use hybrid density functional theory to identify a binary Sb(V) system, Sb2O5, as a promising TCO with high conductivity and transparency when doped with fluorine. We conducted a full point defect analysis, finding F-doped Sb2O5 to exhibit degenerate n-type transparent conducting behavior. The inherently large electron affinity found in antimony oxides also widens their application in organic solar cells. Following our previous work on zinc antimonate, this work provides additional support for designing Sb(V)-based oxides as cost-effective TCOs for a broader range of applications.},
+  file = {/Users/kavanase/Zotero/storage/75HBJQ28/Li et al_2024_Computational Prediction of an Antimony-Based n-Type Transparent Conducting.pdf}
 }
 
 @article{liga_mixed-cation_2023,
