### 2024-03-26 bac6083
Add `get_procar` to handle `PROCAR` loading
