### 2024-03-26 e508eb9
Add `_parse_procar`, expand docstrings, avoid unnecessary `pydefect` `INFO` warning, allow `get_eigenvalue_analysis` to be used directly (not necessarily `doped` `DefectEntry` object)...
