+++ doped/utils/eigenvalues.py
@@ -5,41 +5,59 @@ Contains modified versions of functions from pydefect (https://github.com/kumaga
 and vise (https://github.com/kumagai-group/vise), to avoid requiring additional files (i.e. ``PROCAR``s).
 """
 
-import logging
 import os
 from collections import defaultdict
 from importlib.metadata import version
 from itertools import zip_longest
-from typing import Any, Optional
+from typing import Any, Optional, Union
+import warnings
 
 import matplotlib.pyplot as plt
 import numpy as np
 
-# ruff: noqa: E402
-from vise import user_settings
-
-user_settings.logger.setLevel(logging.CRITICAL)
-
-import pydefect.analyzer.make_band_edge_states
-import pydefect.cli.vasp.make_band_edge_orbital_infos as make_bes
-from easyunfold.procar import Procar as ProcarEasyunfold
-from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
-from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
-from pydefect.analyzer.make_band_edge_states import make_band_edge_states
-from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
-from pydefect.cli.vasp.make_perfect_band_edge_state import (
-    get_edge_info,
-    make_perfect_band_edge_state_from_vasp,
-)
-from pydefect.defaults import defaults
-from pydefect.util.structure_tools import Coordination, Distances
 from pymatgen.core import Element, Species
 from pymatgen.electronic_structure.core import Spin
-from pymatgen.io.vasp.outputs import Outcar, Vasprun
+from pymatgen.io.vasp.outputs import Outcar, Vasprun, Procar
+from pymatgen.entries.computed_entries import ComputedStructureEntry
 from shakenbreak.plotting import _install_custom_font
-from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
 
 from doped.utils.plotting import _get_backend
+from doped import _ignore_pmg_warnings
+from doped.core import DefectEntry
+from doped.utils.parsing import get_procar
+
+
+# suppress pydefect INFO messages
+import logging
+
+try:
+    from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
+    import pydefect.analyzer.make_band_edge_states
+    import pydefect.cli.vasp.make_band_edge_orbital_infos as make_bes
+    from pydefect.analyzer.band_edge_states import BandEdgeOrbitalInfos, OrbitalInfo, PerfectBandEdgeState
+    from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
+    from pydefect.analyzer.make_band_edge_states import make_band_edge_states
+    from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
+    from pydefect.cli.vasp.make_perfect_band_edge_state import (
+        get_edge_info,
+        make_perfect_band_edge_state_from_vasp,
+    )
+    from pydefect.defaults import defaults
+    from pydefect.util.structure_tools import Coordination, Distances
+    from vise import user_settings
+    user_settings.logger.setLevel(logging.CRITICAL)
+
+except ImportError as exc:
+    raise ImportError(
+        "To perform eigenvalue & orbital analysis, you need to install pydefect. "
+        "You can do this by running `pip install pydefect`."
+    ) from exc
+
+# vise suppresses `UserWarning`s, so need to reset
+warnings.simplefilter("default")
+warnings.filterwarnings("ignore", message="`np.int` is a deprecated alias for the builtin `int`")
+warnings.filterwarnings("ignore", message="Use get_magnetic_symmetry()")
[…]
