### 2024-03-26 464253a
Add `_load_and_parse_eigenvalue_data` to `DefectEntry` to efficiently handle eigenvalue parsing, expand `get_eigenvalue_analysis` docstrings, and update so that eigenvalue data still in correct format after saving/reloading
