+++ doped/core.py
@@ -18,7 +18,7 @@ from pymatgen.analysis.defects.utils import CorrectionResult
 from pymatgen.core.composition import Composition, Element
 from pymatgen.core.structure import PeriodicSite, Structure
 from pymatgen.entries.computed_entries import ComputedEntry, ComputedStructureEntry
-from pymatgen.io.vasp.outputs import Locpot, Outcar
+from pymatgen.io.vasp.outputs import Locpot, Outcar, Vasprun, Procar
 from scipy.constants import value as constants_value
 from scipy.stats import sem
 
@@ -223,8 +223,13 @@ class DefectEntry(thermo.DefectEntry):
         """
         # ignore warning about oxidation states not summing to Structure charge:
         warnings.filterwarnings("ignore", message=".*unset_charge.*")
+        self_dict = asdict(self)
+        if self.calculation_metadata and self.calculation_metadata.get("eigenvalue_data"):
+            for key in list(self.calculation_metadata["eigenvalue_data"].keys()):
+                self_dict["calculation_metadata"]["eigenvalue_data"][key] = self.calculation_metadata[
+                    "eigenvalue_data"][key].as_dict()
 
-        return asdict(self)
+        return self_dict
 
     def _check_correction_error_and_return_output(
         self,
@@ -287,17 +292,17 @@ class DefectEntry(thermo.DefectEntry):
                 then the dielectric constant is taken from the ``defect_entry``
                 ``calculation_metadata`` if available.
             defect_locpot:
-                Path to the output VASP LOCPOT file from the defect supercell
-                calculation, or the corresponding ``pymatgen`` Locpot object, or
-                a dictionary of the planar-averaged potential in the form:
-                {i: Locpot.get_average_along_axis(i) for i in [0,1,2]}.
+                Path to the output VASP ``LOCPOT`` file from the defect supercell
+                calculation, or the corresponding ``pymatgen`` ``Locpot``
+                object, or a dictionary of the planar-averaged potential in the
+                form: ``{i: Locpot.get_average_along_axis(i) for i in [0,1,2]}``.
                 If None, will try to use ``defect_locpot`` from the
                 ``defect_entry`` ``calculation_metadata`` if available.
             bulk_locpot:
-                Path to the output VASP LOCPOT file from the bulk supercell
-                calculation, or the corresponding ``pymatgen`` Locpot object, or
-                a dictionary of the planar-averaged potential in the form:
-                {i: Locpot.get_average_along_axis(i) for i in [0,1,2]}.
+                Path to the output VASP ``LOCPOT`` file from the bulk supercell
+                calculation, or the corresponding ``pymatgen`` ``Locpot``
+                object, or a dictionary of the planar-averaged potential in the
+                form: ``{i: Locpot.get_average_along_axis(i) for i in [0,1,2]}``.
                 If None, will try to use ``bulk_locpot`` from the
                 ``defect_entry`` ``calculation_metadata`` if available.
             plot (bool):
@@ -512,34 +517,292 @@ class DefectEntry(thermo.DefectEntry):
             error_tolerance=error_tolerance,
         )
 
-    def get_eigenvalue_analysis(self, filename: Optional[str] = None, plot: bool = True, **kwargs):
+    def _load_and_parse_eigenvalue_data(
+        self,
+        bulk_vr: Optional[Union[str, "Path", Vasprun]] = None,
+        bulk_procar: Optional[Union[str, "Path", "EasyunfoldProcar", Procar]] = None,
+        defect_vr: Optional[Union[str, "Path", Vasprun]] = None,
+        defect_procar: Optional[Union[str, "Path", "EasyunfoldProcar", Procar]] = None,
+        bulk_outcar: Optional[Union[str, "Path", Outcar]] = None,
+        force_reparse: bool = False,
+    ):
         """
[…]
