### 2024-03-26 5c44260
Update eigenvalue data parsing in `analysis.py` accordingly
