+++ doped/analysis.py
@@ -1,7 +1,7 @@
 """
 Code to analyse VASP defect calculations.
 
-These functions are built from a combination of useful modules from pymatgen,
+These functions are built from a combination of useful modules from ``pymatgen``,
 alongside substantial modification, in the efforts of making an efficient,
 user-friendly package for managing and analysing defect calculations, with
 publication-quality outputs.
@@ -14,24 +14,21 @@ from multiprocessing import Pool, Queue, cpu_count
 from typing import Optional, Union
 
 import numpy as np
-from easyunfold.procar import Procar as ProcarEasyunfold
 from filelock import FileLock
 from monty.json import MontyDecoder
 from monty.serialization import dumpfn, loadfn
 from pymatgen.analysis.defects import core
 from pymatgen.analysis.structure_matcher import ElementComparator, StructureMatcher
 from pymatgen.core.sites import PeriodicSite
-from pymatgen.electronic_structure.core import Spin
 from pymatgen.ext.matproj import MPRester
 from pymatgen.io.vasp.inputs import Poscar
-from pymatgen.io.vasp.outputs import Vasprun
+from pymatgen.io.vasp.outputs import Vasprun, Procar, Outcar
 from tqdm import tqdm
 
 from doped import _doped_obj_properties_methods, _ignore_pmg_warnings
 from doped.core import DefectEntry, _guess_and_set_oxi_states_with_timeout, _rough_oxi_state_cost_from_comp
 from doped.generation import get_defect_name_from_defect, get_defect_name_from_entry, name_defect_entries
 from doped.thermodynamics import DefectThermodynamics
-from doped.utils.eigenvalues import get_band_edge_info
 from doped.utils.parsing import (
     _compare_incar_tags,
     _compare_kpoints,
@@ -51,6 +48,7 @@ from doped.utils.parsing import (
     get_orientational_degeneracy,
     get_outcar,
     get_vasprun,
+    get_procar,
 )
 from doped.utils.plotting import format_defect_name
 from doped.utils.symmetry import (
@@ -446,7 +444,7 @@ class DefectsParser:
         bulk_band_gap_path: Optional[str] = None,
         processes: Optional[int] = None,
         json_filename: Optional[Union[str, bool]] = None,
-        parse_projected_eigen: Optional[bool] = True,
+        parse_projected_eigen: Optional[bool] = None,
         **kwargs,
     ):
         r"""
@@ -454,84 +452,92 @@ class DefectsParser:
         for a given host (bulk) material.
 
         Loops over calculation directories in ``output_path`` (likely the same
-        ``output_path`` used with ``DefectsSet`` for file generation in ``doped.vasp``)
-        and parses the defect calculations into a dictionary of:
-        ``{defect_name: DefectEntry}``, where the ``defect_name`` is set to the defect
-        calculation folder name (`if it is a recognised defect name`), else it is
-        set to the default ``doped`` name for that defect. By default, searches for
-        folders in ``output_path`` with ``subfolder`` containing ``vasprun.xml(.gz)``
-        files, and tries to parse them as ``DefectEntry``\s.
+        ``output_path`` used with ``DefectsSet`` for file generation in
+        ``doped.vasp``) and parses the defect calculations into a dictionary of:
+        ``{defect_name: DefectEntry}``, where the ``defect_name`` is set to the
+        defect calculation folder name (`if it is a recognised defect name`),
+        else it is set to the default ``doped`` name for that defect. By default,
+        searches for folders in ``output_path`` with ``subfolder`` containing
+        ``vasprun.xml(.gz)`` files, and tries to parse them as ``DefectEntry``\s.
 
-        By default, tries to use multiprocessing to speed up defect parsing, which
[…]
