### 2024-03-26 2b331d5
Pre-commit and formatting updates
