+++ docs/docs_requirements.txt
@@ -22,4 +22,4 @@ shakenbreak>=3.3.1
 packaging
 pandas>=1.1.0
 pydefect>=0.8.1
-vise>=0.8.1
+vise>=0.9.0
