### 2024-03-27 ceff0b6
Update `vise` requirement following release
Former-commit-id: c894eb6ae07a7be9f90fa59540b7bc252a6fa1e4
