+++ docs/requirements.txt
@@ -10,7 +10,7 @@ shakenbreak>=3.3.1
 packaging
 pandas>=1.1.0
 pydefect>=0.8.1
-vise>=0.8.1
+vise>=0.9.0
 
 # optional dependencies:
 py-sc-fermi
