### 2024-03-27 c894eb6
Update `vise` requirement following release
