### 2024-03-27 10a4c69
Remove vise version checking test, updated new tests
Former-commit-id: 63f97aacd43331e6072549f6ba1f1d51d4750628
