### 2024-03-27 972ca21
non-collinear test eigenvalue parsing
