### 2024-03-27 62b4783
Set `DefectEntry` name during initialisation to avoid unnecessary guessing before later resetting
