### 2024-03-27 d6ec439
Add `lru_cache` for getting potcar stats for neutral nelect guessing, to massively speed up charge state guessing
Former-commit-id: d88f4e49ccced4f742296775e5bdafdc0b418c4a
