### 2024-03-27 e07246e
Add `get_magnetization_from_vasprun` and `get_nelect_from_vasprun` functions to `doped.utils.parsing`
Former-commit-id: d3ea93594ad327dd8f45794bc380c872fa327466
