### 2024-03-27 5da93bc
Add `band_edge_properties_from_vasprun` and `make_perfect_band_edge_state_from_vasp` functions to avoid need for `Outcar` with eigenvalue analysis functions
Former-commit-id: 4da45aca5aef782783b949aa71640fddf258d8aa
