### 2024-03-27 4da45ac
Add `band_edge_properties_from_vasprun` and `make_perfect_band_edge_state_from_vasp` functions to avoid need for `Outcar` with eigenvalue analysis functions
