+++ doped/utils/eigenvalues.py
@@ -19,12 +19,12 @@ import numpy as np
 from pymatgen.core import Element, Species
 from pymatgen.electronic_structure.core import Spin
 from pymatgen.entries.computed_entries import ComputedStructureEntry
-from pymatgen.io.vasp.outputs import Outcar, Procar, Vasprun
+from pymatgen.io.vasp.outputs import Procar, Vasprun
 from shakenbreak.plotting import _install_custom_font
 
 from doped import _ignore_pmg_warnings
 from doped.core import DefectEntry
-from doped.utils.parsing import get_procar
+from doped.utils.parsing import get_magnetization_from_vasprun, get_nelect_from_vasprun, get_procar
 from doped.utils.plotting import _get_backend
 
 if TYPE_CHECKING:
@@ -38,14 +38,11 @@ try:
     from pydefect.analyzer.eigenvalue_plotter import EigenvalueMplPlotter
     from pydefect.analyzer.make_band_edge_states import make_band_edge_states
     from pydefect.analyzer.make_defect_structure_info import MakeDefectStructureInfo
-    from pydefect.cli.vasp.make_perfect_band_edge_state import (
-        get_edge_info,
-        make_perfect_band_edge_state_from_vasp,
-    )
+    from pydefect.cli.vasp.make_perfect_band_edge_state import get_edge_info
     from pydefect.defaults import defaults
     from pydefect.util.structure_tools import Coordination, Distances
     from vise import user_settings
-    from vise.analyzer.vasp.band_edge_properties import VaspBandEdgeProperties
+    from vise.analyzer.vasp.band_edge_properties import BandEdgeProperties, eigenvalues_from_vasprun
 
     user_settings.logger.setLevel(logging.CRITICAL)
 
@@ -98,6 +95,64 @@ def _distances(self, remove_self=True, specie=None) -> list[float]:
             continue
         result.append(distance)
     return result
+
+
+def band_edge_properties_from_vasprun(
+    vasprun: Vasprun, integer_criterion: float = 0.1
+) -> BandEdgeProperties:
+    """
+    Create a ``pydefect`` ``BandEdgeProperties`` object from a ``Vasprun``
+    object.
+
+    Args:
+        vasprun (Vasprun): ``Vasprun`` object.
+        integer_criterion (float):
+            Threshold criterion for determining if a band is unoccupied
+            (< ``integer_criterion``), partially occupied (between
+            ``integer_criterion`` and 1 - ``integer_criterion``), or
+            fully occupied (> 1 - ``integer_criterion``).
+            Default is 0.1.
+
+    Returns:
+        ``BandEdgeProperties`` object.
+    """
+    band_edge_prop = BandEdgeProperties(
+        eigenvalues=eigenvalues_from_vasprun(vasprun),
+        nelect=get_nelect_from_vasprun(vasprun),
+        magnetization=get_magnetization_from_vasprun(vasprun),
+        kpoint_coords=vasprun.actual_kpoints,
+        integer_criterion=integer_criterion,
+    )
+    band_edge_prop.structure = vasprun.final_structure
+    return band_edge_prop
+
+
+def make_perfect_band_edge_state_from_vasp(
+    procar: Procar, vasprun: Vasprun, integer_criterion: float = 0.1
+) -> PerfectBandEdgeState:
+    """
+    Create a ``pydefect`` ``PerfectBandEdgeState`` object from just a
+    ``Vasprun`` and ``Procar`` object, without the need for the ``Outcar``
+    input (as in ``pydefect``).
+
+    Args:
+        procar (Procar): ``Procar`` object.
+        vasprun (Vasprun): ``Vasprun`` object.
[…]
