+++ doped/core.py
@@ -529,7 +529,6 @@ class DefectEntry(thermo.DefectEntry):
         bulk_procar: Optional[Union[str, "Path", "EasyunfoldProcar", Procar]] = None,
         defect_vr: Optional[Union[str, "Path", Vasprun]] = None,
         defect_procar: Optional[Union[str, "Path", "EasyunfoldProcar", Procar]] = None,
-        bulk_outcar: Optional[Union[str, "Path", Outcar]] = None,
         force_reparse: bool = False,
     ):
         """
@@ -546,7 +545,7 @@ class DefectEntry(thermo.DefectEntry):
                 or, failing that, from a ``vasprun.xml(.gz)`` file at
                 ``self.calculation_metadata["bulk_path"]``.
             bulk_procar (str, Path, EasyunfoldProcar, Procar):
-                Either a path to the ``VASP`` ``PROCAR`` output file (with
+                Either a path to the ``VASP`` ``PROCAR(.gz)`` output file (with
                 ``LORBIT > 10`` in the ``INCAR``) or an ``easyunfold``/
                 ``pymatgen`` ``Procar`` object, for the reference bulk supercell
                 calculation. Not required, but speeds up parsing (~50%) if present.
@@ -561,17 +560,12 @@ class DefectEntry(thermo.DefectEntry):
                 or, failing that, from a ``vasprun.xml(.gz)`` file at
                 ``self.calculation_metadata["defect_path"]``.
             defect_procar (str, Path, EasyunfoldProcar, Procar):
-                Either a path to the ``VASP`` ``PROCAR`` output file (with
+                Either a path to the ``VASP`` ``PROCAR(.gz)`` output file (with
                 ``LORBIT > 10`` in the ``INCAR``) or an ``easyunfold``/
                 ``pymatgen`` ``Procar`` object, for the defect supercell calculation.
                 Not required, but speeds up parsing (~50%) if present. If ``None``
                 (default), tries to load from a ``PROCAR(.gz)`` file at
                 ``self.calculation_metadata["defect_path"]``.
-            bulk_outcar (str, Path, Outcar):
-                Either a path to the ``VASP`` ``OUTCAR`` output file or a
-                ``pymatgen`` ``Outcar`` object, for the reference bulk supercell
-                calculation. If ``None`` (default), tries to load from an
-                ``OUTCAR`` file at ``self.calculation_metadata["bulk_path"]``.
             force_reparse (bool):
                 Whether to force re-parsing of the eigenvalue data, even if
                 already present in the ``calculation_metadata``.
@@ -583,7 +577,6 @@ class DefectEntry(thermo.DefectEntry):
         from doped.utils.parsing import (
             _get_output_files_and_check_if_multiple,
             _multiple_files_warning,
-            get_outcar,
             get_procar,
             get_vasprun,
         )
@@ -652,38 +645,15 @@ class DefectEntry(thermo.DefectEntry):
         bulk_vr, bulk_procar = parsed_vr_procar_dict["bulk"]
         defect_vr, defect_procar = parsed_vr_procar_dict["defect"]
 
-        if bulk_outcar is not None and not isinstance(bulk_outcar, Outcar):
-            bulk_outcar = get_outcar(bulk_outcar)
-
-        if bulk_outcar is None:  # try parsing from bulk path
-            bulk_path = self.calculation_metadata["bulk_path"]
-            try:
-                bulk_outcar_path, multiple = _get_output_files_and_check_if_multiple("OUTCAR", bulk_path)
-                if multiple:
-                    _multiple_files_warning(
-                        "OUTCAR",
-                        bulk_path,
-                        bulk_outcar_path,
-                        dir_type="bulk",
-                    )
[…]
