### 2024-03-28 2c4b988
Redefine `_add_eigenvalues` to make plotting with many kpoints and bands much more efficient (~10 seconds -> 0.5 seconds)
Former-commit-id: 3da0cc453992f290bec4498e6c23b786d9f1518a
