+++ doped/core.py
@@ -645,15 +645,12 @@ class DefectEntry(thermo.DefectEntry):
         bulk_vr, bulk_procar = parsed_vr_procar_dict["bulk"]
         defect_vr, defect_procar = parsed_vr_procar_dict["defect"]
 
-        if defect_procar and bulk_procar:
-            band_orb, vbm_info, cbm_info = get_band_edge_info(
-                bulk_vr=bulk_vr,
-                defect_vr=defect_vr,
-                bulk_procar=bulk_procar,
-                defect_procar=defect_procar,
-            )
-        else:
-            band_orb, vbm_info, cbm_info = get_band_edge_info(bulk_vr=bulk_vr, defect_vr=defect_vr)
+        band_orb, vbm_info, cbm_info = get_band_edge_info(
+            bulk_vr=bulk_vr,
+            defect_vr=defect_vr,
+            bulk_procar=bulk_procar,  # may be None, in which case Vasprun.projected_eigenvalues used
+            defect_procar=defect_procar,  # may be None, in which case Vasprun.projected_eigenvalues used
+        )
 
         self.calculation_metadata["eigenvalue_data"] = {
             "band_orb": band_orb,
@@ -777,11 +774,7 @@ class DefectEntry(thermo.DefectEntry):
                 "necessary VASP output files for the defect and bulk supercells (see docstring)."
             )
 
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore")
-            bes, fig = get_eigenvalue_analysis(self, filename=filename, **kwargs)
-
-        return (bes, fig) if plot else bes
+        return get_eigenvalue_analysis(self, plot=plot, filename=filename, **kwargs)
 
     def _get_chempot_term(self, chemical_potentials=None):
         chemical_potentials = chemical_potentials or {}
