### 2024-03-28 5acf184
Expand eigenvalue tests
Former-commit-id: cad2ee54709116483b95ea70cab022f3541ec758
