+++ tests/test_analysis.py
@@ -27,19 +27,18 @@ from doped.analysis import (
 )
 from doped.core import _orientational_degeneracy_warning
 from doped.generation import DefectsGenerator, get_defect_name_from_defect, get_defect_name_from_entry
+from doped.utils.eigenvalues import get_eigenvalue_analysis
 from doped.utils.parsing import (
     get_defect_site_idxs_and_unrelaxed_structure,
     get_defect_type_and_composition_diff,
     get_orientational_degeneracy,
     get_outcar,
+    get_procar,
     get_vasprun,
 )
 from doped.utils.symmetry import point_symmetry
 
 mpl.use("Agg")  # don't show interactive plots if testing from CLI locally
-
-
-# TODO: Remove `parse_projected_eigen = False` when new `vise` version released
 
 
 def if_present_rm(path):
@@ -71,9 +70,6 @@ class DefectsParsingTestCase(unittest.TestCase):
 
         self.Sb2Se3_DATA_DIR = os.path.join(self.module_path, "data/Sb2Se3")
         self.Sb2Se3_dielectric = np.array([[85.64, 0, 0], [0.0, 128.18, 0], [0, 0, 15.00]])
-
-        self.Cu2SiSe3_EXAMPLE_DIR = os.path.join(self.module_path, "../examples/Cu2SiSe3")
-        self.MgO_EXAMPLE_DIR = os.path.join(self.module_path, "../examples/MgO")
 
         self.Sb2Si2Te6_dielectric = [44.12, 44.12, 17.82]
         self.Sb2Si2Te6_DATA_DIR = os.path.join(self.EXAMPLE_DIR, "Sb2Si2Te6")
@@ -324,7 +320,6 @@ class DefectsParsingTestCase(unittest.TestCase):
                 output_path=self.CdTe_EXAMPLE_DIR,
                 dielectric=9.13,
                 json_filename="CdTe_example_defect_dict.json",
-                parse_projected_eigen=False,
             )  # for testing in test_thermodynamics.py
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(default_dp, w)
@@ -368,7 +363,8 @@ class DefectsParsingTestCase(unittest.TestCase):
         warnings.filterwarnings("ignore", "Multiple")
         with warnings.catch_warnings(record=True) as w:
             dp = DefectsParser(
-                output_path=self.CdTe_EXAMPLE_DIR, dielectric=9.13, parse_projected_eigen=False
+                output_path=self.CdTe_EXAMPLE_DIR,
+                dielectric=9.13,
             )
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, multiple_outcars_warning=False)
@@ -388,11 +384,13 @@ class DefectsParsingTestCase(unittest.TestCase):
         print([warn.message for warn in w])  # for debugging
         self._check_default_CdTe_DefectsParser_outputs(dp, w, dist_tol=0.1)
 
+    @custom_mpl_image_compare(filename="CdTe_Te_Cd_+1_eigenvalue_plot.png")
     def test_DefectsParser_CdTe_no_dielectric_json(self):
         # test no dielectric and no JSON:
         with warnings.catch_warnings(record=True) as w:
             dp = DefectsParser(
-                output_path=self.CdTe_EXAMPLE_DIR, json_filename=False, parse_projected_eigen=False
+                output_path=self.CdTe_EXAMPLE_DIR,
+                json_filename=False,
             )
         print([warn.message for warn in w])  # for debugging
         assert any(
@@ -402,6 +400,8 @@ class DefectsParsingTestCase(unittest.TestCase):
         )
         self._check_DefectsParser(dp, skip_corrections=True)
[…]
