+++ pyproject.toml
@@ -30,7 +30,8 @@ dependencies = [
     "pydefect>=0.8.1",  # for Kumagai (eFNV) correction & PHS analysis
     "filelock",
     "vise>=0.9.0",  # for eigenvalue analysis, >=0.9.0 required for SOC compatibility
-    "easyunfold",  # for PROCAR parsing, if any dependency issues arise can make it an optional dependency
+    "easyunfold>=0.3.6",  # for efficient PROCAR parsing (and supporting SOC),
+    # if any dependency issues arise can make easyunfold an optional dependency
 ]
 
 [project.urls]
