### 2024-03-28 a5e5c6c
Update y-limits and legend behaviour, and add `ylims` and `legend_kwargs` option for easy user customisation
Former-commit-id: caa6e5e8f455855e5b49cf0c9966fd91626f9bf9
