### 2024-04-08 #6433 @atimmins7
@kavanase @rkurchin

Doped is a thoughtfully designed piece of software for analyzing charged defects. It is well built, well documented, has useful examples, and will certainly be of benefit to the research community. I fully recommend publishing doped in JOSS. Below I have added some suggestions/comments.

*As a note, I primarily use Quantum Espresso and don't have access to VASP. As a result, I could not verify some specific functionality that actively required VASP. Otherwise functionality was tested through the available tutorials and data sets.

1. Installation > Requirements. The tutorial states: "doped requires pymatgen>=2022.10.22 and its dependencies." This appears to be out of date?
2. Tips and Tricks > Layered / Low Dimensional Materials. Regarding the error "Estimated error in the Kumagai (eFNV) ..." -  The meaning of the error is explained in the `get_kumagai_correction` arguments (and those for the FNV), but I didn't see this explained in the tutorial. It may be helpful to briefly explain how this error is being determined or what it represents.
3. Tips and Tricks > Perturbed Host States. Not super clear what information “OrbDiff” is indicating. May be good to specify.
4. Tips and Tricks > Perturbed Host States. The info shown with `print(bes)` appears to not be aligned with the 'Spin Up' and 'Spin Down' plots. From the printed info, I expected the VBM for spin up to be occupied at the gamma point. Are the titles of the plots perhaps switched?
5. Tutorials > Defect Calculation Parsing. `v_Cd_thermo.get_formation_energies?` not found
6. Tutorials > Defect Calculation Parsing. Output following code `correction = F_O_1_entry.get_kumagai_correction(error_tolerance=0.0001)` results in error "core.py:280: UserWarning: Estimated error in the Kumagai (eFNV) charge correction for defect F_O_1 is 0.001 eV (i.e. which is greater than the `error_tolerance`: 0.000 eV)." Given that a very small error tolerance has the digits truncated, is it perhaps worth it to change the error to reflect the number of digits provided by the user?
7.  Tutorials > Defect Thermodynamics and Doping >  Dopability Limits. Possible typo in statement "For example here we see that (coordinated by Te anions) is our dominant (lowest-energy) compensating native donor under p-type (Te-rich) conditions, and we can see the doping window of 0.85 eV corresponds to its formation energy at the VBM under Te-rich conditions in the plot above.” I think ‘0.85 eV’ may be a typo. Looking at the table and plot, it looks like it should be ‘0.489’ eV?
8. Tutorials > Defect Thermodynamics and Doping > CdTe: Cd-Rich. Functions `get_quenched_fermi_level_and_concentrations` and `get_equilibrium_fermi_level` do not include keyword argument fermi_dos and running the code blocks with this keyword argument resulted in errors. This issue came up a handful of times in the provided tutorials.
9. Tutorials > Defect Thermodynamics and Doping > Section 'CdTe: Approximating Temperature-Dependent Band Gap'. My understanding is that the experimental data ('# CdTe gap vs Temp, 2014 study') is added such that we can compare it to the results predicted by doped. Later there is a plot showing how the band gap changes with respect to anneal temp. However, its hard to compare experimental and predicted values as they are split up between plots. Perhaps adjust the experimental plot so that is shows the band gap predicted by doped?
10. Tutorials > Competing Phases > Eigenvalue / Electronic Structure Analysis. It appears that the titles for the plots (up vs down) are flipped wrt to the bes printed data. Same comment as mentioned for Tips & Tricks.
11.  Tutorials > Plot Customization. The chemical potential limit specification of competing phases may be good to briefly explain (e.g. limit="Cd-CdTe")? It might be helpful to briefly clarify what is meant by this setting.
12. Tutorials > Full Defect Workflow. In section 8.2, you have the following: “Since we have calculated the chemical potentials in the previous section [ Chemical Potentials](https://doped.readthedocs.io/en/latest/doped.chemical_potentials.html#module-doped.chemical_potentials), we can just load the results from the JSON file here:” It looks like this is meant to point to chemical potentials section, but the hyperlink points to the doc string.
