+++ tests/test_chemical_potentials.py
@@ -63,9 +63,13 @@ class ChemPotsTestCase(unittest.TestCase):
 
         assert len(stable_cpa.elemental) == 2
         assert len(self.ext_cpa.elemental) == 3
-        assert any(entry["formula"] == "O2" for entry in stable_cpa.data)
+        assert any(entry["Formula"] == "O2" for entry in stable_cpa.data)
         assert np.isclose(
-            next(entry["energy_per_fu"] for entry in self.ext_cpa.data if entry["formula"] == "La2Zr2O7"),
+            next(
+                entry["DFT Energy (eV/fu)"]
+                for entry in self.ext_cpa.data
+                if entry["Formula"] == "La2Zr2O7"
+            ),
             -119.619571095,
         )
 
@@ -88,7 +92,7 @@ class ChemPotsTestCase(unittest.TestCase):
         )
         self.ext_cpa.from_csv(self.csv_path_ext)
         chempot_df = self.ext_cpa.calculate_chempots()
-        assert next(iter(chempot_df["La_limiting_phase"])) == "La2Zr2O7"
+        assert next(iter(chempot_df["La-Limiting Phase"])) == "La2Zr2O7"
         assert np.isclose(next(iter(chempot_df["La"])), -9.46298748)
 
     def test_ext_cpa_chempots(self):
@@ -107,8 +111,8 @@ class ChemPotsTestCase(unittest.TestCase):
         stable_cpa = chemical_potentials.CompetingPhasesAnalyzer(self.stable_system)
         stable_cpa.from_csv(self.csv_path)
         chempot_df = stable_cpa.calculate_chempots(sort_by="Zr")
-        assert np.isclose(next(iter(chempot_df["Zr"])), -0.199544)
-        assert np.isclose(list(chempot_df["Zr"])[1], -10.975428439999998)
+        assert np.isclose(next(iter(chempot_df["Zr"])), -0.199544, atol=1e-4)
+        assert np.isclose(list(chempot_df["Zr"])[1], -10.975428439999998, atol=1e-4)
 
         with pytest.raises(KeyError):
             stable_cpa.calculate_chempots(sort_by="M")
@@ -119,7 +123,7 @@ class ChemPotsTestCase(unittest.TestCase):
         path = self.path / "ZrO2"
         cpa.from_vaspruns(path=path, folder="relax", csv_path=self.csv_path)
         assert len(cpa.elemental) == 2
-        assert cpa.data[0]["formula"] == "O2"
+        assert cpa.data[0]["Formula"] == "O2"
 
         cpa_no = chemical_potentials.CompetingPhasesAnalyzer(self.stable_system)
         with pytest.raises(FileNotFoundError):
@@ -134,7 +138,7 @@ class ChemPotsTestCase(unittest.TestCase):
         assert len(ext_cpa.elemental) == 3
         # sorted by num_species, then alphabetically, then by num_atoms_in_fu, then by
         # formation_energy
-        assert [entry["formula"] for entry in ext_cpa.data] == [
+        assert [entry["Formula"] for entry in ext_cpa.data] == [
             "La",
             "O2",
             "Zr",
@@ -171,18 +175,21 @@ class ChemPotsTestCase(unittest.TestCase):
         # -278.46392292, 'formation_energy': -10.951109995000003}, {'formula': 'La2Zr2O7',
         # 'kpoints': '3x3x3', 'energy_per_fu': -119.619571095, 'energy_per_atom':
         # -10.874506463181818, 'energy': -239.23914219, 'formation_energy': -40.87683184}]
-        assert np.isclose(ext_cpa.data[0]["energy_per_fu"], -5.00458616)
[…]
