+++ doped/analysis.py
@@ -999,12 +999,20 @@ class DefectsParser:
 
         def _call_multiple_corrections_tolerance_warning(correction_errors, type="FNV"):
             long_name = "Freysoldt" if type == "FNV" else "Kumagai"
-            correction_errors_string = "\n".join(
-                f"{name}: {error:.3f} eV" for name, error in correction_errors
-            )
+            if error_tolerance >= 0.01:  # if greater than 10 meV, round energy values to meV:
+                error_tol_string = f"{error_tolerance:.3f}"
+                correction_errors_string = "\n".join(
+                    f"{name}: {error:.3f} eV" for name, error in correction_errors
+                )
+            else:  # else give in scientific notation:
+                error_tol_string = f"{error_tolerance:.2e}"
+                correction_errors_string = "\n".join(
+                    f"{name}: {error:.2e} eV" for name, error in correction_errors
+                )
+
             warnings.warn(
                 f"Estimated error in the {long_name} ({type}) charge correction for certain "
-                f"defects is greater than the `error_tolerance` (= {error_tolerance:.3f} eV):"
+                f"defects is greater than the `error_tolerance` (= {error_tol_string} eV):"
                 f"\n{correction_errors_string}\n"
                 f"You may want to check the accuracy of the corrections by plotting the site "
                 f"potential differences (using `defect_entry.get_{long_name.lower()}_correction()`"
@@ -1318,6 +1326,11 @@ def _parse_vr_and_poss_procar(
     parse_procar: bool = True,
 ):
     procar = None
+
+    failed_eig_parsing_warning_message = (
+        f"Could not parse eigenvalue data from vasprun.xml.gz files in {label} folder at {output_path}"
+    )
+
     try:
         vr = get_vasprun(
             vr_path,
@@ -1326,6 +1339,7 @@ def _parse_vr_and_poss_procar(
         )  # vr.eigenvalues not needed for defects except for vr-only eigenvalue analysis
     except Exception as vr_exc:
         vr = get_vasprun(vr_path, parse_projected_eigen=False, parse_eigen=label == "bulk")
+        failed_eig_parsing_warning_message += f", got error:\n{vr_exc}"
 
         if parse_procar:
             procar_path, multiple = _get_output_files_and_check_if_multiple("PROCAR", output_path)
@@ -1334,15 +1348,14 @@ def _parse_vr_and_poss_procar(
                     procar = get_procar(procar_path)
 
                 except Exception as procar_exc:
-                    if parse_projected_eigen is not None:  # otherwise no warning
-                        warning_message = (
-                            f"Could not parse eigenvalue data from vasprun.xml.gz files in "
-                            f"{label} folder at {output_path}, got error:\n{vr_exc}\nThen got "
-                        )
-                        warnings.warn(
-                            f"{warning_message}the following error when attempting to parse projected "
-                            f"eigenvalues from the defect PROCAR(.gz):\n{procar_exc}"
-                        )
+                    failed_eig_parsing_warning_message += (
+                        f"\nThen got the following error when attempting to parse projected eigenvalues "
+                        f"from the defect PROCAR(.gz):\n{procar_exc}"
+                    )
+
+    if vr.projected_eigenvalues is None and procar is None and parse_projected_eigen is True:
+        # only warn if parse_projected_eigen is set to True (not None)
+        warnings.warn(failed_eig_parsing_warning_message)
 
     return vr, procar if parse_procar else vr
 
@@ -1785,7 +1798,7 @@ class DefectParser:
                     defect_procar=defect_procar,
                 )
             except Exception as exc:
-                if parse_projected_eigen is not None:  # otherwise no warning
[…]
