+++ tests/test_analysis.py
@@ -610,13 +610,13 @@ class DefectsParsingTestCase(unittest.TestCase):
                 "Estimated error in the Freysoldt (FNV) ",
                 "Estimated error in the Kumagai (eFNV) ",
                 "charge correction for certain defects is greater than the `error_tolerance` (= "
-                "0.001 eV):",
-                "v_Cd_-2: 0.011 eV",
-                "v_Cd_-1: 0.008 eV",
-                "Int_Te_3_1: 0.003 eV",
-                "Te_Cd_+1: 0.002 eV",
-                "Int_Te_3_Unperturbed_1: 0.005 eV",
-                "Int_Te_3_2: 0.012 eV",
+                "1.00e-03 eV):",
+                "v_Cd_-2: 1.13e-02 eV",
+                "v_Cd_-1: 7.91e-03 eV",
+                "Int_Te_3_1: 3.10e-03 eV",
+                "Te_Cd_+1: 2.02e-03 eV",
+                "Int_Te_3_Unperturbed_1: 4.91e-03 eV",
+                "Int_Te_3_2: 1.24e-02 eV",
                 "You may want to check the accuracy of the corrections by",
                 "(using `defect_entry.get_freysoldt_correction()` with `plot=True`)",
                 "(using `defect_entry.get_kumagai_correction()` with `plot=True`)",
@@ -1913,8 +1913,8 @@ class DopedParsingTestCase(unittest.TestCase):
             )
         assert (
             f"Estimated error in the Kumagai (eFNV) charge correction for defect "
-            f"{int_F_minus1_ent.name} is 0.003 eV (i.e. which is greater than the `error_tolerance`: "
-            f"0.001 eV). You may want to check the accuracy of the correction by plotting the site "
+            f"{int_F_minus1_ent.name} is 2.58e-03 eV (i.e. which is greater than the `error_tolerance`: "
+            f"1.00e-03 eV). You may want to check the accuracy of the correction by plotting the site "
             f"potential differences (using `defect_entry.get_kumagai_correction()` with "
             f"`plot=True`). Large errors are often due to unstable or shallow defect charge states ("
             f"which can't be accurately modelled with the supercell approach; see "
@@ -2250,53 +2250,6 @@ def _remove_metadata_keys_from_dict(d: dict) -> dict:
             d[key] = _remove_metadata_keys_from_dict(d[key])
 
     return d
-
-
-def _compare_band_edge_states_dicts(d1, d2, orb_diff_tol: float = 0.1):
-    """
-    Compare two dictionaries of band edge states, removing metadata keys and
-    allowing a slight difference in the ``vbm/cbm_orbital_diff`` values to
-    account for rounding errors with ``PROCAR``s.
-    """
-    if isinstance(d1, str):
-        d1 = loadfn(d1)
-    if isinstance(d2, str):
-        d2 = loadfn(d2)
-
-    d1 = d1.as_dict()
-    d2 = d2.as_dict()
-
-    cbm_orbital_diffs1 = [subdict.pop("cbm_orbital_diff") for subdict in d1["states"]]
-    cbm_orbital_diffs2 = [subdict.pop("cbm_orbital_diff") for subdict in d2["states"]]
-    for i, j in zip(cbm_orbital_diffs1, cbm_orbital_diffs2):
-        print(f"cbm_orbital_diffs: {i:.3f} vs {j:.3f}")
-        assert np.isclose(i, j, atol=orb_diff_tol)
-    vbm_orbital_diffs1 = [subdict.pop("vbm_orbital_diff") for subdict in d1["states"]]
-    vbm_orbital_diffs2 = [subdict.pop("vbm_orbital_diff") for subdict in d2["states"]]
[…]
