### 2024-04-10 #6433 @kavanase
Thanks @atimmins7 for your review and helpful comments! 😃
I've now resolved each of the issues raised, with responses below:

1. > Installation > Requirements. The tutorial states: "doped requires pymatgen>=2022.10.22 and its dependencies." This appears to be out of date?

Ah yes this was missed in a previous update due to changes in `pymatgen`, now updated to
> The doped dependencies are listed in the pyproject.toml file on [the GitHub repository](https://github.com/SMTG-Bham/doped/blob/main/pyproject.toml).

https://github.com/SMTG-Bham/doped/pull/61/commits/451c954441dc4e1a1c186eb7edc9eaaa8e39c498

2. > Tips and Tricks > Layered / Low Dimensional Materials. Regarding the error "Estimated error in the Kumagai (eFNV) ..." - The meaning of the error is explained in the get_kumagai_correction arguments (and those for the FNV), but I didn't see this explained in the tutorial. It may be helpful to briefly explain how this error is being determined or what it represents.

This is now briefly explained in the `get_kumagai/freysoldt_correction` docstrings, tips page and parsing tutorials 👍
https://github.com/SMTG-Bham/doped/pull/61/commits/451c954441dc4e1a1c186eb7edc9eaaa8e39c498

3. > Tips and Tricks > Perturbed Host States. Not super clear what information “OrbDiff” is indicating. May be good to specify.

This is now updated in the tips docs page and advanced analysis tutorial 👍
https://github.com/SMTG-Bham/doped/pull/61/commits/451c954441dc4e1a1c186eb7edc9eaaa8e39c498

4. > Tips and Tricks > Perturbed Host States. The info shown with print(bes) appears to not be aligned with the 'Spin Up' and 'Spin Down' plots. From the printed info, I expected the VBM for spin up to be occupied at the gamma point. Are the titles of the plots perhaps switched?

Ah good spot! This was a mislabelling in the code, and has been fixed now (with docs/tutorials images updated). Thanks!
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

5. > Tutorials > Defect Calculation Parsing. v_Cd_thermo.get_formation_energies? not found

This was missed in a previous tutorials update (where the object name was changed) – fixed now!
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

6. > Tutorials > Defect Calculation Parsing. Output following code correction = F_O_1_entry.get_kumagai_correction(error_tolerance=0.0001) results in error "core.py:280: UserWarning: Estimated error in the Kumagai (eFNV) charge correction for defect F_O_1 is 0.001 eV (i.e. which is greater than the error_tolerance: 0.000 eV)." Given that a very small error tolerance has the digits truncated, is it perhaps worth it to change the error to reflect the number of digits provided by the user?

This has now been updated so that the error and tolerance values in eV are rounded to 3 decimal places (i.e. 1 meV) if the tolerance value is 10 meV or more, otherwise scientific notation with 2 decimal places is used, to avoid excessive rounding for these cases:
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

7. > Tutorials > Defect Thermodynamics and Doping > Dopability Limits. Possible typo in statement "For example here we see that (coordinated by Te anions) is our dominant (lowest-energy) compensating native donor under p-type (Te-rich) conditions, and we can see the doping window of 0.85 eV corresponds to its formation energy at the VBM under Te-rich conditions in the plot above.” I think ‘0.85 eV’ may be a typo. Looking at the table and plot, it looks like it should be ‘0.489’ eV?

Ah yes good spot! The $Cd_i^{+2}$ and $V_{Cd}^{-2}$ energies were mixed up here, fixed now!
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

8. > Tutorials > Defect Thermodynamics and Doping > CdTe: Cd-Rich. Functions get_quenched_fermi_level_and_concentrations and get_equilibrium_fermi_level do not include keyword argument fermi_dos and running the code blocks with this keyword argument resulted in errors. This issue came up a handful of times in the provided tutorials.

Thanks for checking this! This was due to a typo when updating in response to query 5 from reviewer 1 above (https://github.com/SMTG-Bham/doped/commit/c29cd1b9068d733f8259dd9bda975bc6b0e38d7f), fixed now.
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

9. > Tutorials > Defect Thermodynamics and Doping > Section 'CdTe: Approximating Temperature-Dependent Band Gap'. My understanding is that the experimental data ('# CdTe gap vs Temp, 2014 study') is added such that we can compare it to the results predicted by doped. Later there is a plot showing how the band gap changes with respect to anneal temp. However, its hard to compare experimental and predicted values as they are split up between plots. Perhaps adjust the experimental plot so that is shows the band gap predicted by doped?

So here the experimental gap is not being predicted by `doped`, but rather we are just fitting a function to the temperature dependent band gap from experiment, and then using this temperature-dependent band gap in the defect & carrier concentration calculations (as they are affected by the band edge positions, particularly at high temperatures). This section of that tutorial was not overly clear, so I have updated this to better explain what that code is doing & why.
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

10. > Tutorials > Competing Phases > Eigenvalue / Electronic Structure Analysis. It appears that the titles for the plots (up vs down) are flipped wrt to the bes printed data. Same comment as mentioned for Tips & Tricks.

Yep same issue as Q4, fixed now and updated!
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

11. > Tutorials > Plot Customization. The chemical potential limit specification of competing phases may be good to briefly explain (e.g. limit="Cd-CdTe")? It might be helpful to briefly clarify what is meant by this setting.

This is now explained in the tutorials 👍
https://github.com/SMTG-Bham/doped/pull/61/commits/9804968ebcc1a97350ab7e98736a3b79c076e7f1

12. > Tutorials > Full Defect Workflow. In section 8.2, you have the following: “Since we have calculated the chemical potentials in the previous section [Chemical Potentials](https://doped.readthedocs.io/en/latest/doped.chemical_potentials.html#module-doped.chemical_potentials), we can just load the results from the JSON file here:” It looks like this is meant to point to chemical potentials section, but the hyperlink points to the doc string.

The link has now been updated to point to the chemical potentials section 👍
https://github.com/SMTG-Bham/doped/pull/61/commits/bbba1fcb605c7e00aa80bd811b781e6470b1b9e5
