### 2024-04-10 #6433 @kavanase
I've made a release of the `doped` repo following the JOSS review updates, with version tag `2.4.0-JOSS`.

This is now archived on Zenodo with matching title name, author list and license at:
`10.5281/zenodo.10957359`

I've also double-checked the author and affiliation list, so all points on the author to-do list above are ✅
