### 2024-04-10 #6433 @rkurchin
Some editorial comments:
* missing an Oxford comma at the end of line 15
* somewhat similarly, line 17 reads a bit strangely, I wonder if adding an "and" instead of the comma between "efficient" and "user-friendly" would be smoother
* line 24: "which" -> "that"
* 30: need parentheses or something around citation
* 30: missing Oxford comma after "user-friendly"
* 40: I would change the ampersand to writing out the word "and"
* 43: either "is a Python package" or "is Python software"
* 43: comma after "-processing"
* 45-46: same issue from Summary
* 49: comma after outputs
* Figure 2: I would add an "overall" caption before the subcaptions to describe the purpose of the figure, especially because the way the paper renders, the reader comes to the figure before the reference to it in the text. Also, the first two subplots are labeled like "a." and the rest are labeled like "(c)" so make that consistent. And the rendering of the chemical formulae in parts (i) and (k) are clunky looking, make sure that you use `\mathrm`, `\text` or equivalent to avoid the weird italicization issues in both the formulae and the subscripts, as in $\text{Cu}_2\text{SiS}_3$ rather than $Cu_2SiS_3$ and also in subscript of vacancies, etc.
* 90: no need for a hyphen
* 118: is "2/3" here meaning the fraction or "two or three"? If the former, maybe write as "2/3 of an order of magnitude" and if the latter, spell out the numbers as I've done here to clarify
* 133: no need for the hyphen in "highly-customisable"
* 139: comma after close parenthesis
* 148: comma after "works"

References:
* line 278: some capitalization mishaps in the oxidation state that need some curly brackets to fix them...
* 304-310, 366, 383: missing subscripts in chemical formulae, presuming they are really there in the article titles...
* 356, 363: some errant mathmode stuff

A question of technical curiosity that is certainly beyond the scope of what's needed for this paper: Have you considered incorporating directional screening in your supercell optimization? If you're using the eFNV scheme that allows for anisotropic screening, it has always seemed to me that you should really optimize over something like the permittivity times the distance rather than just the distance, and I've long wondered how much of a difference that would make in supercell size convergence, etc. If you ever wanted to work together on testing this out, I'd love to chat about it!
